% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_date_range.R
\name{as_date_range_custom}
\alias{as_date_range_custom}
\title{Convert non-ambiguous labels for customised periods
or cohorts}
\usage{
as_date_range_custom(x, month_start = "Jan")
}
\arguments{
\item{x}{A vector of period or cohort labels.}

\item{month_start}{An element of \code{\link[base]{month.name}},
or \code{\link[base]{month.abb}}. Each
period  or cohort starts on
the first day of this month.}
}
\value{
A factor the same length as \code{x}.
}
\description{
Convert standard labels for periods
or cohorts with varying widths
(denominated in years), such as
such as \code{"<2020"}, \code{"2020-2030"},
\code{"2030-2032"}
to labels that give precise date ranges
such as \code{"(-Inf, 2020-06-30]"},
\code{"[2020-07-01, 2030-06-30]"},
\code{"[2030-07-01, , 2032-06-30]"}.
}
\details{
The vector of period or cohort labels
\code{x} is typically
created using functions
\code{\link{format_period_custom}}
or \code{\link{format_cohort_custom}}.
}
\examples{
x <- c("<2010", "2010-2012", "2012-2020")
as_date_range_custom(x)
as_date_range_custom(x, month_start = "Jul")
}
\seealso{
\code{\link{as_date_range_year}},
\code{\link{as_date_range_multi}},
\code{\link{as_date_range_quarter}},
\code{\link{as_date_range_month}}
}
