% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_date_range.R
\name{as_date_range_quarter}
\alias{as_date_range_quarter}
\title{Convert standard labels for one-quarter periods
or cohorts to labels specifying precise dates}
\usage{
as_date_range_quarter(x)
}
\arguments{
\item{x}{A vector of period or cohort labels.}
}
\value{
A factor the same length as \code{x}.
}
\description{
Convert standard labels for one-quarter
(three-month) periods
or cohorts, such as
such as \code{"<2020 Q2"}, \code{"2020 Q2"},
\code{"2020 Q3"}
to labels that give precise date ranges
such as \code{"(-Inf, 2020-03-31]"},
\code{"[2020-04-01, 2020-06-30]"},
\code{"[2020-07-01, 2020-10-31]"}.
}
\details{
The vector of period or cohort labels
\code{x} is typically
created by functions
\code{\link{format_period_quarter}},
or \code{\link{format_cohort_quarter}}.
}
\examples{
x <- c("<2010 Q4", "2010 Q4", "2011 Q1")
as_date_range_quarter(x)
}
\seealso{
\code{\link{as_date_range_year}},
\code{\link{as_date_range_multi}},
\code{\link{as_date_range_custom}},
\code{\link{as_date_range_month}}.
}
