% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_date_range.R
\name{as_date_range_year}
\alias{as_date_range_year}
\title{Convert non-ambiguous labels for one-year periods
or cohorts}
\usage{
as_date_range_year(x, month_start = "Jan", label_year_start = TRUE)
}
\arguments{
\item{x}{A vector of period or cohort labels.}

\item{month_start}{An element of \code{\link[base]{month.name}},
or \code{\link[base]{month.abb}}. Each
period  or cohort starts on
the first day of this month.}

\item{label_year_start}{Whether to label a period
by the calendar year at the beginning of the period
or the calendar year at the end. Defaults to \code{TRUE}.}
}
\value{
A factor the same length as \code{x}.
}
\description{
Convert standard labels for one-year periods
such as "2020", "2001", or "<2005" to
labels that give precise date ranges
such as \code{"[2020-01-01, 2020-12-31]"},
\code{"[2001-10-01, 2002-10-01]"},
or \code{"(-Inf, 2005-06-30]"}.
}
\details{
The vector of period or cohort labels \code{x} is typically
created using functions
\code{\link{format_period_year}}
or \code{\link{format_cohort_year}}.
}
\examples{
x <- c("<2010", "2010", "2011")
as_date_range_year(x)
as_date_range_year(x,
                   month_start = "Jul")
as_date_range_year(x,
                   month_start = "Jul",
                   label_year_start = FALSE)
}
\seealso{
\code{\link{as_date_range_multi}},
\code{\link{as_date_range_custom}},
\code{\link{as_date_range_quarter}},
\code{\link{as_date_range_month}}
}
