% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_cohort}
\alias{clean_cohort}
\alias{clean_cohort_df}
\title{Tidy cohort labels}
\usage{
clean_cohort(x, language = "English")

clean_cohort_df(x, language = "English")
}
\arguments{
\item{x}{A numeric or character vector.}

\item{language}{The language in which text
labels are written. Defaults to English.}
}
\value{
\code{clean_cohort} returns a character vector with the same
length as \code{x} in which labels that have been
parsed are translated to dem formats.
\code{clean_cohort_df} returns a data frame with columns
\code{"input"}, \code{"output"}, and \code{"is_valid"}.
}
\description{
Try to parse cohort
labels and convert them
to the format used by the dem packages.
}
\details{
Intervals that are open on the left
such as \code{"<2000"} are
allowed. Intervals that are open on the right
such as \code{"2000+"} are not.

By default, \code{clean_cohort} assumes that any
text labels are written in English. However,
other languages can be specified using
the \code{language} argument. Current choices are
ADD OVER TIME.

Function \code{clean_cohort_df} returns a data frame
showing how each unique element in \code{x} is
interpreted by function \code{clean_cohort} and whether
the element can be interpreted as a valid
cohort label. \code{clean_cohort_df} can be
used to check whether \code{clean_cohort} is
giving the desired results.
}
\examples{
x <- c("before 2000",  ## open on left
       "after 2000",   ## open on right
       "Millenials",
       "2020 Jan",
       "Q3 2020",
       "January 2020",
       "2025 first quarter",
       "untranslatable",
       "2020-2025")
x
clean_cohort(x)
clean_cohort_df(x)
}
