% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_period}
\alias{clean_period}
\alias{clean_period_df}
\title{Tidy period labels}
\usage{
clean_period(x, language = "English")

clean_period_df(x, language = "English")
}
\arguments{
\item{x}{A numeric or character vector.}

\item{language}{The language in which text
labels are written. Defaults to English.}
}
\value{
\code{clean_period} returns a character vector with the same
length as \code{x} in which labels that have been
parsed are translated to dem formats.
\code{clean_period_df} returns a data frame with columns
\code{"input"}, \code{"output"}, and \code{"is_valid"}.
}
\description{
Try to parse period
labels and convert them
to the format used by the dem packages.
}
\details{
Open intervals such as \code{"<2020"}
or \code{"2020+"} are not allowed.

By default, \code{clean_period} assumes that any
text labels are written in English. However,
other languages can be specified using
the \code{language} argument. Current choices are
ADD OVER TIME.

Function \code{clean_period_df} returns a data frame
showing how each unique element in \code{x} is
interpreted by function \code{clean_period} and whether
the element can be interpreted as a valid
period label. \code{clean_period_df} can be
used to check whether \code{clean_period} is
giving the desired results.
}
\examples{
x <- c("before 2000", ## open on left
       "after 2000",  ## open on right
       "2020 Jan",
       "Q3 2020",
       "January 2020",
       "2025 first quarter",
       "untranslatable",
       "2020-2025")
x
clean_period(x)
clean_period_df(x)
}
