% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_age.R
\name{date_to_age_custom}
\alias{date_to_age_custom}
\title{Convert dates to customized age groups}
\usage{
date_to_age_custom(date, dob, breaks = NULL, open_last = TRUE)
}
\arguments{
\item{date}{Dates of events.
A vector of class \code{\link[base]{Date}},
or a vector that can be coerced to class
\code{Date} using function \code{\link[base]{as.Date}}.}

\item{dob}{Dates of birth.
A vector of class \code{\link[base]{Date}},
or a vector that can be coerced to class
\code{Date} using function \code{\link[base]{as.Date}}.}

\item{breaks}{A vector of strictly increasing integer values.}

\item{open_last}{Whether the final age group
has no upper limit. Defaults to \code{TRUE}.}
}
\value{
A factor with the same length as
\code{date} or \code{dob}, whichever
is longer.
}
\description{
Given the dates when events occurred,
and the dates of birth of the people experiencing the events,
allocate the events to age groups.
\code{date_to_age_custom} is the most flexible
of the \code{date_to_age} functions
in that the age groups can have any combination of widths,
though the widths must be defined in whole numbers of years.
}
\details{
\code{date} and \code{dob} must have the same length,
unless one of them has length 1, in which case the
length-1 argument is recycled.

\code{breaks} is used to specify the points at which
each age group starts and finishes. If 
\code{open_last} is \code{TRUE}, and \code{b} is
the last value for \code{breaks}, then the oldest
age group is \code{[b, Inf)} years. 
If \code{open_last} is \code{FALSE}, \code{a} is the
second-to-last value for \code{breaks} and \code{b}
is the last value, then the oldest age
group is \code{[a, b)} years.

The return value is a factor. The levels of this
factor contain all intermediate age groups,
including ones that do not appear in the data.
}
\examples{
date_to_age_custom(date = c("2024-03-27",
                            "2022-11-09"),
                   dob = c("2001-03-21",
                           "2000-07-13"),
                   breaks = c(0, 15, 60))
date_to_age_custom(date = c("2024-03-27",
                            "2022-11-09"),
                   dob = c("2001-03-21",
                           "2000-07-13"),
                   breaks = c(15, 40, 65))

## alternative specifications for oldest age group
date_to_age_custom(date = c("2024-03-27",
                            "2022-11-09"),
                   dob = c("2001-03-21",
                           "2000-07-13"),
                   breaks = c(15, 65, 100))
date_to_age_custom(date = c("2024-03-27",
                            "2022-11-09"),
                   dob = c("2001-03-21",
                           "2000-07-13"),
                   breaks = c(15, 65, 100),
                   open_last = FALSE)
}
\seealso{
Other functions for creating age groups are
\code{\link{date_to_age_year}},
\code{\link{date_to_age_multi}},
\code{\link{date_to_age_lifetab}},
\code{\link{date_to_age_births}},
\code{\link{date_to_age_quarter}},
and \code{\link{date_to_age_month}}.

\code{\link{make_labels_age}} describes the rules
for constructing labels for age groups.

\code{\link{plot_date_to_age_custom}} creates plots
that illustrate how \code{date_to_age_custom}
works.
}
