% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_age_group.R
\name{date_to_age_group}
\alias{date_to_age_group}
\alias{date_to_age_group_year}
\alias{date_to_age_group_multi}
\alias{date_to_age_group_lifetab}
\alias{date_to_age_group_fert}
\alias{date_to_age_group_custom}
\title{Convert dates to age groups}
\usage{
date_to_age_group_year(date, dob, age_max = 100, open_right = TRUE,
  as_factor = TRUE)

date_to_age_group_multi(date, dob, width = 5, age_max = 100,
  open_right = TRUE, as_factor = TRUE)

date_to_age_group_lifetab(date, dob, age_max = 100, as_factor = TRUE)

date_to_age_group_fert(date, dob, age_min = 15, age_max = 50,
  width = 1, recode_up = TRUE, recode_down = TRUE,
  as_factor = TRUE)

date_to_age_group_custom(date, dob, breaks = NULL, open_right = TRUE,
  as_factor = TRUE)
}
\arguments{
\item{date}{The date of the event or measurement.}

\item{dob}{The individual's date of birth.}

\item{as_factor}{Whether the return value should be a factor
Defaults to \code{TRUE}.}

\item{width}{The width, in years, of the age groups.
A positive integer.}

\item{breaks}{The points dividing the intervals.}

\item{age_open}{The minimum age for the open age group 
age, measured in months, quarters, or years.}
}
\value{
If \code{as_factor} is \code{TRUE}, a factor;
otherwise a character vector. The length equals the length
of \code{date} or the length of \code{dob}, whichever
is greater.
}
\description{
Convert precise dates into broader age groups,
in a dataset describing individual people or events.
}
\details{
Raw data on individual people or events often do not
give people's ages but instead give the
date when the data was collected or the event occurred,
plus the person's date of birth. Functions \code{date_to_age_group_month},
\code{date_to_age_group_quarter}, and \code{date_to_age_group_year}
can be used to calculate ages from information on dates.
The functions focus on the typical cases, rather than try to
cover every possible input or output.

With one exception, the age groups constructed
by \code{date_to_age_group_month} always have a
width of one month, and the age groups constructed
by \code{date_to_age_group_year} always have a
width of one quarter.  The exception is the
highest age group, which is typically 'open', i.e. has
no upper limit. The age groups constructed by \code{date_to_age_group_year}
by default have a width of one year. However, setting \code{width}
to a number other than 1 will give wider intervals.
For intervals with varying lengths, use \code{breaks}.
See below for examples.

The \code{age_open} argument specifies the start of
the open age group. By default this is 1200 months
for \code{date_to_age_group_month},
400 quarters for \code{date_to_age_group_quarter}, 
and for \code{date_to_age_group_year}.  If you don't want an
open age group, set \code{age_open} to \code{NULL}.

\code{date} and \code{dob} must have the same length,
unless one of them has length 1, in which case the
length-1 argument is recycled.

By default, the functions returns a factor. The levels of this
factor contain all intermediate age groups between the lowest
and highest, even if these were not found in the data.
To return a character vector instead, set \code{as_factor}
to \code{FALSE}.
}
\seealso{
To turn dates into periods, cohorts, or Lexis triangles,
use functions such as \code{date_to_period_year},
\code{date_to_cohort_year}, and \code{date_to_triangle_year}.
If none of these functions can do the job, try \code{\link[base]{cut}},
possibly in combination with \code{\link{make_age_labels}}.

WHAT ABOUT LOWEST AGE GROUP? HOW TO SPECIFY OPEN AGE GROUP WHEN
USING BREAKS ARG?


If as_factor is TRUE, and date or dob have NAs, then the
levels of the answer will include NA.
}
