% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_age_group.R
\name{date_to_age_group_custom}
\alias{date_to_age_group_custom}
\title{Convert dates to customized age groups}
\usage{
date_to_age_group_custom(date, dob, breaks = NULL, open_last = TRUE,
  as_factor = TRUE)
}
\arguments{
\item{date}{Dates of events or measurements.}

\item{dob}{Dates of birth.}

\item{breaks}{A vector of strictly increasing integer values.}

\item{open_last}{Whether the final age group
has no upper limit. Defaults to \code{TRUE}.}

\item{as_factor}{Whether the return value is a factor.
Defaults to \code{TRUE}.}
}
\value{
If \code{as_factor} is \code{TRUE}, then the
return value is a factor; otherwise it is a character vector.
The length of the return value equals the length
of \code{date} or the length of \code{dob}, whichever
is greater.
}
\description{
Given dates when events occurred or measurements were made,
and dates of birth, allocate the events or measurements
to age groups. \code{date_to_age_group_custom} is the most flexible
of the \code{date_to_age_group} functions
in that the age groups can have any combination of widths,
though the widths must be defined in whole numbers of years.
}
\details{
\code{date} and \code{dob} are both vectors of class
\code{\link[base]{Date}}, or vectors that can be coerced to class
\code{Date} via function \code{\link[base]{as.Date}}.

\code{date} and \code{dob} must have the same length,
unless one of them has length 1, in which case the
length-1 argument is recycled.

\code{breaks} is used to specify the points at which
each age group starts and finishes. If 
\code{open_last} is \code{TRUE}, and \code{b} is
the last value for \code{breaks}, then the oldest
age group is \code{[b, Inf)} years. 
If \code{open_last} is \code{FALSE}, \code{a} is the
second-to-last value for \code{breaks} and \code{b}
is the last value, then the oldest age
group is \code{[a, b)} years.

When \code{as_factor} is \code{TRUE} the levels of
the factor include all intermediate age groups,
including age groups that not appear in the data.
}
\examples{
date_to_age_group_custom(date = c("2024-03-27", "2022-11-09"),
                         dob = c("2001-03-21", "2000-07-13"),
                         breaks = c(0, 15, 60))
date_to_age_group_custom(date = c("2024-03-27", "2022-11-09"),
                         dob = c("2001-03-21", "2000-07-13"),
                         breaks = c(15, 40, 65))

## replicate date of birth
date_to_age_group_custom(date = c("2024-03-27", "2022-11-09"),
                         dob = "2001-01-01",
                         breaks = c(0, 15, 60))

## return non-factor
date_to_age_group_custom(date = c("2024-03-27", "2022-11-09"),
                         dob = "2001-01-01",
                         breaks = c(0, 15, 60),
                         as_factor = FALSE)

## alternative specifications for oldest age group
date_to_age_group_custom(date = c("2024-03-27", "2022-11-09"),
                         dob = c("2001-03-21", "2000-07-13"),
                         breaks = c(15, 65, 100))
date_to_age_group_custom(date = c("2024-03-27", "2022-11-09"),
                         dob = c("2001-03-21", "2000-07-13"),
                         breaks = c(15, 65, 100),
                         open_last = FALSE)
}
\seealso{
Other functions for creating age groups are
\code{\link{date_to_age_group_year}},
\code{\link{date_to_age_group_multi}},
\code{\link{date_to_age_group_lifetab}},
\code{\link{date_to_age_group_fert}},
\code{\link{date_to_age_group_quarter}},
and \code{\link{date_to_age_group_month}}.
See \code{\link{make_labels_age_group}} for the rules
on constructing labels for age groups.
}
