% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_age_group.R
\name{date_to_age_group_fert}
\alias{date_to_age_group_fert}
\title{Convert dates to age groups when measuring fertility}
\usage{
date_to_age_group_fert(date, dob, break_min = 15, break_max = 50,
  width = 5, recode_up = FALSE, recode_down = FALSE,
  as_factor = TRUE)
}
\arguments{
\item{date}{Dates when births being measured occur.}

\item{dob}{Dates of birth of monthers.}

\item{break_min}{An integer or \code{NULL}.
Defaults to 15.}

\item{break_max}{An integer or \code{NULL}.
Defaults to 50.}

\item{width}{The width in years of the age intervals.
A positive integer defaulting to 5.}

\item{recode_up}{If \code{TRUE}, births to women
aged less than \code{break_min} are treated as occurring to
women in the lowest repoductive age group.}

\item{recode_down}{If \code{TRUE}, births to women
aged \code{break_max} or more are treated as
occurring to women in the highest reproductive
age group.}

\item{as_factor}{Whether the return value is a factor.
Defaults to \code{TRUE}.}
}
\value{
If \code{as_factor} is \code{TRUE}, then the return
value is a factor; otherwise it is a character vector.
The length of the return value equals the length
of \code{date} or the length of \code{dob}, whichever
is greater.
}
\description{
Given the dates when births occur,
and the dates of birth of the mothers,
allocate the births to age groups. These age groups all
have the same width, which is measured in years,
and in an integer.
}
\details{
\code{date} and \code{dob} must have the same length,
unless one of them has length 1, in which case the
length-1 argument is recycled.

\code{break_min} and \code{break_max} specify
the range of ages over which reproduction
is assumed to occur. If, for instance,
\code{break_min} is \code{15} and \code{break_max}
is \code{50}, all births are assumed to
occur to women aged 15 to 49 (inclusive).

Datasets sometimes contain a few births to mothers
younger than the assumed minimum age of reproduction,
or births to mothers older than the assumed maximum age
of reproduction. Demographers often recode such births,
so that ones to unexpectedly young mothers are
treated as occurring just above the minimum age
for reproduction, and ones to unexpectedly old mothers
are treated as occurring just below the maximum
age for reproduction. This recoding can be justified
on the grounds that some of the original ages may have
been misreported, but it also alleviates any problems
with tabulations having small counts at extreme ages.
Recoding of mothers' ages outside the expected range
is controlled by parameters \code{recode_up}
and \code{recode_down}. The default
is for no recoding to occur.
}
\examples{
date_to_age_group_fert(date = c("2024-03-27", "2022-11-09"),
                       dob = c("2001-03-21", "2000-07-13"))

## alternative values for 'width'
date_to_age_group_fert(date = c("2024-03-27", "2022-11-09"),
                       dob = c("2001-03-21", "2000-07-13"),
                       width = 10,
                       break_min = 20)
date_to_age_group_fert(date = c("2024-03-27", "2022-11-09"),
                       dob = c("2001-03-21", "2000-07-13"),
                       width = 1)

## replicate date of birth
date_to_age_group_fert(date = c("2024-03-27", "2022-11-09"),
                       dob = "2001-05-18")

## return non-factor
date_to_age_group_fert(date = c("2024-03-27", "2022-11-09"),
                       dob = "2001-05-18",
                       as_factor = FALSE)

## allow youngest and oldest age groups to be
## set by the data
date_to_age_group_fert(date = c("2052-01-02", "2019-09-22", "2022-10-08"),
                       dob = c("2000-01-01", "2001-03-17", "2010-07-05"),
                       break_min = NULL,
                       break_max = NULL)

## recode ages outside the expected range
date_to_age_group_fert(date = c("2052-01-02", "2019-09-22", "2022-10-08"),
                       dob = c("2000-01-01", "2001-03-17", "2010-07-05"),
                       recode_up = TRUE,
                       recode_down = TRUE)
}
\seealso{
Other functions for creating age groups are
\code{\link{date_to_age_group_year}},
\code{\link{date_to_age_group_multi}},
\code{\link{date_to_age_group_lifetab}},
\code{\link{date_to_age_group_custom}},
\code{\link{date_to_age_group_quarter}},
and \code{\link{date_to_age_group_month}}.
See \code{\link{make_labels_age_group}} for the rules
on constructing labels for age groups.
}
