% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_age_group.R
\name{date_to_age_group_year}
\alias{date_to_age_group_year}
\title{Convert dates to one-year age groups}
\usage{
date_to_age_group_year(date, dob, break_max = 100, open_right = TRUE,
  as_factor = TRUE)
}
\arguments{
\item{date}{Dates of events or measurements.}

\item{dob}{Dates of birth.}

\item{break_max}{An integer, defining the
maximum age group.}

\item{open_right}{Whether the final age group
has no upper limit. Defaults to \code{TRUE}.}

\item{as_factor}{Whether the return value should be a factor.
Defaults to \code{TRUE}.}
}
\value{
If \code{as_factor} is \code{TRUE}, a factor;
otherwise a character vector. The length equals the length
of \code{date} or the length of \code{dob}, whichever
is greater.
}
\description{
Given dates of birth, and dates when events occurred
or measurements were made, derive age groups. These
age groups all have widths of one year, except possibly
the final age group.
}
\details{
\code{date} and \code{dob} are both vectors of class
\code{\link[base]{Date}}, or vectors that can be coerced to class
\code{Date} via function \code{\link[base]{as.Date}}.

\code{date} and \code{dob} must have the same length,
unless one of them has length 1, in which case the
length-1 argument is recycled.

\code{break_max} and \code{open_right} are used to specify
the oldest age group.
When \code{break_max} is non-\code{NULL} and
\code{open_right} is \code{TRUE}, the oldest
age group is \code{[break_max, Inf)} years. When
\code{break_max} is non-\code{NULL} and 
\code{open_right} is \code{FALSE}, the oldest age
group is \code{[break_max-1, break_max)} years.
When \code{break_max} is \code{NULL}, the oldest age
group depends on the highest value in the data, and
\code{open_right} has no effect.

When \code{as_factor} is \code{TRUE} the levels of
the factor include all intermediate age groups,
including age groups that not appear in the data.
}
\examples{
date_to_age_group_year(date = c("2024-03-27", "2022-11-09"),
                       dob = c("2001-03-21", "2000-07-13"))
date_to_age_group_year(date = c("2024-03-27", "2022-11-09"),
                       dob = "2011-05-18")
date_to_age_group_year(date = c("2024-03-27", "2022-11-09"),
                       dob = "2011-05-18",
                       as_factor = FALSE)
date_to_age_group_year(date = "2019-09-22",
                       dob = "1910-01-01")
date_to_age_group_year(date = "2019-09-22",
                       dob = "1910-01-01",
                       break_max = 80)
date_to_age_group_year(date = "2019-09-22",
                       dob = "1910-01-01",
                       break_max = NULL)
date_to_age_group_year(date = "2019-09-22",
                       dob = "1910-01-01",
                       break_max = 110,
                       open_right = FALSE)
}
\seealso{
Other functions for creating age groups are
\code{\link{date_to_age_group_multi}},
\code{\link{date_to_age_group_lifetab}},
\code{\link{date_to_age_group_fert}},
\code{\link{date_to_age_group_custom}},
\code{\link{date_to_age_group_quarter}},
and \code{\link{date_to_age_group_month}}.
Other functions for working with one-year intervals are
\code{\link{date_to_period_year}},
\code{\link{date_to_cohort_year}},
and \code{\link{date_to_triangle_year}}.
}
