% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_age.R
\name{date_to_age_lifetab}
\alias{date_to_age_lifetab}
\title{Convert dates to age groups used in abridged life table}
\usage{
date_to_age_lifetab(date, dob, break_max = 100)
}
\arguments{
\item{date}{Date of death.}

\item{dob}{Dates of birth.
A vector of class \code{\link[base]{Date}},
or a vector that can be coerced to class
\code{Date} using function \code{\link[base]{as.Date}}.}

\item{break_max}{An integer or \code{NULL}.
Defaults to 100.}
}
\value{
A factor with the same length as
\code{date} or \code{dob}, whichever
is longer.
}
\description{
Given dates of death and dates of birth,
allocate the deaths to age groups. These
age groups are the ones typically used in
"abridged" (ie not single-year) life tables: "0", "1-4",
and "5-9", "10-14", and so on up to the
highest age group, which is always open.
}
\details{
\code{date} and \code{dob} must have the same length,
unless one of them has length 1, in which case the
length-1 argument is recycled.

\code{break_max} is used to specify
the oldest age group.
If \code{break_max} is \code{NULL}, the oldest
age group is derived from the data.

The return value is a factor. The levels of this
factor contain all intermediate age groups,
including ones that do not appear in the data.

If \code{date} or \code{dob} contain \code{NA}, then the
levels of the factor created by \code{date_to_age_lifetab}
will also contain \code{NA}.
}
\examples{
date_to_age_lifetab(date = c("2024-03-27",
                             "2022-11-09"),
                    dob = c("2001-03-21",
                            "2000-07-13"))

## set oldest age group to 80+
date_to_age_lifetab(date = "2019-09-22",
                    dob = "1910-01-01",
                    break_max = 80)
}
\seealso{
Other functions for creating
age groups from dates are
\code{\link{date_to_age_year}},
\code{\link{date_to_age_multi}},
\code{\link{date_to_age_births}},
\code{\link{date_to_age_custom}},
\code{\link{date_to_age_quarter}},
and \code{\link{date_to_age_month}}.

Existing labels for life table age groups
can be processed using \code{\link{clean_age}}
and \code{\link{format_age_lifetab}}.
}
