% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_age.R
\name{date_to_age_multi}
\alias{date_to_age_multi}
\title{Convert dates to multi-year age groups}
\usage{
date_to_age_multi(
  date,
  dob,
  width = 5,
  break_min = 0,
  break_max = 100,
  open_last = TRUE
)
}
\arguments{
\item{date}{Dates of events.
A vector of class \code{\link[base]{Date}},
or a vector that can be coerced to class
\code{Date} using function \code{\link[base]{as.Date}}.}

\item{dob}{Dates of birth.
A vector of class \code{\link[base]{Date}},
or a vector that can be coerced to class
\code{Date} using function \code{\link[base]{as.Date}}.}

\item{width}{The width in years of the age intervals.
A positive integer. Defaults to 5.}

\item{break_min}{An integer or \code{NULL}.
Defaults to 0.}

\item{break_max}{An integer or \code{NULL}.
Defaults to 100.}

\item{open_last}{Whether the final age group
has no upper limit. Defaults to \code{TRUE}.}
}
\value{
A factor with the same length as
\code{date} or \code{dob}, whichever
is longer.
}
\description{
Given the dates when events occurred,
and the dates of birth of the people experiencing the events,
allocate the events to age groups. These
age groups all have the same width, except possibly
the final age group. The width is measured in years
and must be an integer.
}
\details{
\code{date} and \code{dob} must have the same length,
unless one of them has length 1, in which case the
length-1 argument is recycled.

If \code{break_min} or \code{break_max} is set to \code{NULL},
rather than to a specific value, then \code{date_to_age_year}
finds the narrowest range that accommodates the data.

When \code{open_last} is \code{TRUE}, the final age group
is "open", which means that it has no upper limit.
(The definition of an open interval in demography is
is different from the definition of an open interval
in mathematics.)

The return value is a factor. The levels of this
factor contain all intermediate age groups,
including ones that do not appear in the data.
}
\examples{
date_to_age_multi(date = c("2024-03-27",
                           "2022-11-09"),
                  dob = c("2001-03-21",
                          "2000-07-13"))

## alternative value for 'width'
date_to_age_multi(date = c("2024-03-27",
                           "2022-11-09"),
                  dob = c("2001-03-21",
                          "2000-07-13"),
                  width = 10)

## final age group closed
date_to_age_multi(date = c("2024-03-27",
                           "2022-11-09"),
                  dob = c("2001-03-21",
                          "2000-07-13"),
                  width = 10,
                  open_last = FALSE)

## allow highest age group to be set by the data
date_to_age_multi(date = c("2024-03-27",
                           "2022-11-09"),
                  dob = c("2001-03-21",
                          "2000-07-13"),
                  width = 10,
                  break_max = NULL)
}
\seealso{
Other functions for creating age groups are
\code{\link{date_to_age_year}},
\code{\link{date_to_age_lifetab}},
\code{\link{date_to_age_births}},
\code{\link{date_to_age_custom}},
\code{\link{date_to_age_quarter}},
and \code{\link{date_to_age_month}}.

Other functions for working with multi-year intervals are
\code{\link{date_to_period_multi}},
\code{\link{date_to_cohort_multi}},
and \code{\link{date_to_triangle_multi}}.

\code{\link{make_labels_age}} describes the rules
for constructing labels for age groups.

\code{\link{plot_date_to_age_multi}} creates plots
that illustrate how \code{date_to_age_multi}
works.
}
