% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_age.R
\name{date_to_age_quarter}
\alias{date_to_age_quarter}
\title{Convert dates to quarter (three-month) age groups}
\usage{
date_to_age_quarter(date, dob)
}
\arguments{
\item{date}{Dates of events.
A vector of class \code{\link[base]{Date}},
or a vector that can be coerced to class
\code{Date} using function \code{\link[base]{as.Date}}.}

\item{dob}{Dates of birth.
A vector of class \code{\link[base]{Date}},
or a vector that can be coerced to class
\code{Date} using function \code{\link[base]{as.Date}}.}
}
\value{
an integer vector with the same length as
\code{date} or \code{dob}, whichever
is longer.
}
\description{
Given dates and dates of birth, create one-quarter
(three-month) age groups.
}
\details{
\code{date} and \code{dob} must have the same length,
unless one of them has length 1, in which case the
length-1 argument is recycled.
}
\examples{
date_to_age_quarter(date = c("2024-03-27",
                             "2022-11-09"),
                    dob = c("2001-03-21",
                            "2000-07-13"))

## replicate date of event
date_to_age_quarter(date = "2024-03-27",
                    dob = c("2001-03-21",
                            "2000-07-13"))

}
\seealso{
The output from \code{date_to_age_quarter}
is typically processed further using function
\code{\link{format_age_quarter}}.

Other functions for creating
age groups from dates are
\code{\link{date_to_age_year}}
and \code{\link{date_to_age_month}}.

Other functions for creating one-quarter
intervals from dates are
\code{\link{date_to_period_quarter}},
\code{\link{date_to_cohort_quarter}},
and \code{\link{date_to_triangle_quarter}}.
}
