% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_age.R
\name{date_to_age_year}
\alias{date_to_age_year}
\title{Convert dates to one-year age groups}
\usage{
date_to_age_year(date, dob)
}
\arguments{
\item{date}{Dates of events.
A vector of class \code{\link[base]{Date}},
or a vector that can be coerced to class
\code{Date} using function \code{\link[base]{as.Date}}.}

\item{dob}{Dates of birth.
A vector of class \code{\link[base]{Date}},
or a vector that can be coerced to class
\code{Date} using function \code{\link[base]{as.Date}}.}
}
\value{
an integer vector with the same length as
\code{date} or \code{dob}, whichever
is longer.
}
\description{
Given the dates when events occurred,
and the dates of birth of the people experiencing the events,
allocate the events to age groups.
All the resulting age groups have widths
of one year.
}
\details{
A person belongs to age group \code{"x"} if that
person was exactly \code{x} years
old at their most recent birthday. For instance, a person
belongs to age group {"5"} if that person had their
5th birthday two days ago, and a person belongs to age
group \code{"0"} if that person had their 0th birthday
(ie was born) three months ago.

\code{date} and \code{dob} must have the same length,
unless one of them has length 1, in which case the
argument with length 1 is recycled.
}
\examples{
date_to_age_year(date = c("2024-03-27",
                          "2022-11-09"),
                 dob = c("2001-03-21",
                         "2000-07-13"))

## replicate date of birth
date_to_age_year(date = c("2024-03-27",
                          "2022-11-09"),
                 dob = "2011-05-18")
}
\seealso{
The output from \code{date_to_age_year}
is often processed further using function
\code{\link{format_age_year}}.

Other functions for creating
age groups from dates are
\code{\link{date_to_age_quarter}}
and \code{\link{date_to_age_month}}.

Other functions for creating one-year
units from dates are
\code{\link{date_to_period_year}},
\code{\link{date_to_cohort_year}},
and \code{\link{date_to_triangle_year}}.
}
