% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_cohort.R
\name{date_to_cohort_custom}
\alias{date_to_cohort_custom}
\title{Convert dates to customized cohorts}
\usage{
date_to_cohort_custom(date, breaks, open_left = TRUE, as_factor = TRUE)
}
\arguments{
\item{date}{Dates of events defining cohorts.}

\item{breaks}{Dates defining starts and ends of cohorts.}

\item{open_left}{Whether the first cohort
has no lower limit. Defaults to \code{TRUE}.}

\item{as_factor}{Whether the return value is a factor.
Defaults to \code{TRUE}.}
}
\value{
If \code{as_factor} is \code{TRUE}, then the return
value is a factor; otherwise it is a character vector.
The return value has the same length as \code{date}.
}
\description{
Allocate dates of events
to cohorts with varying widths, though all widths
are measured in whole years.
The events are typically, but not always, births.
}
\details{
The interface for \code{date_to_cohort_custom} is the similar
to that of \code{\link{date_to_period_custom}}, except that
\code{date_to_cohort_custom} also has arguments \code{break_min}
and \code{open_left}.

\code{date} is a vector of class \code{\link[base]{Date}},
or can be coerced to class \code{Date}
via function \code{\link[base]{as.Date}}.

\code{breaks} is also vector of class \code{\link[base]{Date}},
or can be coerced to to one. \code{breaks} is
used to define the points where each cohort starts and finishes.
The dates in \code{breaks} must all be the first day of
the same month of the year. For instance,
\code{breaks} could consist of
the values \code{"2010-01-01"} and \code{"2017-01-01"},
but not \code{"2010-01-01"} and \code{"2017-01-02"},
or \code{"2010-01-01"} and \code{"2017-02-01"}.

The definition of the first cohort depends on the
value of \code{open_left}. If \code{open_left} is
is \code{TRUE} (the default), then the first cohort consists of
all dates before the first break. Setting
\code{open_left} to \code{TRUE} is essentially
the same as having no upper limit on the oldest age group.
If \code{open_left} is \code{FALSE}, then the first cohort
is defined by the first two dates.

When \code{as_factor} is \code{TRUE} the levels of
the factor include all intermediate cohorts,
including cohorts that not appear in the data.
}
\examples{
## cohorts start on 1 January
date_to_cohort_custom(date = c("2024-03-27",
                               "2018-11-09",
                               "2020-05-13",
                               "2021-03-02"),
                      breaks = c("2020-01-01",
                                 "2021-01-01",
                                 "2026-01-01"))

## cohorts start on 1 March
date_to_cohort_custom(date = c("2024-03-27",
                               "2018-11-09",
                               "2020-05-13",
                               "2021-03-02"),
                      breaks = c("2020-03-01",
                                 "2021-03-01",
                                 "2026-03-01"))

## first cohort is closed
date_to_cohort_custom(date = c("2024-03-27",
                               "2018-11-09",
                               "2020-05-13",
                               "2021-03-02"),
                      breaks = c("2018-01-01",
                                 "2021-01-01",
                                 "2026-01-01"),
                      open_left = FALSE)


## return non-factor
date_to_cohort_custom(date = c("2024-03-27",
                               "2018-11-09",
                               "2020-05-13",
                               "2021-03-02"),
                      breaks = c("2020-01-01",
                                 "2021-01-01",
                                 "2026-01-01"),
                      as_factor = FALSE)
}
\seealso{
Other functions for creating cohorts are
\code{\link{date_to_cohort_year}},
\code{\link{date_to_cohort_multi}},
\code{\link{date_to_cohort_quarter}},
and \code{\link{date_to_cohort_month}}.
Other functions for working with customised intervals are
\code{\link{date_to_age_group_custom}},
and \code{\link{date_to_period_custom}}.
See \code{\link{make_labels_period}} for the rules
on constructing labels for periods and cohorts.
}
