% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_cohort.R
\name{date_to_cohort_month}
\alias{date_to_cohort_month}
\title{Convert dates to month cohorts}
\usage{
date_to_cohort_month(date, break_min = NULL, open_left = NULL,
  as_factor = TRUE)
}
\arguments{
\item{date}{Dates of events defining cohorts.}

\item{break_min}{An integer or \code{NULL}.}

\item{open_left}{Whether the first cohort
has no lower limit. If \code{break_min} is \code{NULL},
then \code{open_left} is ignored. If \code{break_min} is
non-\code{NULL}, then \code{open_left} defaults to
\code{TRUE}.}

\item{as_factor}{Whether the return value is a factor.
Defaults to \code{TRUE}.}
}
\value{
If \code{as_factor} is \code{TRUE}, then the return
value is a factor; otherwise it is a character vector.
The return value has the same length as \code{date}.
}
\description{
Allocate dates of events (typically, but not always, births)
to cohorts of length one month.
}
\details{
The interface for \code{date_to_cohort_month} is the similar
to that of \code{\link{date_to_period_month}}, except that
\code{date_to_cohort_month} also has arguments \code{break_min}
and \code{open_left}.

\code{date} is a vector of class \code{\link[base]{Date}},
or can be coerced to class \code{Date}
via function \code{\link[base]{as.Date}}.

\code{break_min} is a single value of
class \code{\link[base]{Date}}, or a value that
can be coerced to class \code{Date}
via function \code{\link[base]{as.Date}}.
\code{break_min} must be the first day of
a month, e.g \code{"2020-01-01"} or
\code{"2020-04-01"}.

Supplying a value for \code{break_min} is essentially equivalent
to supplying a value for \code{break_max} in
\code{\link{date_to_age_group_month}}, in that it is a way of
defining how the oldest ages are handled. If
\code{open_left} is \code{TRUE}, then the first cohort
is \code{(-Inf, break_min)}; otherwise the
first cohort is \code{[break_min, break_min - 1)}.

When \code{as_factor} is \code{TRUE} the levels of
the factor include all intermediate cohorts,
including cohorts that not appear in the data.
}
\examples{
date_to_cohort_month(date = c("2024-03-27",
                              "2022-11-09"))

## Specify oldest cohort, with open_left
## at default value of TRUE
date_to_cohort_month(date = c("2024-03-27",
                              "2019-08-22",
                              "2022-11-09"),
                       break_min = "2020-01-01")

## Specify oldest cohort, with open_left = FALSE
date_to_cohort_month(date = c("2024-03-27",
                              "2019-08-22",
                              "2022-11-09"),
                    break_min = "2015-07-01",
                    open_left = FALSE)

## return non-factor
date_to_cohort_month(date = c("2024-03-27",
                              "2022-11-09"),
                    as_factor = FALSE)
}
\seealso{
Other functions for creating cohorts are
\code{\link{date_to_cohort_year}},
\code{\link{date_to_cohort_multi}},
\code{\link{date_to_cohort_custom}},
and \code{\link{date_to_cohort_quarter}}.
Other functions for working with one-month intervals are
\code{\link{date_to_age_group_month}},
\code{\link{date_to_period_month}},
and \code{\link{date_to_triangle_month}}.
See \code{\link{make_labels_period_month}} for the rules
on constructing labels for periods and cohorts.
}
