% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_cohort.R
\name{date_to_cohort_multi}
\alias{date_to_cohort_multi}
\title{Convert dates to multi-year cohorts}
\usage{
date_to_cohort_multi(
  date,
  width = 5,
  origin = 2000,
  month_start = "Jan",
  break_min = NULL,
  open_first = NULL
)
}
\arguments{
\item{date}{Dates of events defining cohorts.
A vector of class \code{\link[base]{Date}},
or a vector that can be coerced to class \code{Date}
via function \code{\link[base]{as.Date}}.}

\item{width}{The length, in whole years, of the cohorts.
Defaults to 5.}

\item{origin}{An integer. Defaults to 2000.
Ignored if value supplied for \code{break_min}.}

\item{month_start}{An element of \code{\link[base]{month.name}},
or \code{\link[base]{month.abb}}. Cohorts start on
the first day of this month.}

\item{break_min}{The start date of the first cohort,
or \code{NULL} (the default.)
If non-\code{NULL}, \code{break_min} can be a single value of
class \code{\link[base]{Date}}, or a value that
can be coerced to class \code{Date}
via function \code{\link[base]{as.Date}};
in either case, it must be the first day of
a month.}

\item{open_first}{Whether the first cohort
has no lower limit. If \code{break_min} is non-\code{NULL},
then \code{open_first} defaults to \code{TRUE};
otherwise it defaults to \code{FALSE}.}
}
\value{
A factor with the same length as \code{date}.
}
\description{
Identify cohorts, based on dates of births or other events.
The cohorts all have the same width, which by default is 5 years.
}
\details{
The interface for \code{date_to_cohort_multi} is the similar
to that of \code{\link{date_to_period_multi}}, except that
\code{date_to_cohort_multi} also has arguments \code{break_min}
and \code{open_first}.

Supplying a date for \code{break_min} defines
the first cohort. It is similar
to supplying a value for \code{break_max} in
\code{\link{date_to_age_multi}}. If
\code{open_first} is \code{TRUE}, then \code{break_min}
forms the upper limit for the first cohort, and there is no lower
limit. If \code{open_first} is \code{FALSE}, then
\code{break_min} forms the lower limit for the first cohort.
\code{open_first} defaults to \code{TRUE}
if a value for \code{break_min} is supplied,
and to \code{FALSE} otherwise.

If \code{break_min} is specified, then
cohorts start on the day and month supplied;
otherwise they start on first day of \code{month_start}.
Cohorts end \code{widths}-years-minus-one-day later.
The default value for \code{month_start} is \code{"January"},
so when \code{break_min} is not specified,
cohorts by default start on 1 January and
end on 31 December. \code{month_start} can be a
full month name or an abbreviation.

The location of the periods can be shifted
by using different values for \code{origin},
unless a value for \code{break_min} has been
supplied, in which case \code{origin} is ignored.

The return value is a factor. The levels of this
factor contain all intermediate cohorts,
including cohorts that do not appear in the data.

If \code{date} contains \code{NA}, then the
levels of the factor created by \code{date_to_cohort_multi}
also contain \code{NA}.
}
\examples{
date_to_cohort_multi(date = c("2024-03-27",
                              "2018-11-09",
                              "2021-03-02"))

## width is 10
date_to_cohort_multi(date = c("2024-03-27",
                              "2018-11-09",
                              "2021-03-02"),
                     width = 5)

## origin is 2001
date_to_cohort_multi(date = c("2024-03-27",
                              "2018-11-09",
                              "2021-03-02"),
                     origin = 2001)

## starts on 1 July rather than 1 January
date_to_cohort_multi(date = c("2024-03-27",
                              "2018-11-09",
                              "2021-03-02"),
                     origin = 2001,
                     month_start = "Jul")

## specify oldest cohort, with 'open_first'
## at default value of 'TRUE'
date_to_cohort_multi(date = c("2024-03-27",
                              "2019-08-22",
                              "2022-11-09"),
                     break_min = "2015-01-01")

## specify oldest cohort, with 'open_first'
## set to 'FALSE'
date_to_cohort_multi(date = c("2024-03-27",
                              "2019-08-22",
                              "2022-11-09"),
                     break_min = "2015-07-01",
                     open_first = FALSE)
}
\seealso{
Other functions for creating
cohorts from dates are
\code{\link{date_to_cohort_year}},
\code{\link{date_to_cohort_custom}},
\code{\link{date_to_cohort_quarter}},
and \code{\link{date_to_cohort_month}}.

Other functions for creating multi-year
intervals from dates are
\code{\link{date_to_age_multi}},
\code{\link{date_to_period_multi}},
and \code{\link{date_to_triangle_multi}}.

Existing labels for multi-year cohorts
can be processed using \code{\link{clean_age}}
and \code{\link{format_cohort_multi}}.
}
