% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_cohort.R
\name{date_to_cohort_quarter}
\alias{date_to_cohort_quarter}
\title{Convert dates to quarter (three-month) cohorts}
\usage{
date_to_cohort_quarter(date)
}
\arguments{
\item{date}{Dates of events defining cohorts (typically births).
A vector of class \code{\link[base]{Date}},
or a vector that can be coerced to class \code{Date}
via function \code{\link[base]{as.Date}}.}
}
\value{
A character vector with the same length as \code{date}.
}
\description{
Assign dates to one-quarter (three-month) cohorts.
Quarters are defined as follows:
\tabular{lll}{
  \strong{Quarter} \tab \strong{Start} \tab \strong{End} \cr
  Q1 \tab 1 January \tab 31 March \cr
  Q2 \tab 1 April \tab 30 June \cr
  Q3 \tab 1 July \tab 30 September \cr
  Q4 \tab 1 October \tab 31 December
}
}
\examples{
date_to_cohort_quarter(date = c("2024-03-27",
                                "2022-11-09",
                                "2023-05-11"))
}
\seealso{
The output from \code{date_to_cohort_quarter}
is often processed further using \code{\link{format_cohort_quarter}}.

Other functions for creating cohorts from dates are
\code{\link{date_to_cohort_quarter}} and
and \code{\link{date_to_cohort_month}}.

Other functions for creating one-quarter units
from dates are
\code{\link{date_to_age_quarter}},
\code{\link{date_to_period_quarter}},
and \code{\link{date_to_triangle_quarter}}.
The interface for \code{date_to_cohort_quarter} is identical
to that of \code{\link{date_to_period_quarter}}.
}
