% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_cohort.R
\name{date_to_cohort_year}
\alias{date_to_cohort_year}
\title{Convert dates to one-year cohorts}
\usage{
date_to_cohort_year(date, month_start = "Jan", label_year_start = TRUE)
}
\arguments{
\item{date}{Dates of events defining cohorts (typically births).
A vector of class \code{\link[base]{Date}},
or a vector that can be coerced to class \code{Date}
via function \code{\link[base]{as.Date}}.}

\item{month_start}{An element of \code{\link[base]{month.name}},
or \code{\link[base]{month.abb}}. Cohorts start on
the first day of this month.}

\item{label_year_start}{Logical. Whether to label a cohort
by the calendar year at the beginning of the cohort
or the calendar year at the end. Defaults to \code{TRUE}.}
}
\value{
An integer vector with the same length as \code{date}.
}
\description{
Assign dates to one-year cohorts.
}
\details{
Cohorts start on the first day of \code{month_start},
and end one-year-minus-one-day later.
The default value for \code{month_start} is \code{"Jan"},
so cohorts by default start on 1 January and
end on 31 December. \code{month_start} can be a
full month name or an abbreviation.

If a cohort starts on 1 January, then the first day and last day
of the cohort belong to the same calendar year.
But if a cohort starts on any other day, then
the first day belongs to one calendar
year and the last day belongs to the next calendar year.
For instance, if a cohort extends from
1 July 2000 to 30 June 2001, then the first
day belongs to calendar year 2000, and the last
day belongs to calendar year 2001. Some people
use the first calendar year to label such cohorts,
and others use the last calendar year.
For instance, if a cohort extends from
1 July 2000 to 30 June 2001, some people
label the cohort \code{"2000"}, and
others label it \code{"2001"}. Function
\code{date_to_cohort_year} by default uses
the start year. To use the end year, set
\code{label_year_start} to \code{FALSE}.
}
\examples{
date_to_cohort_year(date = c("2024-03-27", "2022-11-09"))

## starts on 1 July rather than 1 January
date_to_cohort_year(date = c("2024-03-27", "2022-11-09"),
                    month_start = "Jul")

## periods starts on 1 July, rather than 1 January, and uses
## calendar year at end, rather than beginning, for the label
date_to_cohort_year(date = c("2024-03-27", "2022-11-09"),
                    month_start = "Jul",
                    label_year_start = FALSE)
}
\seealso{
The output from \code{date_to_cohort_year}
is ofen processed further using function
\code{\link{format_cohort_year}}.

Other functions for creating
cohorts from dates are
\code{\link{date_to_cohort_quarter}} and
and \code{\link{date_to_cohort_month}}.

Other functions for creating one-year
intervals from dates are
\code{\link{date_to_age_year}},
\code{\link{date_to_period_year}},
and \code{\link{date_to_triangle_year}}.
The interface for \code{date_to_cohort_year} is identical
to that of \code{\link{date_to_period_year}}.
}
