% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_period.R
\name{date_to_period_custom}
\alias{date_to_period_custom}
\title{Convert dates to customized periods}
\usage{
date_to_period_custom(date, breaks, as_factor = TRUE)
}
\arguments{
\item{date}{Dates of events or measurements.}

\item{breaks}{Dates defining starts and ends of periods.}

\item{as_factor}{Whether the return value is a factor.
Defaults to \code{TRUE}.}
}
\value{
If \code{as_factor} is \code{TRUE}, then the return
value is a factor; otherwise it is a character vector.
The return value has the same length as \code{date}.
}
\description{
Allocate dates to periods with varying widths,
though all widths are measured in whole years.
}
\details{
\code{date} is a vector of class \code{\link[base]{Date}},
or can be coerced to class \code{Date}
via function \code{\link[base]{as.Date}}.

\code{breaks} is also vector of class \code{\link[base]{Date}},
or can be coerced to to one. \code{breaks} is
used to define the points where each period starts and finishes.
The dates in \code{breaks} must all be the first day of
the same month of the year. For instance,
\code{breaks} could consist of
the values \code{"2010-01-01"} and \code{"2017-01-01"},
but not \code{"2010-01-01"} and \code{"2017-01-02"},
or \code{"2010-01-01"} and \code{"2017-02-01"}.
The final period finishes
one day before the final break.
If, for instance, \code{breaks} has values \code{"2010-01-01"},
\code{"2017-01-01"}, and \code{"2021-01-01"}, then the first period
starts on 1 January 2010 and ends on 31 December 2016,
and the second period starts on 1 January 2017 and ends
on 31 December 2020.

When \code{as_factor} is \code{TRUE} the levels of
the factor include all intermediate periods,
including periods that not appear in the data.
}
\examples{
## periods start on 1 January
date_to_period_custom(date = c("2024-03-27",
                              "2018-11-09",
                              "2021-03-02"),
                      breaks = c("2000-01-01",
                                 "2019-01-01",
                                 "2026-01-01"))

## periods start on 1 March
date_to_period_custom(date = c("2024-03-27",
                              "2018-11-09",
                              "2021-03-02"),
                      breaks = c("2000-03-01",
                                 "2019-03-01",
                                 "2026-03-01"))

## return non-factor
date_to_period_custom(date = c("2024-03-27",
                               "2018-11-09",
                               "2021-03-02"),
                      breaks = c("2000-01-01",
                                 "2019-01-01",
                                 "2026-01-01"),
                      as_factor = FALSE)
}
\seealso{
Other functions for creating periods are
\code{\link{date_to_period_year}},
\code{\link{date_to_period_multi}},
\code{\link{date_to_period_quarter}},
and \code{\link{date_to_period_month}}.

Other functions for working with customized intervals are
\code{\link{date_to_age_custom}},
and \code{\link{date_to_cohort_custom}}.

\code{\link{make_labels_period}} describes the rules
for constructing labels for periods.

\code{\link{plot_date_to_period_custom}} creates plots
that illustrate how \code{date_to_period_custom}
works.
}
