% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_period.R
\name{date_to_period_month}
\alias{date_to_period_month}
\title{Convert dates to one-month periods}
\usage{
date_to_period_month(date)
}
\arguments{
\item{date}{Dates of events or measurements.
A vector of class \code{\link[base]{Date}},
or a vector than can be coerced to class \code{Date}
via function \code{\link[base]{as.Date}}.}
}
\value{
A factor with the same length as \code{date}.
}
\description{
Assign dates to one-month periods.
}
\examples{
date_to_period_month(date = c("2024-03-27",
                              "2020-01-03",
                              "2022-11-09"))
}
\seealso{
The output from \code{date_to_period_month}
is often processed further using \code{\link{format_period_month}}.

Other functions for creating periods from dates are
\code{\link{date_to_period_year}} and
and \code{\link{date_to_period_quarter}}.

Other functions for creating one-month units
from dates are
\code{\link{date_to_age_month}},
\code{\link{date_to_cohort_month}},
and \code{\link{date_to_triangle_month}}.
The interface for \code{date_to_period_month} is identical
to that of \code{\link{date_to_cohort_month}}.
}
