% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_period.R
\name{date_to_period_multi}
\alias{date_to_period_multi}
\title{Convert dates to multi-year periods}
\usage{
date_to_period_multi(date, width = 5, origin = 2000, month_start = "Jan")
}
\arguments{
\item{date}{Dates of events or measurements.}

\item{width}{The length, in whole years, of the periods.
Defaults to 5.}

\item{origin}{An integer. Defaults to 2000.}

\item{month_start}{An element of \code{\link[base]{month.name}},
or \code{\link[base]{month.abb}}. Each period starts on
the first day of this month.}
}
\value{
A factor with the same length as \code{date}.
}
\description{
Allocate dates to multi-year periods. The periods all have
the same width, which by default is 5 years.
}
\details{
\code{date} is a vector of class \code{\link[base]{Date}},
or can be coerced to class \code{Date}
via function \code{\link[base]{as.Date}}.

Periods are \code{width} years long.
They start on the first day of \code{month_start},
and end \code{width}-years-minus-one-day later.
The default value for \code{month_start} is \code{"Jan"},
so periods by default start on 1 January and
end on 31 December. \code{month_start} can be a
full month name or an abbreviation.

The location of the periods can be shifted
by using different values for \code{origin}.

The return value is a factor.
The levels of this factor include all intermediate periods,
including periods that do not appear in the data.

If \code{date} contains \code{NA}, then the
levels of the factor created by \code{date_to_period_multi}
also contain \code{NA}.
}
\examples{
date_to_period_multi(date = c("2024-03-27",
                              "2018-11-09",
                              "2021-03-02"))

## width is 10
date_to_period_multi(date = c("2024-03-27",
                              "2018-11-09",
                              "2021-03-02"),
                     width = 5)

## origin is 2001
date_to_period_multi(date = c("2024-03-27",
                              "2018-11-09",
                              "2021-03-02"),
                     origin = 2001)

## period starts on 1 July, not 1 January
date_to_period_multi(date = c("2024-03-27",
                              "2018-11-09",
                              "2021-03-02"),
                     origin = 2001,
                     month_start = "Jul")
}
\seealso{
Other functions for creating periods are
\code{\link{date_to_period_year}},
\code{\link{date_to_period_custom}},
\code{\link{date_to_period_quarter}},
and \code{\link{date_to_period_month}}.

Other functions for working with multi-year intervals are
\code{\link{date_to_age_multi}},
\code{\link{date_to_cohort_multi}},
and \code{\link{date_to_triangle_multi}}.

\code{\link{make_labels_period}} describes the rules
for constructing labels for periods.

\code{\link{plot_date_to_period_multi}} creates plots
that illustrate how \code{date_to_period_multi}
works.
}
