% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_period.R
\name{date_to_period_quarter}
\alias{date_to_period_quarter}
\title{Convert dates to quarter periods}
\usage{
date_to_period_quarter(date, as_factor = TRUE)
}
\arguments{
\item{date}{Dates of events or measurements.}

\item{as_factor}{Whether the return value is a factor.
Defaults to \code{TRUE}.}
}
\value{
If \code{as_factor} is \code{TRUE}, then the return
value is a factor; otherwise it is a character vector.
The return value has the same length as \code{date}.
}
\description{
Allocate dates to periods with periods one quarter
(ie three month) long. Q1 (the first quarter)
starts on 1 January
and ends on 31 March; Q2 starts on 1 April and ends on
30 June; Q3 starts on 1 July and ends on 30 September;
Q4 starts on 1 October and ends on 31 December.
}
\details{
\code{date} is a vector of class \code{\link[base]{Date}},
or can be coerced to class \code{Date}
via function \code{\link[base]{as.Date}}.

When \code{as_factor} is \code{TRUE} the levels of
the factor include all intermediate periods,
including periods that not appear in the data.
}
\examples{
date_to_period_quarter(date = c("2024-03-27",
                                "2020-01-03",
                                "2022-11-09"))
## return non-factor
date_to_period_quarter(date = c("2024-03-27",
                                "2020-01-03",
                                "2022-11-09"),
                       as_factor = FALSE)
}
\seealso{
Other functions for creating periods are
\code{\link{date_to_period_year}},
\code{\link{date_to_period_multi}},
\code{\link{date_to_period_custom}},
and \code{\link{date_to_period_month}}.
Other functions for working with quarter intervals are
\code{\link{date_to_age_group_quarter}},
and \code{\link{date_to_cohort_quarter}},
and \code{\link{date_to_triangle_quarter}}.
See \code{\link{make_labels_period_quarter}} on the rules
for constructing labels for quarter periods.
}
