% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_triangle.R
\name{date_to_triangle_fert}
\alias{date_to_triangle_fert}
\title{Convert dates to Lexis triangles used when measuring fertility}
\usage{
date_to_triangle_fert(date, dob, width = 5, break_min = 15,
  break_max = 50, recode_up = FALSE, recode_down = FALSE,
  origin = 2000, month_start = "Jan", as_factor = TRUE)
}
\arguments{
\item{date}{Dates when births being measured occur.}

\item{dob}{Dates of birth of monthers.}

\item{width}{The width in years of the periods and
age groups. A positive integer defaulting to 5.}

\item{break_min}{An integer or \code{NULL}.
Defaults to 15.}

\item{break_max}{An integer or \code{NULL}.
Defaults to 50.}

\item{recode_up}{If \code{TRUE}, births to women
aged less than \code{break_min} are treated as occurring to
women in the lowest repoductive age group.}

\item{recode_down}{If \code{TRUE}, births to women
aged \code{break_max} or more are treated as
occurring to women in the highest reproductive
age group.}

\item{origin}{An integer. Defaults to 2000.}

\item{month_start}{An element of \code{\link[base]{month.name}},
or \code{\link[base]{month.abb}}. The period starts on
the first day of this month.}

\item{as_factor}{Whether the return value is a factor.
Defaults to \code{TRUE}.}
}
\value{
If \code{as_factor} is \code{TRUE}, then the return
value is a factor; otherwise it is a character vector.
The return value has the same length as \code{date}.
}
\description{
Given dates when births occurred, together with the dates of birth
All the Lexis triangles have the same widths.
}
\details{
An birth occurring during period \code{[t, t+n)} to
a woman in age group \code{[a, a+n)} belongs to an
upper Lexis triangle if the woman was already in age group
\code{[a, a+n)} at time \code{t}, and belongs to a
lower Lexis triangle if the woman joins age group
\code{[a, a+n)} during period \code{[t, t+n)}.

\code{date} and \code{dob} are both vectors of class
\code{\link[base]{Date}}, or vectors that can be coerced to class
\code{Date} via function \code{\link[base]{as.Date}}.
\code{date} and \code{dob} must have the same length,
unless one of them has length 1, in which case the
length-1 argument is recycled.

Periods are \code{width} years long.
They start on the first day of \code{month_start},
and end \code{width}-years-minus-one-day later.
The default value for \code{month_start} is \code{"Jan"},
so periods by default start on 1 January and
end on 31 December. \code{month_start} can be a
full month name or an abbreviation.

The location of the periods can be shifted
by using different values for \code{origin}.

\code{break_min} and \code{break_max} specify
the range of ages over which reproduction
is assumed to occur. If, for instance,
\code{break_min} is \code{15} and \code{break_max}
is \code{50}, all births are assumed to
occur to women aged 15 to 49 (inclusive).

Datasets sometimes contain a few births to mothers
younger than the assumed minimum age for reproduction,
or to mothers older than the assumed maximum age
for reproduction. Demographers often recode such births,
so that births to unexpectedly young mothers are
treated as occurring just above the minimum age
for reproduction, and births to unexpectedly old mothers
are treated as occurring just below the maximum
age for reproduction. This recoding can be justified
on the grounds that some of the original ages may have
been misreported, but it also alleviates any problems
with tabulations having small counts at extreme ages.
Recoding of mothers' ages outside the expected range
is controlled by parameters \code{recode_up}
and \code{recode_down}. The default
is for no recoding to occur.

The allocation of a birth to a Lexis triangles becomes
tricky when children and mothers have the same birthday,
and when their dates of birth are some multiple of
\code{width} years apart. Always allocating such
births to the lower triangle, or always allocating
them to the upper triangle, would lead to a slight
imbalance between upper and lower triangles. Instead,
if the birthdays match and it is first day of the month,
the birth is allocated to \code{"Lower"}; if the birthdays match and
it is the second day of the month, the birth is allocated to
\code{"Upper"}; if the birthdays match and it is the third day
of the month, the birth is allocated to \code{"Lower"}; and so on.
See below for an example.

When \code{as_factor} is \code{TRUE}, the levels of
the factor includes both \code{"Lower"} and
\code{"Upper"}, even when they do not both appear
in the data.
}
\examples{
date_to_triangle_fert(date = c("2024-03-27", "2022-11-09"),
                      dob = c("2001-03-21", "2000-07-13"))

## alternative values for 'width'
date_to_triangle_fert(date = c("2024-03-27", "2022-11-09"),
                      dob = c("2001-03-21", "2000-07-13"),
                      width = 10,
                      break_min = 20)
date_to_triangle_fert(date = c("2024-03-27", "2022-11-09"),
                      dob = c("2001-03-21", "2000-07-13"),
                      width = 1)

## births of children and mothers occur on
## same month and day of month, and dates
## are 4 * 5 years apart
date_to_triangle_fert(date = c("2020-03-01",
                               "2020-03-02",
                               "2020-03-03",
                               "2020-03-04",
                               "2020-03-05"),
                      dob = c("2000-03-01",
                              "2000-03-02",
                              "2000-03-03",
                              "2000-03-04",
                              "2000-03-05"))

## return non-factor
date_to_triangle_year(date = c("2024-03-27",
                               "2022-11-09"),
                      dob = "1996-03-01",
                      as_factor = FALSE)
}
\seealso{
Other functions for creating Lexis triangles are
\code{\link{date_to_triangle_year}},
\code{\link{date_to_triangle_multi}},
\code{\link{date_to_triangle_quarter}},
and \code{\link{date_to_triangle_month}}.
\code{date_to_triangle_fert} is typically used in combination with
\code{\link{date_to_age_group_year}}
and \code{\link{date_to_period_year}}.
}
