% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_triangle.R
\name{date_to_triangle_multi}
\alias{date_to_triangle_multi}
\title{Convert dates to multi-year Lexis triangles}
\usage{
date_to_triangle_multi(
  date,
  dob,
  width = 5,
  break_max = 100,
  open_last = TRUE,
  origin = 2000,
  month_start = "Jan",
  as_factor = TRUE
)
}
\arguments{
\item{date}{Dates of events.}

\item{dob}{Dates of birth.}

\item{width}{The width in years of the periods and
age groups. A positive integer defaulting to 5.}

\item{break_max}{An integer or \code{NULL}.
Defaults to 100.}

\item{open_last}{Whether the final age group
has no upper limit. Defaults to \code{TRUE}.}

\item{origin}{An integer. Defaults to 2000.}

\item{month_start}{An element of \code{\link[base]{month.name}},
or \code{\link[base]{month.abb}}. The period starts on
the first day of this month.}

\item{as_factor}{Whether the return value is a factor.
Defaults to \code{TRUE}.}
}
\value{
If \code{as_factor} is \code{TRUE}, then the return
value is a factor; otherwise it is a character vector.
The return value has the same length as \code{date}.
}
\description{
Based on an age-time plan where age groups and periods have
identical widths, use dates of events and dates of birth
to allocate events to Lexis triangles.
}
\details{
The allocation of an event to a Lexis triangle depends
on the timing of the event and on the timing of change
in age group. A person moves up into a new age group
once during each period. An event is allocated
to an upper Lexis triangle if the event occurs
before the move to the new age group.
An event is allocated to a lower Lexis triangle
if the event occurs with or after the move to the
new age group.

Consider, for instance, events occurring to a person
who was born on 28 March 2001, in an age-time plan
where periods start on 1 January 2000, 1 January 2005, etc.
\code{date_to_triangle_multi} allocates events
to Lexis triangles as follows:
\tabular{ll}{
  \emph{Date of event} \tab \emph{Lexis triangle} \cr
  28 March 2001 to 31 December 2004 \tab \code{"Lower"} \cr
  1 January 2005 to 27 March 2006 \tab \code{"Upper"} \cr
  28 March 2006 to 31 December 2009 \tab \code{"Lower"} \cr
  1 January 2010 to 27 March 2011 \tab \code{"Upper"} \cr
  etc \tab etc
}

\code{date} and \code{dob} are both vectors of class
\code{\link[base]{Date}}, or vectors that can be coerced to class
\code{Date} via function \code{\link[base]{as.Date}}.

\code{date} and \code{dob} must have the same length,
unless one of them has length 1, in which case the
length-1 argument is recycled.

\code{break_max} and \code{open_last} are used to specify
the oldest age group.
When \code{break_max} is non-\code{NULL} and
\code{open_last} is \code{TRUE}, the oldest
age group is \code{[break_max, Inf)} years. When
\code{break_max} is non-\code{NULL} and 
\code{open_last} is \code{FALSE}, the oldest age
group is \code{[break_max-width, break_max)} years.

If \code{break_max} is \code{NULL}, then
\code{date_to_triangle_multi} derives a value for
\code{break_max} based on the highest age in the data
and on the value for \code{open_last}.

Periods start on the first day of \code{month_start},
and end \code{width}-years-minus-one-day later.
The default value for \code{month_start} is \code{"Jan"},
so periods by default start on 1 January and
end on 31 December.

The location of the periods can be shifted
by using different values for \code{origin}.

When \code{as_factor} is \code{TRUE}, the levels of
the factor includes both \code{"Lower"} and
\code{"Upper"}, even when \code{"Lower"}
and \code{"Upper"} do not both appear
in the data.
}
\examples{
date_to_triangle_multi(date = c("2027-03-27",
                                "2022-11-09"),
                       dob = "2010-05-12")

## width is 10
date_to_triangle_multi(date = c("2027-03-27",
                                "2022-11-09"),
                       dob = "2010-05-12",
                       width = 10)


## starts on 1 July rather than 1 January
date_to_triangle_multi(date = c("2027-03-27",
                                "2022-11-09"),
                       dob = "2010-05-12",
                       month_start = "Jul")

## periods start in 2021, 2026, 2031, etc
date_to_triangle_multi(date = c("2027-03-27",
                                "2022-11-09"),
                       dob = "2010-05-12",
                       origin = 2001)

## open age group starts at 10 years
date_to_triangle_multi(date = c("2027-03-27",
                                "2022-11-09"),
                       dob = "2003-05-12",
                       break_max = 10)

## return non-factor
date_to_triangle_multi(date = c("2024-03-27",
                                "2022-11-09"),
                       dob = "2012-03-01",
                       as_factor = FALSE)
}
\seealso{
Other functions for creating Lexis triangles are
\code{\link{date_to_triangle_year}},
\code{\link{date_to_triangle_births}},
\code{\link{date_to_triangle_quarter}},
and \code{\link{date_to_triangle_month}}.

\code{date_to_triangle_multi} is typically used in combination with
\code{\link{date_to_age_multi}} and
\code{\link{date_to_period_multi}}.

\code{\link{plot_date_to_triangle_multi}} creates plots
that illustrate how \code{date_to_triangle_multi}
works.
}
