% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtabs.R
\name{dtabs_survey}
\alias{dtabs_survey}
\title{Cross-tabulation of binary outcomes
in weighted survey data}
\source{
Ghitza Y and Gelman A. 2013. Deep interactions with MRP:
Election turnout and voting patterns among small electoral subgroups.
\emph{American Journal of Political Science}, 57(3): 762-776.

For an alternative way of calculating effective
numbers of responses, see Chen C, Wakefield J, and Lumley T.
2014. The use of sampling weights
in Bayesian hierarchical models for small area estimation.
\emph{Spatial and Spatio-temporal Epidemiology}. 11: 33-43.
}
\usage{
dtabs_survey(
  data = parent.frame(),
  formula,
  wt,
  method = c("ghitza_med", "ghitza_mean"),
  na_rm = FALSE
)
}
\arguments{
\item{data}{A data frame or matrix.}

\item{formula}{A \code{\link[stats]{formula}}.
The formula must contain a response, and that
response must be a binary varibale.}

\item{wt}{A vector of weights. Required.}

\item{method}{The method used to calculate
the overall design effect. Defaults to \code{"ghitza_med"}.}

\item{na_rm}{Whether to remove \code{NA}s from the
response before tabulating it.
Defaults to \code{FALSE}.}
}
\value{
A list of two arrays. The first
array is named \code{"successes"}
and the second is named \code{"trials"}.
}
\description{
Tabulate binary outcomes, for combinations of
cross-classifying variables, from individual-level
survey data.
}
\details{
The survey has a complex
design, involving techniques such as clustering
and stratification. The complex design implies
that different members
of the target population have different probabilities
of being included in the survey. These differential
inclusion probabilities complicate subsequent analyses,
including the calculation of means and variances.

However, by exploiting information contained in the survey
weights, it is possible to calculate the "effective"
number of trials and successes for each
combination of the cross-classifying variables.
These effective numbers can be analysed as if
they were generated by simple random sampling.

The effective number of trials and successes equals the
actual number of trials and successes
divided by the "design effect".
The method that \code{dtab_survey} uses for calculating
design effects implies that they are always
greater than or equal to one. The larger the design
effect, the less information the survey contains,
releative the the information that would be contained
in a simple random sample.

The right hand side of the \code{formula} argument
should give the cross-classifying
variables, separated by \code{+} signs. The left hand side
should give the response variable. This variable must be
a binary outcome: a mixture of \code{1}s and \code{0}s
or a mixture of \code{TRUE}s and \code{FALSE}s.
A dot on the right hand side
is shorthand for "all variables not included
on the left hand side".

The \code{wt} vector is typically a column
within \code{data}. It should consist entirely
of non-negative values. If a record has a weight
of 0, the record is removed from the dataset
before the calculations are made.

The method for calculating design effects,
and hence effective numbers of trials and
successes, is specified by the \code{method}
argument. The default method, \code{"ghitza_med"}
is based on Ghitza and Gelman (2013),
and proceeds as follows:
\enumerate{
  \item Calculate design effect. This is proportional
    to the variance of the weights within each
    cell of the tabulation. An overall design
    effect is calculated by taking the median of
    the cell-level design effects.
  \item Calculate effective number of trials.
    The effective number of trials within each
    cell is calculated by dividing the (unweighted) number of
    responses in each cell by the overall design effect.
  \item Calculate weighted proportion of successes in
    each cell. The proportion of successes is calculated
    by taking the weighted mean of the outcome variable,
    with the weights equal to the survey weights.
  \item Calculate effective number of succcesses.
    Multiply the weighted proportion of successes in each
    cell by the effective number of trials in that
    cell.
}

Ghitza and Gelman (2013) in fact calcuate the overall
design effect by taking the mean, rather than
the median, of the cell-level design effects.
The mean is, however, sensitive to outliers, as
can occur when cells have few observations.
The original method can be obtained by setting
\code{method} to \code{"ghitza_mean"}.

The effective numbers of success and trials obtained
in this way will not generally be integers. To
perform further calculations on these data it may
be necessary to \code{\link{round}} them.

If there are no observations for a particular
combination of the cross-classifying variables,
then the corresponding cell is not included in the
calculation of design effects, and the number of
effective trials and successes for that cell
are set to 0. If there is one observation for a particular
combination of the cross-classifying variables,
then a design effect for that cell cannot be calculated,
and the cell is ignored in the calculation of the overall
design effect. However, estimates of the effective
number of trials and successes for that cell are provided.
}
\examples{
## create a small synthetic dataset
df <- data.frame(is_obese = rbinom(n = 25,
                                   size = 1,
                                   prob = 0.3),
                 age = sample(c("10-14", "15-19", "20-24"),
                              size = 25,
                              replace = TRUE),
                 sex = sample(c("F", "M"),
                              size = 25,
                              replace = TRUE),
                 wt = runif(n = 25,
                            min = 5,
                            max = 40))

## obtain tabulations
dtabs_survey(df, is_obese ~ age + sex, wt = wt)

## just for comparison, obtain tabulations without
## accounting for survey design
dtabs(df, is_obese ~ age + sex)
}
\seealso{
\code{dtabs_survey} is a modified version of function
\code{\link{dtabs}}, which is designed for unweighted data,
such as administrative or census data, or
for aggregate-level data. \code{dtabs} is in turn
based on function \code{\link[stats]{xtabs}}.
}
