% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip_year_label.R
\name{flip_year_label}
\alias{flip_year_label}
\title{Change format of one-year period or cohort labels}
\usage{
flip_year_label(x, current_uses_start = TRUE, month_start = "Jan")
}
\arguments{
\item{x}{A vector of period or cohort labels
that are single-year or open on the left.}

\item{current_uses_start}{Logical. Whether the
the current labels use calendar year at
the start of the period/cohort.}

\item{month_start}{An element of \code{\link[base]{month.name}},
or \code{\link[base]{month.abb}}. Each period/cohort
starts on the first day of this month.}
}
\value{
A factor with the same length as \code{x}.
}
\description{
Change one-year period or cohort labels from using
calendar year at the start of the period/cohort
to calendar year at the end, or vice versa.
}
\details{
As discussed in \code{\link{date_to_period_year}}
and \code{\link{date_to_cohort_year}}, single-year
labels are ambiguous. The label \code{"2020"}, for instance,
could refer to any period from
"1 February 2019 - 31 January 2020" to
"1 December 2020 - 31 November 2021", depending on
the starting month and on whether the label
uses the calendar year at the start of the period/cohort
or the calendar year at the end.

If \code{current_uses_start} is \code{TRUE},
then \code{flip_year_label} produces new
labels that use calendar year at the end
of the period/cohort. If \code{current_uses_start}
is \code{FALSE}, then \code{flip_year_label}
produces new labels that use calendar year at the
start of the period/cohort.
}
\examples{
## currently using calendar year at start of period
flip_year_label(x = c("2001", "2005"),
                current_uses_start = TRUE,
                month_start = "Jul")

## currently using calendar year at end of period
flip_year_label(x = c("2001", "2005"),
                current_uses_start = FALSE,
                month_start = "Jul")

## periods start in January, so changing labelling
## rule has no effect
flip_year_label(x = c("2001", "2005"),
                current_uses_start = TRUE,
                month_start = "Jan")
}
\seealso{
\code{\link{date_to_period_year}},
\code{\link{date_to_cohort_year}},
\code{\link{format_period_year}},
\code{\link{format_cohort_year}}
}
