% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_age.R
\name{format_age_births}
\alias{format_age_births}
\title{Put age groups labels into format
required for tabulating births}
\usage{
format_age_births(
  x,
  break_min = 15,
  break_max = 50,
  width = 5,
  recode_up = FALSE,
  recode_down = FALSE
)
}
\arguments{
\item{x}{A vector of age group labels. The
age of the parent at the time of the birth of
the child.}

\item{break_min}{An integer or \code{NULL}.
Defaults to 15.}

\item{break_max}{An integer or \code{NULL}.
Defaults to 50.}

\item{width}{The width in years of the age intervals.
A positive integer defaulting to 5.}

\item{recode_up}{If \code{TRUE}, births to parents
aged less than \code{break_min} are treated as occurring to
people in the lowest repoductive age group.}

\item{recode_down}{If \code{TRUE}, births to parents
aged \code{break_max} or more are treated as
occurring to people in the highest reproductive
age group.}
}
\value{
A factor with the same length as \code{x}.
}
\description{
Given a vector of age group labels, create a factor that contains
levels for all ages between \code{break_min} and \code{break_max}.
}
\details{
\code{break_min} and \code{break_max} specify
the range of ages over which reproduction
is assumed to occur. If, for instance,
\code{break_min} is \code{15} and \code{break_max}
is \code{50}, all births are assumed to
occur to women aged 15 to 49 (inclusive).
If \code{break_min} or \code{break_max} is set to \code{NULL},
rather than to a specific value, then \code{date_to_age_births}
finds the narrowest range that accommodates the values
in \code{x}.

Datasets sometimes contain a few births to parents
younger than the assumed minimum age of reproduction,
or births to parents older than the assumed maximum age
of reproduction. Demographers often recode such births,
so that ones to unexpectedly young parents are
treated as occurring just above the minimum age
for reproduction, and ones to unexpectedly old parents
are treated as occurring just below the maximum
age for reproduction. This recoding can be justified
on the grounds that some of the original ages may have
been misreported, but it also alleviates any problems
with tabulations having small counts at extreme ages.
Recoding of parents' ages outside the expected range
is controlled by parameters \code{recode_up}
and \code{recode_down}. The default
is for no recoding to occur.
}
\examples{
format_age_births(x = c("20-24", "37", NA, "32", "21-24"))

format_age_births(x = c("20-24", "37", "32", "21-24"),
                  width = 10,
                  break_min = 20)

format_age_births(x = c("20", "37", "15"),
                  width = 1,
                  break_max = 45)

## allow youngest and oldest age groups to be
## set by the data
format_age_births(x = c("21", "33", "22-24"),
                  break_min = NULL,
                  break_max = NULL)

## recode ages outside the expected range
format_age_births(x = c("22", "13", "54"),
                  recode_up = TRUE,
                  recode_down = TRUE)
}
\seealso{
Other functions for creating age groups are
\code{\link{format_age_year}},
\code{\link{format_age_multi}},
\code{\link{format_age_lifetab}},
\code{\link{format_age_custom}},
\code{\link{format_age_quarter}},
and \code{\link{format_age_month}}.

\code{\link{date_to_age_births}} creates
reproductive age groups from dates.

\code{\link{make_labels_age}} describes the rules
for constructing labels for age groups.
}
