% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_age.R
\name{format_age_custom}
\alias{format_age_custom}
\title{Put age group labels into format required
for customized age groups}
\usage{
format_age_custom(x, breaks = NULL, open_last = TRUE)
}
\arguments{
\item{x}{A vector of age group labels.}

\item{breaks}{A vector of strictly increasing integer values.}

\item{open_last}{Whether the final age group
has no upper limit. Defaults to \code{TRUE}.}
}
\value{
A factor with length equal to \code{x}.
}
\description{
Given a vector of age group labels, create a factor
that contains levels for all age groups
defined by \code{breaks}, including, possibly,
an open age group.\code{format_age_custom} is the most flexible
of the \code{format_age} functions
in that the age groups can have any combination of widths,
though the widths must be defined in whole numbers of years.
}
\details{
\code{breaks} is used to specify the points at which
each age group starts and finishes. If 
\code{open_last} is \code{TRUE}, and \code{b} is
the last value for \code{breaks}, then the oldest
age group is \code{[b, Inf)} years. 
If \code{open_last} is \code{FALSE}, \code{a} is the
second-to-last value for \code{breaks} and \code{b}
is the last value, then the oldest age
group is \code{[a, b)} years.
}
\examples{
format_age_custom(x = c("90+", "19-40", "22", NA),
                  breaks = c(0, 15, 60))
format_age_custom(x = c("50-59", "19-40", "31"),
                  breaks = c(15, 45, 60),
                  open_last = FALSE)
}
\seealso{
Other functions for creating age groups are
\code{\link{format_age_year}},
\code{\link{format_age_multi}},
\code{\link{format_age_lifetab}},
\code{\link{format_age_births}},
\code{\link{format_age_quarter}},
and \code{\link{format_age_month}}.

\code{\link{date_to_age_custom}} creates
customized age groups from dates.

\code{\link{make_labels_age}} describes the rules
for constructing labels for age groups.
}
