% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_age.R
\name{format_age_month}
\alias{format_age_month}
\title{Put age group labels into format required
for one-month age groups}
\usage{
format_age_month(x, break_min = 0, break_max = 1200, open_last = TRUE)
}
\arguments{
\item{x}{A vector of age group labels.}

\item{break_min}{An integer or \code{NULL}.
Defaults to 0.}

\item{break_max}{An integer or \code{NULL}.
Defaults to 1200.}

\item{open_last}{Whether the final age group
has no upper limit. Defaults to \code{TRUE}.}
}
\value{
A factor with the same length as \code{x}.
}
\description{
Given a vector of age group labels, create a factor
that contains levels for all ages between \code{break_min}
and \code{break_max}, and that may contain an open age group
(ie an age group with no upper limit.)
If, for instance, \code{break_min} is \code{0},
\code{break_max} is \code{1200}, and \code{open_last} is
\code{TRUE} (the defaults), then \code{format_age_year}
creates a factor with levels \code{"0"}, \code{"1"},
\dots, \code{"1199"}, \code{"1200+"}. Even when an age
group between \code{break_min} and \code{break_max}
is not included in \code{x}, \code{factor_month}
still creates a level for it.
}
\details{
A person belongs to age group \code{"x"} if that
person was exactly \code{x} months
old at their most recent birthday.
For instance, a person belongs to age
group \code{"60"} if that person had their
5th birthday (= 60 months) two days ago.

If \code{break_min} or \code{break_max} is set to \code{NULL},
rather than to a specific value, then \code{format_age_year}
finds the narrowest range that accommodates the data.

All age groups in \code{x} must be single-quarter age groups,
except above \code{break_max}, where open age groups
are allowed.

If \code{x} contains \code{NA}, then the
levels of the factor created by \code{format_age_month}
also contain \code{NA}.
}
\examples{
format_age_month(x = c(22, 0, 300))

format_age_month(x = c("3", NA, "12", "1400+"))

## specify highest age group
format_age_month(x = c("3", "12", "1400+"),
                 break_max = 24)

## let lowest age group be determined by the data
format_age_month(x = c("3", "12", "1400+"),
                 break_min = NULL,
                 break_max = 24)

## make final age group closed
format_age_month(x = c("3", "12"),
                 break_min = NULL,
                 break_max = 24,
                 open_last = FALSE)
}
\seealso{
Other functions for creating age groups are
\code{\link{format_age_year}},
\code{\link{format_age_multi}},
\code{\link{format_age_lifetab}},
\code{\link{format_age_births}},
\code{\link{format_age_custom}},
\code{\link{format_age_quarter}}.
}
