% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_age.R
\name{format_age_quarter}
\alias{format_age_quarter}
\title{Create consistent, complete quarter (three month) age groups}
\usage{
format_age_quarter(x, break_min = 0, break_max = 400, open_last = TRUE)
}
\arguments{
\item{x}{A vector of age group labels.}

\item{break_min}{An integer or \code{NULL}.
Defaults to 0.}

\item{break_max}{An integer or \code{NULL}.
Defaults to 400.}

\item{open_last}{Whether the final age group
has no upper limit. Defaults to \code{TRUE}.}
}
\value{
A factor with the same length as \code{x}.
}
\description{
Given a vector of age group labels, create a \code{\link[base]{factor}}
that contains levels for all ages between \code{break_min}
and \code{break_max}. The labels may contain an open age group,
ie an age group with no upper limit. Apart from the open age
group, all the age groups have a width of one quarter
(three months).
}
\details{
Even when an age
group between \code{break_min} and \code{break_max}
is not included in \code{x}, \code{factor_quarter}
still creates a level for it.

If \code{break_min} or \code{break_max} is set to \code{NULL},
rather than to a specific value, then \code{format_age_year}
finds the narrowest range that accommodates the data.

If \code{x} contains \code{NA}, then the
levels of the factor created by \code{format_age_quarter}
also contain \code{NA}.
}
\examples{
format_age_quarter(x = c(0, 21, 5))

format_age_quarter(x = c("10", "22", "500+"))

## specify highest age group
format_age_quarter(x = c("10", "22", "500+"),
                   break_max = 48)

## let lowest age group be determined by the data
format_age_quarter(x = c("10", "22", "500+"),
                   break_min = NULL,
                   break_max = 48)

## make final age group closed
format_age_quarter(x = c("10", "22"),
                   break_min = NULL,
                   break_max = 48,
                   open_last = FALSE)
}
\seealso{
Other functions for creating age groups are
\itemize{
  \item \code{\link{format_age_year}}
  \item \code{\link{format_age_multi}}
  \item \code{\link{format_age_lifetab}}
  \item \code{\link{format_age_births}}
  \item \code{\link{format_age_custom}}
  \item \code{\link{format_age_month}}
}

\code{\link{date_to_age_quarter}} calculates
ages from dates.
}
