% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_cohort.R
\name{format_cohort_custom}
\alias{format_cohort_custom}
\title{Put cohort labels into the format required
for customised cohorts}
\usage{
format_cohort_custom(
  x,
  breaks,
  open_first = NULL,
  month_start = "Jan",
  label_year_start = TRUE
)
}
\arguments{
\item{x}{A vector of cohort labels.}

\item{breaks}{A vector of strictly increasing integer values.}

\item{open_first}{Whether the oldest cohort
has no lower limit.}

\item{month_start}{An element of \code{\link[base]{month.name}},
or \code{\link[base]{month.abb}}. Cohorts start on
the first day of this month.}

\item{label_year_start}{Logical. Whether to label a cohort
by the calendar year at the beginning of the cohort
or the calendar year at the end. Defaults to \code{TRUE}.}
}
\value{
A factor with the same length as
\code{x}.
}
\description{
Given a vector of cohort labels, create a
\code{\link[base]{factor}}
that contains levels for all cohorts
defined by \code{breaks}. \code{format_cohort_custom}
is the most flexible
of the \code{format_cohort} functions
in that the cohorts can have any combination of widths,
though the widths must be defined in whole numbers of years.
}
\details{
If \code{open_first} is \code{TRUE}, then the earliest
cohort has no lower limit. (This is equivalent
to an open age group with no upper limit.)

The elements of \code{x} must be
multi-year labels such as \code{"1950-1960"} and
\code{"2020-2025"}, or labels for intervals
that are open on the left, such as \code{"<2000"}
and \code{<"1960"}.

\code{open_first} defaults to \code{TRUE}
if any of the intervals in \code{x} is open,
and to \code{FALSE} otherwise.

If \code{breaks} has length 0, then
\code{open_first} must be \code{FALSE}.
If \code{breaks} has length 1, then
\code{open_first} must be \code{TRUE}.

If \code{x} contains \code{NA}, then the
levels of the factor created by \code{format_cohort_custom}
also contain \code{NA}.
}
\examples{
format_cohort_custom(x = c(2019, 2011, 2000, 2015),
                     breaks = c(1990, 2000, 2020))

## change interpretation of single-year labels
format_cohort_custom(x = c(2019, 2011, 2000, 2015),
                     breaks = c(1990, 2000, 2020),
                     month_start = "Jul",
                     label_year_start = FALSE)

## multi-year labels
format_cohort_custom(x = c("2000", "2005-2010", "1995-1999"),
                     breaks = c(1990, 2000, 2020))

format_cohort_custom(x = c("2000", "2005-2010", "1995-1999"),
                     breaks = c(1995, 2005, 2010, 2020),
                     open_first = TRUE)
}
\seealso{
Other functions for reformating
cohort labels are 
\code{\link{format_cohort_year}},
\code{\link{format_cohort_multi}},
\code{\link{format_cohort_quarter}},
and \code{\link{format_cohort_month}}.

\code{\link{date_to_cohort_year}} calculates
cohorts from dates.
}
