% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_cohort.R
\name{format_cohort_month}
\alias{format_cohort_month}
\title{Put cohort labels into the format required
for one-month cohorts}
\usage{
format_cohort_month(x, break_min = NULL, open_first = NULL)
}
\arguments{
\item{x}{A vector of cohort labels.}

\item{break_min}{The lower limit of the cohort,
or \code{NULL} (the default).}

\item{open_first}{Whether the first cohort
has no lower limit.}
}
\value{
A factor with the same length as
\code{x}.
}
\description{
Given a vector of cohort labels, create a
\code{\link[base]{factor}} where the levels
contain a complete set of cohorts.
If \code{open_first} is \code{TRUE}, then the earliest
cohort has no lower limit. (This is equivalent
to an open age group with no upper limit.)
}
\details{
The elements of \code{x} must be month
labels, such as \code{"2001 Jan"} or \code{"2055 Sep"}, including
labels for cohorts that are open on the left,
such as \code{"<2000 Mar"}
or \code{<"1960 Dec"}.

\code{open_first} defaults to \code{TRUE}
if a value for \code{break_min} is supplied,
or if any intervals in \code{x} is open,
and to \code{FALSE} otherwise.

If \code{x} contains \code{NA}, then the
levels of the factor created by \code{format_cohort_month}
also contain \code{NA}.
}
\examples{
format_cohort_month(x = c("2003 Dec", "2005 Jan", NA, "2004 Sep"))

## 'open_first' defaults to TRUE, since 'x'
## contains an open interval
format_cohort_month(x = c("<2003 Dec", "2005 Jan", NA, "2004 Sep"))

## 'open_first' defaults to TRUE, since
## a value for 'break_min' is supplied
format_cohort_month(x = c("2003 Dec", "2005 Jan", NA, "2004 Sep"),
                    break_min = "2004 Jun")

## 'break_min' specified, and 'open_first' set to FALSE
format_cohort_month(x = c("2005 Jan", NA, "2004 May"),
                      break_min = "2004 Jan",
                      open_first = FALSE)
}
\seealso{
Other functions for reformating
cohort labels are 
\code{\link{format_cohort_year}},
\code{\link{format_cohort_multi}},
\code{\link{format_cohort_custom}},
and \code{\link{format_cohort_quarter}}.

\code{\link{date_to_cohort_month}} creates
month cohorts from dates.
}
