% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_cohort.R
\name{format_cohort_multi}
\alias{format_cohort_multi}
\title{Create consistent, complete multi-year cohorts}
\usage{
format_cohort_multi(
  x,
  width = 5,
  origin = 2000,
  break_min = NULL,
  open_first = NULL,
  month_start = "Jan",
  label_year_start = TRUE,
  label_open_multi = NULL
)
}
\arguments{
\item{x}{A vector of cohort labels.}

\item{width}{The width, in whole years, of the cohorts
to be created. Defaults to 5.}

\item{origin}{An integer. Defaults to 2000.}

\item{break_min}{An integer or \code{NULL}
(the default). If an integer, it is the
year in which the oldest cohort begins.}

\item{open_first}{Whether the oldest cohort
has no lower limit.}

\item{month_start}{An element of \code{\link[base]{month.name}},
or \code{\link[base]{month.abb}}. Cohorts start on
the first day of this month.}

\item{label_year_start}{Logical. Whether to label a cohort
by the calendar year at the beginning of the cohort
or the calendar year at the end. Defaults to \code{TRUE}.}

\item{label_open_multi}{Whether intervals that are open
on the left should be interpreted as multi-year
or single-year labels.}
}
\value{
A factor with the same length as
\code{x}.
}
\description{
Given a vector of cohort labels, create a
\code{\link[base]{factor}}  containing
levels for the earliest and latest cohorts
in \code{x}, and for all cohorts in between.
All cohorts, with the possible exception of
a first "open" cohort, have the same width,
which is controlled by the \code{width} argument.
}
\details{
The elements of \code{x} can be
single-year labels such as \code{"2020"},
multi-year labels such as \code{"1950-1960"},
and intervals that are open on the left,
such as \code{"<2000"}.

As discussed in \code{\link{date_to_cohort_year}},
single-year labels such as \code{"2000"} are ambiguous.
Correctly aligning single-year and multi-year cohorts
requires knowing which month the single-year cohorts start on,
which is controlled by the \code{month_start}
argument, and whether single-year cohorts
are labelled according 
to the calendar year at the start or end of the cohort,
which is controlled by the \code{label_year_start}
argument.

\code{open_first} defaults to \code{TRUE}
if a value for \code{break_min} is supplied,
or if any intervals in \code{x} are open,
and to \code{FALSE} otherwise.

The location of the cohorts can be shifted
by using different values for \code{origin}.

If \code{x} contains \code{NA}, then the
levels of the factor created by \code{format_cohort_multi}
also contain \code{NA}.

There is a (slightly obscure)
combination of settings that make an open
interval such as \code{"<2010"} ambiguous.
The settings are
\enumerate{
  \item \code{x} contains a mix of single-year labels
such as \code{"2018"} and multi-year labels such as
\code{"2020-2025"}
  \item \code{month_start} is not January.
  \item \code{label_year_start} is \code{FALSE}.
}
With these settings, it is unclear whether \code{"<2010"}
should be treated as a type of single-year label,
in which case it refers to the period before
\code{"2009-<month_start>-01"}, or as a type of
multi-year label, in which case it refers to the
period before \code{"2010-<month_start>-01"}.
Supplying a value for \code{label_open_multi}
removes the ambiguity.
When \code{label_open_multi} is \code{TRUE},
open intervals
interpreted as a type of multi-year label,
and when \code{label_open_multi} is \code{FALSE},
they are interpreted as
a type of single-year label.
}
\examples{
format_cohort_multi(x = c(2000, 2005, NA, 2004))

format_cohort_multi(x = c("2000", "2005-2010", NA, "1995-1999"))

## contains open interval
format_cohort_multi(x = c("2000", "2005-2010", NA, "<1995"))

## changing the interpretation of the labels results in the
## reclassification of cohort "2000"
format_cohort_multi(x = c(2000, 2005, NA, 2004),
                    month_start = "Jul",
                    label_year_start = FALSE)
format_cohort_multi(x = c("2000", "2005-2010", NA, "1995-1999"),
                    month_start = "Jul",
                    label_year_start = FALSE)

## 'break_min' is higher than the minimum of 'x'
format_cohort_multi(x = c("2000", "2005-2010", NA, "1995-1999"),
                    break_min = 2005)

## 'break_min' is lower then the minimum of 'x'
format_cohort_multi(x = c("2000", "2005-2010", NA, "1995-1999"),
                    break_min = 1990)

## 'break_min' supplied, but 'open_first' is FALSE
format_cohort_multi(x = c("2000", "2005-2010", NA, "1995-1999"),
                    break_min = 1990,
                    open_first = FALSE)

## non-default value for 'width'
format_cohort_multi(x = c("2000", "2005-2010", NA, "1995-1999"),
                    width = 10)

## non-default value for 'origin', to shift labels by one year
format_cohort_multi(x = c("2000", "2005-2010", NA, "1995-1999"),
                    width = 10,
                    origin = 2001)

## supply value for 'label_open_multi' to remove
## ambiguity of the "<2000" label
format_cohort_multi(x = c("2025", "2030-2035", "<2021"),
                    month_start = "Jul",
                    label_year_start = FALSE,
                    label_open_multi = FALSE)
}
\seealso{
Other functions for reformating
cohort labels are
\itemize{
  \code{\link{format_cohort_year}}
  \code{\link{format_cohort_custom}}
  \code{\link{format_cohort_quarter}}
  \code{\link{format_cohort_month}}
}

\code{\link{date_to_cohort_year}} calculates
cohorts from dates.
}
