% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_cohort.R
\name{format_cohort_quarter}
\alias{format_cohort_quarter}
\title{Put cohort labels into the format required
for quarter (three-month) cohorts}
\usage{
format_cohort_quarter(x, break_min = NULL, open_first = NULL)
}
\arguments{
\item{x}{A vector of cohort labels.}

\item{break_min}{An integer or \code{NULL}
(the default). If an integer, it is the
year in which the oldest cohort begins.}

\item{open_first}{Whether the oldest cohort
has no lower limit.}
}
\value{
A factor with the same length as
\code{x}.
}
\description{
Given a vector of cohort labels, create a
\code{\link[base]{factor}} where the levels
contain a complete set of cohorts. 
If \code{open_first} is \code{TRUE}, then the earliest
cohort has no lower limit. (This is equivalent
to an open age group with no upper limit.)
}
\details{
The elements of \code{x} must be quarter
labels, such as \code{"2001 Q2"} or \code{"2055 Q1"}, including
labels for cohorts that are open on the left,
such as \code{"<2000 Q3"}
or \code{<"1960 Q4"}.

\code{open_first} defaults to \code{TRUE}
if a value for \code{break_min} is supplied,
or if any intervals in \code{x} is open,
and to \code{FALSE} otherwise.

Quarters are defined as follows:
\tabular{lll}{
  \strong{Quarter} \tab \strong{Start} \tab \strong{End} \cr
  Q1 \tab 1 January \tab 31 March \cr
  Q2 \tab 1 April \tab 30 June \cr
  Q3 \tab 1 July \tab 30 September \cr
  Q4 \tab 1 October \tab 31 December
}

If \code{x} contains \code{NA}, then the
levels of the factor created by \code{format_cohort_quarter}
also contain \code{NA}.
}
\examples{
format_cohort_quarter(x = c("2000 Q4", "2005 Q1", NA, "2004 Q3"))

## 'open_first' defaults to TRUE, since 'x'
## contains an open interval
format_cohort_quarter(x = c("<2000 Q4", "2005 Q1", NA, "2004 Q3"))

## 'open_first' defaults to TRUE, since
## a value for 'break_min' is supplied
format_cohort_quarter(x = c("2000 Q4", "2005 Q1", NA, "2004 Q3"),
                      break_min = "2004 Q1")

## 'break_min' specified, and 'open_first' set to FALSE
format_cohort_quarter(x = c("2005 Q1", NA, "2004 Q3"),
                      break_min = "2004 Q1",
                      open_first = FALSE)
}
\seealso{
Other functions for reformating
cohort labels are 
\code{\link{format_cohort_year}},
\code{\link{format_cohort_multi}},
\code{\link{format_cohort_custom}},
and \code{\link{format_cohort_month}}.

\code{\link{date_to_cohort_year}} calculates
cohorts from dates.
}
