% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_cohort.R
\name{format_cohort_year}
\alias{format_cohort_year}
\title{Create consistent, complete single-year cohorts}
\usage{
format_cohort_year(x, break_min = NULL, open_first = NULL)
}
\arguments{
\item{x}{A vector of cohort labels.}

\item{break_min}{An integer or \code{NULL}
(the default). If an integer, it is the
year in which the oldest cohort begins.}

\item{open_first}{Whether the oldest cohort
has no lower limit.}
}
\value{
A factor with the same length as
\code{x}.
}
\description{
Given a vector of cohort labels, create a
\code{\link[base]{factor}} with a complete
set of cohorts. The labels
may include an "open" cohort with no
lower limit. Apart from the open cohort,
all cohorts have a width of one year.
}
\details{
\code{x} must consist of labels for single-year
cohorts, such as \code{"2001"}, or
open cohorts, such as \code{"<2000"}.
\code{x} must not contain labels for
multi-year cohorts such
as \code{"2000-2005"}

\code{open_first} defaults to \code{TRUE}
if a value for \code{break_min} is supplied,
or if any elements of \code{x} are open,
and to \code{FALSE} otherwise.

If \code{x} has \code{NA}s, then the
levels of the factor created by \code{format_cohort_year}
also contain \code{NA}.
}
\examples{
## the 'levels' contain all values from
## '2000' to '2005', even when these do not
## appear in the data
format_cohort_year(x = c(2000, 2005, NA, 2004))

## 'x' contains an open interval, so
## 'open_first' defaults to TRUE
format_cohort_year(x = c("<2000", "2005", NA, "2004"))

## specify 'break_min', which also makes
## 'open_first' default to TRUE
format_cohort_year(x = c(2000, 2005, NA, 2004),
                   break_min = 1998)

## set 'break_min' above minimum value of 'x'
format_cohort_year(x = c(2000, 2005, NA, 2004),
                   break_min = 2001)

## set 'open_first' to FALSE 
format_cohort_year(x = c(2000, 2005, NA, 2004),
                   open_first = FALSE)
format_cohort_year(x = c(2000, 2005, NA, 2004),
                   break_min = 1998,
                   open_first = FALSE)

## if no value for 'break_min' is supplied, and one or
## more of the labels refers to an open cohort,
## then the levels start with the latest open cohort
format_cohort_year(x = c("2000", "2005", "<2002",
                         "2004", "<2001"))
}
\seealso{
Other functions for reformatting
cohort labels are
\itemize{
  \item \code{\link{format_cohort_multi}}
  \item \code{\link{format_cohort_custom}}
  \item \code{\link{format_cohort_quarter}}
  \item \code{\link{format_cohort_month}}
}

\code{\link{date_to_cohort_year}} creates
single-year cohorts from dates.
}
