% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_period.R
\name{format_period_custom}
\alias{format_period_custom}
\title{Put period labels into the format required
for customised periods}
\usage{
format_period_custom(x, breaks, month_start = "Jan", label_year_start = TRUE)
}
\arguments{
\item{x}{A vector of period labels.}

\item{breaks}{A vector of strictly increasing integer values.}

\item{month_start}{An element of \code{\link[base]{month.name}},
or \code{\link[base]{month.abb}}. Cohorts start on
the first day of this month.}

\item{label_year_start}{Logical. Whether to label a cohort
by the calendar year at the beginning of the cohort
or the calendar year at the end. Defaults to \code{TRUE}.}
}
\value{
A factor with the same length as
\code{x}.
}
\description{
Given a vector of period labels, create a
\code{\link[base]{factor}}
that contains levels for all periods
defined by \code{breaks}. \code{format_period_custom}
is the most flexible
of the \code{format_period} functions
in that the periods can have any combination of widths,
though the widths must be defined in whole numbers of years.
}
\details{
The elements of \code{x} can be
single-year labels such as \code{"2000"}
and \code{"1975"}, multi-year labels
such as \code{"1950-1960"} and \code{"2020-2025"},
or a mixture of the two.
\code{x} cannot contain open intervals
such as \code{"<2020"}.

If \code{x} contains \code{NA}, then the
levels of the factor created by \code{format_period_custom}
also contain \code{NA}.
}
\examples{
format_period_custom(x = c("2000-2001", "2004", "2005-2010", "1996-1998"),
                     breaks = c(1990, 2000, 2020))

format_period_custom(x = c("2000-2001", "2004", "2005-2010", "1996-1998"),
                     breaks = c(1995, 2005, 2010, 2020))
}
\seealso{
Other functions for reformating
period labels are 
\code{\link{format_period_year}},
\code{\link{format_period_multi}},
\code{\link{format_period_quarter}},
and \code{\link{format_period_month}}.

\code{\link{date_to_period_custom}} creates
customized periods from dates.

\code{\link{make_labels_period}} describes the rules
for constructing labels for periods.
}
