% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_period.R
\name{format_period_year}
\alias{format_period_year}
\title{Put period labels into the format required
for single-year periods}
\usage{
format_period_year(x)
}
\arguments{
\item{x}{A vector of period labels.}
}
\value{
A factor with the same length as
\code{x}.
}
\description{
Given a vector of period labels, create a
\code{\link[base]{factor}} that contains
levels for the earliest and latest periods in \code{x},
and for all periods in between.
For instance, if the earliest period in \code{x}
is \code{"1990"}, and the latest is \code{"2010"},
then \code{format_period_year} creates a factor
with levels \code{"1990"}, \code{"1991"}, \dots,
\code{"2010"}.
}
\details{
The elements of \code{x} must all be single years,
eg \code{"2001"} or \code{"2055"}.

If \code{x} contains \code{NA}, then the
levels of the factor created by \code{format_period_year}
also contain \code{NA}.
}
\examples{
format_period_year(x = c("2000", "2005", NA, "2004"))
}
\seealso{
Other functions for reformating
period labels are 
\code{\link{format_period_multi}},
\code{\link{format_period_custom}},
\code{\link{format_period_quarter}},
and \code{\link{format_period_month}}.

\code{\link{date_to_period_year}} creates
single-year periods from dates.

\code{\link{make_labels_period}} describes the rules
for constructing labels for periods.
}
