% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_triangle.R
\name{format_triangle_births}
\alias{format_triangle_births}
\title{Format labels for Lexis triangles
used when tabulating births}
\usage{
format_triangle_births(
  x,
  age,
  period,
  width = 5,
  break_min = 15,
  break_max = 50,
  recode_up = FALSE,
  recode_down = FALSE,
  month_start = "Jan",
  label_year_start = TRUE,
  origin = 2000
)
}
\arguments{
\item{x}{A vector of Lexis triangle labels.}

\item{age}{A vector of age groups, the same length
as \code{x}.}

\item{period}{A vector of periods, the same length
as \code{x}.}

\item{break_min}{An integer or \code{NULL}.
Defaults to 15.}

\item{break_max}{An integer or \code{NULL}.
Defaults to 50.}

\item{recode_up}{If \code{TRUE}, births to parents
aged less than \code{break_min} are treated as occurring to
people in the youngest repoductive age group.}

\item{recode_down}{If \code{TRUE}, births to parents
aged \code{break_max} or more are treated as
occurring to people in the oldest reproductive
age group.}
}
\value{
A factor with the same length as \code{x}.
}
\description{
Create labels for Lexis triangles used,
along with age groups and periods,
for tabulating births.
The age groups and periods must all have the same
length.
}
\details{
\code{age} and \code{period} define the
age groups and periods to which the
Lexis triangles
in \code{x} belong. Age groups can be single-year
(eg \code{"23"}) or multi-year (eg \code{"20-24"}).
Periods can also be single-year
(eg \code{"2023"}), multi-year (eg \code{"2020-2025"}).

\code{break_min} and \code{break_max} specify
the range of ages over which reproduction
is assumed to occur, and \code{recode_up} and
\code{recode_down} control the way that reported
ages outside this range are handled. See
\code{\link{date_to_age_births}} for details.

\code{format_triangle_births} returns different labels
from \code{x} in two situations: (i) when the intervals
in \code{age} and \code{period} are narrower than
\code{width}, so that aggregation is required, and
(ii) when ages lie outside the range set
by \code{break_min} and \code{break_max}.
If an age is recoded upwards to fall within
the youngest reproductive age group,
then the corresponding Lexis triangle is set to
\code{"Lower"}. If an age is recoded downwards to
fall within the lowest reproductive age group, then
the corresponding Lexis triangle is set to \code{"Upper"}.
}
\examples{
format_age_births(x = c("20-24", "37", NA, "32", "21-24"))

format_age_births(x = c("20-24", "37", "32", "21-24"),
                  width = 10,
                  break_min = 20)

format_age_births(x = c("20", "37", "15"),
                  width = 1,
                  break_max = 45)

## allow youngest and oldest age groups to be
## determined by the data
format_age_births(x = c("21", "33", "22-24"),
                  break_min = NULL,
                  break_max = NULL)

## recode ages outside the expected range
format_age_births(x = c("22", "13", "54"),
                  recode_up = TRUE,
                  recode_down = TRUE)
}
\seealso{
Other functions for creating Lexis triangles are
\code{\link{format_triangle_year}},
\code{\link{format_triangle_multi}},
\code{\link{format_triangle_quarter}},
and \code{\link{format_triangle_month}}.

\code{\link{date_to_triangle_births}} creates
Lexis triangles from dates.
}
