% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_triangle.R
\name{format_triangle_month}
\alias{format_triangle_month}
\title{Format labels for one-month Lexis triangles}
\usage{
format_triangle_month(x, age, break_max = 1200, open_last = TRUE)
}
\arguments{
\item{x}{A vector of Lexis triangle labels.}

\item{age}{A vector of age groups, the same length
as \code{x}.}

\item{break_max}{An integer or \code{NULL}.
Defaults to 1200.}

\item{open_last}{Whether the final age group
has no upper limit. Defaults to \code{TRUE}.}
}
\value{
A factor with the same length as
\code{x}.
}
\description{
Create labels for one-month Lexis triangles
to be used with labels for one-month
age groups and periods.
}
\details{
\code{age} gives the age group to which each triangle
in \code{x} belongs. All age groups in \code{age}
must be single-month, except for any open age groups.
The lower limit of open age groups in \code{age} must
be at least as high as \code{break_min}
(if \code{break_min} is non-\code{NULL}).

\code{break_max} and \code{open_first}
control the boundaries of Lexis triangles
for the oldest age group.
}
\examples{
## we construct 'x' and 'age' from
## dates information ourselves before
## calling 'format_triangle_month'
x <- date_to_triangle_month(date = c("2024-03-27",
                                     "2022-11-09"),
                            dob = "2020-01-01",
                            month_start = "Jul")
age <- date_to_age_month(date = c("2024-03-27",
                                  "2022-11-09"),
                         dob = "2020-01-01")
format_triangle_month(x = x,
                      age = age)
format_triangle_month(x = x,
                      age = age,
                      break_max = 12)

## someone else has constructed
## 'x' and 'age' from
## dates information
x <- c("Lower", "Lower", "Lower")
age <- c("10", "12+", "6")
format_triangle_month(x = x,
                      age = age,
                      break_max = 12)
}
\seealso{
Other functions for reformating
triangle labels are 
\code{\link{format_triangle_year}},
\code{\link{format_triangle_multi}},
\code{\link{format_triangle_births}},
and \code{\link{format_triangle_quarter}}.

\code{\link{date_to_triangle_month}} creates
one-month Lexis triangles from dates.
}
