% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_triangle.R
\name{format_triangle_multi}
\alias{format_triangle_multi}
\title{Format labels for multi-year Lexis triangles}
\usage{
format_triangle_multi(
  x,
  age,
  period,
  width = 5,
  break_max = 100,
  open_last = TRUE,
  month_start = "Jan",
  label_year_start = TRUE,
  origin = 2000
)
}
\arguments{
\item{x}{A vector of Lexis triangle labels.}

\item{age}{A vector of age groups, the same length
as \code{x}.}

\item{period}{A vector of periods, the same length
as \code{x}.}

\item{break_max}{An integer or \code{NULL}.
Defaults to 100.}

\item{open_last}{Whether the final age group
has no upper limit. Defaults to \code{TRUE}.}
}
\value{
A factor with the same length as
\code{x}.
}
\description{
Format labels for multi-year Lexis triangles to
be used with multi-year age groups and periods.
These age groups and periods (apart from a
possible open age group) all have the same width,
which is set by the \code{width} parameter.
}
\details{
\code{age} and \code{period} define the
age groups and periods to which the
Lexis triangles in \code{x} belong. These age groups
and periods can be narrower than \code{width},
Age groups can be single-year (\code{"23"}),
multi-year (\code{"20-24"})
or open (\code{"100+"}), and periods 
can be single-year (\code{"2023"})
or multi-year (\code{"2020-2025"}).

The values for \code{width}, \code{break_max}, \code{open_last},
and \code{origin} together define a new system
of Lexis triangles. \code{format_triangle_multi}
calculates where the triangles defined by
\code{x}, \code{age}, and \code{period} fall within
this new system. For instance, if an upper triangle
defined by \code{x}, \code{age}, and \code{period}
falls entirely within a lower triangle in the new
system, then \code{format_triangle_multi}
returns \code{"Lower"}.

\code{open_last} determines whether the
triangles need to account for an
open age group, and \code{break_max}
specifies the cut-off for the open age group.
See \code{\link{format_age_multi}} for a description
of how \code{open_last} and \code{break_max}
determine age groups.

\code{x} and \code{period} must be based on the same
starting month, so that if \code{x} uses years that
start in July and end in June,
then \code{period} must do so too. If
\code{x} was created using function
\code{\link{date_to_triangle_year}}
and \code{period} was created using function
\code{\link{date_to_period_year}},
then both should have used the
same value for \code{month_start}.
If \code{x} and \code{period} were not
calculated from raw dates data, 
then it may be necessary to check the
documentation for \code{x} and \code{period}
to see which
months of the year were used.
}
\examples{
## we construct 'x', 'age', and 'period'
## from dates information ourselves before
## calling 'format_triangle_multi'
date_original <- c("2024-03-27", "2022-11-09")
dob_original <- "2020-01-01"
x <- date_to_triangle_multi(date = date_original,
                            dob = dob_original,
                            month_start = "Jul")
age <- date_to_age_multi(date = date_original,
                         dob = dob_original)
period <- date_to_period_multi(date = date_original,
                               month_start = "Jul")
format_triangle_multi(x = x,
                      age = age,
                      period = period)

## someone else has constructed
## 'x', 'age', and 'period' from
## dates information
x_processed <- c("Lower", "Lower", "Lower")
age_processed <- c("10", "15+", "5")
period_processed <- c(2002, 2015, 2011)
format_triangle_multi(x = x_processed,
                      age = age_processed,
                      period = period_processed)

## alternative value for 'width'
format_triangle_multi(x = x_processed,
                      age = age_processed,
                      period = period_processed,
                      width = 10)

## alternative value for 'break_max'
format_triangle_multi(x = x_processed,
                      age = age_processed,
                      period = period_processed,
                      break_max = 10)
}
\seealso{
Other functions for reformating
triangle labels are 
\code{\link{format_triangle_year}},
\code{\link{format_triangle_births}},
\code{\link{format_triangle_quarter}},
and \code{\link{format_triangle_month}}.

\code{\link{date_to_triangle_multi}} creates
multi-year Lexis triangles from dates.
}
