% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_triangle.R
\name{format_triangle_quarter}
\alias{format_triangle_quarter}
\title{Format labels for quarter Lexis triangles}
\usage{
format_triangle_quarter(x, age, break_max = 400, open_last = TRUE)
}
\arguments{
\item{x}{A vector of Lexis triangle labels.}

\item{age}{A vector of age groups, the same length
as \code{x}.}

\item{break_max}{An integer or \code{NULL}.
Defaults to 400.}

\item{open_last}{Whether the final age group
has no upper limit. Defaults to \code{TRUE}.}
}
\value{
A factor with the same length as
\code{x}.
}
\description{
Create labels for one-quarter (three-month)
Lexis triangles to be used with one-quarter
age groups and periods.
}
\details{
\code{age} gives the age group to which each triangle
in \code{x} belongs. All age groups in \code{age}
must have a width of one quarter,
except for any open age groups.

\code{break_max} and \code{open_first}
control the boundaries of Lexis triangles
for the oldest age group.
}
\examples{
## we construct 'x' and 'age' from
## dates information ourselves before
## calling 'format_triangle_quarter'
x <- date_to_triangle_quarter(date = c("2024-03-27",
                                    "2022-11-09"),
                              dob = "2020-01-01")
age <- date_to_age_quarter(date = c("2024-03-27",
                                    "2022-11-09"),
                           dob = "2020-01-01")
format_triangle_quarter(x = x,
                        age = age)
format_triangle_quarter(x = x,
                        age = age,
                        break_max = 5)

## someone else has constructed
## 'x' and 'age' from
## dates information
x <- c("Lower", "Lower", "Lower")
age <- c("10", "15+", "5")
format_triangle_quarter(x = x,
                        age = age,
                        break_max = 10)
}
\seealso{
Other functions for reformating
triangle labels are 
\code{\link{format_triangle_year}},
\code{\link{format_triangle_multi}},
\code{\link{format_triangle_births}},
and \code{\link{format_triangle_month}}.

\code{\link{date_to_triangle_quarter}} creates
one-quarter Lexis triangles from dates.
}
