% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute_date}
\alias{impute_date}
\title{Impute a vector of dates}
\usage{
impute_date(year, month = NULL, day = NULL)
}
\arguments{
\item{year}{A numeric vector specifying years. Required.}

\item{month}{A numeric or character vector
specifying months. Optional.}

\item{day}{A numeric vector specifying days. Optional}
}
\value{
A vector of class \code{\link[base]{Date}}.
}
\description{
Given a vector of years, and possibly vectors 
of months and days, randomly generate a vector of dates. 
The dates are generated
from a uniform distribution:
all dates that fall within the range implied by 
the \code{year}, \code{month}, and \code{day} arguments
have the same chance of being selected. 
For instance, if \code{year}
is \code{2000}, \code{month} is \code{"March"}, and
no value is supplied for \code{date}, then every date from
\code{"2000-03-01"} to \code{"2000-03-31"} has the same
1/31 chance of being selected.
}
\details{
Months can be specified in one of four ways:
\enumerate{
  \item Numbers \code{1}, \code{2}, \dots, \code{12}
  \item Codes \code{"01"}, \code{"02"}, \dots, \code{"12"}
  \item English full names: \code{"January"}, \code{"February"}, \dots,
    \code{"December"}
  \item English abbreviations: \code{"Jan"}, \code{"Feb"}, \dots, 
    \code{"Dec"}
}

\code{month} and \code{day}, if supplied, must have the
same length as \code{year}.

\code{year}, \code{month}, and \code{day} can all
include \code{NA}s. The corresponding elements of the return
value will also be \code{NA}.
}
\examples{
impute_date(year = 2000:2004)
impute_date(year = 2000:2004,
            month = c("Feb", "Jun", "May", "Apr", "Feb"))
impute_date(year = 2000:2004,
            day = c(3, 31, 11, 2, NA))
}
\seealso{
\code{\link{impute_dob}}
}
