% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_valid.R
\name{is_valid_age}
\alias{is_valid_age}
\title{Check for valid age group labels}
\usage{
is_valid_age(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A logical vector, the same length as \code{x}.
}
\description{
Identify the elements of a vector that are
valid labels for age groups. Each element is assessed
independently, so the labels are not necessarily
consistent with each other.
}
\examples{
## years
x <- c("1-4", "12-10", "100+", "old")
is_valid_age(x)

## quarters
x <- c("44q", "100q+")
is_valid_age(x)

## months
x <- c("3m", "12m+")
is_valid_age(x)

## NAs are not valid age groups
x <- c("10-14", NA)
is_valid_age(x)

## age groups are allowed to overlap,
## and to mix years, quarters, and months,
## provided that each individual
## label is valid
x <- c("10-19", "15-19", "3q", "0m")
is_valid_age(x)
}
\seealso{
\code{\link{make_labels_age}} describes the rules
governing labels for age groups.
}
