% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_valid.R
\name{is_valid_period}
\alias{is_valid_period}
\title{Check for valid period labels}
\usage{
is_valid_period(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A logical vector, the same length as \code{x}.
}
\description{
Identify the elements of a vector that are
valid labels for periods. Each element is assessed
independently, so \code{x} can contain
a mix of different single-year, multiple-year,
quarter, and month formats, and labels
can overlap.
}
\details{
\code{NA}s are treated as valid labels.
}
\examples{
## years
x <- c("2030-2035", "2021", "2000", "Now", NA,
       "2020 Q4", "2021 Sep", "<2022 Jul")
is_valid_period(x)
}
