% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Labels-generators.R
\name{Label-generators}
\alias{Label-generators}
\alias{LabCategories}
\alias{LabTriangles}
\alias{LabPool}
\alias{LabQuantiles}
\alias{LabIntegers}
\alias{LabGroupedIntEnumerations}
\alias{LabGroupedIntEndpoints}
\alias{LabCalendarQuarters}
\alias{LabCalendarMonths}
\alias{LabDurationsQuarters}
\alias{LabDurationsMonths}
\title{Create object of class "Label"}
\usage{
LabCategories(labels, include_na)

LabTriangles(include_na)

LabPool(include_na)

LabQuantiles(labels, include_na)

LabIntegers(int_min, int_max, include_na)

LabGroupedIntEnumerations(breaks, open_first, open_last, include_na)

LabGroupedIntEndpoints(breaks, open_first, open_last, include_na)

LabCalendarQuarters(break_min, break_max, open_first, open_last, include_na)

LabCalendarMonths(break_min, break_max, open_first, open_last, include_na)

LabDurationsQuarters(break_min, break_max, open_first, open_last, include_na)

LabDurationsMonths(break_min, break_max, open_first, open_last, include_na)
}
\arguments{
\item{labels}{A chararacter vector with no
duplicates, blanks, or missing values,
listing all categories.}

\item{include_na}{Logical. Whether
to allow for an \code{NA} category.}

\item{int_min}{An integer, specifying the
bottom of the range of values.}

\item{int_max}{An integer, specifying the
top of the range of values.}

\item{breaks}{Integers or dates, defining
intervals.}

\item{open_first}{Logical. Whether the
first interval has a lower limit.}

\item{open_last}{Logical. Whether the
last interval has an upper limit.}

\item{break_min}{An integer or date, defining
interval limits.}

\item{break_max}{An integer or date, defining
interval limits.}
}
\value{
An object of class \code{\linkS4class{Label}}.
}
\description{
Generator functions for objects of class
\code{\linkS4class{Label}}. End-users would not
normally call these functions directly.
}
