% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels_to_dates.R
\name{labels_to_dates_month}
\alias{labels_to_dates_month}
\title{Convert a vector of month labels to list of pairs of dates}
\usage{
labels_to_dates_month(x, name)
}
\arguments{
\item{x}{A vector of labels.}

\item{name}{The name for \code{x} to be used in error messages}
}
\value{
A list, the items of which are Date
vectors of length 2.
}
\description{
Given a vector of labels for months
return a list of dates defining those months.
Assume that if \code{x} includes an open interval
that open interval is allowed to be there.
}
\details{
The items in the list all have length two. If an item represents
a closed interval, then the first element gives the date of the first
day of the interval, and the second element gives the date of
the first day of the next interval. If the item represents an
interval that is open on the left, then the first element
is \code{as.Date(NA)} and the second element is the first
day of the next interval. If the item represents a missing value,
then the first element and second element are both \code{as.Date(NA)}.

The elements are not assumed to be ordered, and there may be
more than one open interval and more than one missing value.
}
\examples{
x <- c("<2019 Aug", NA, "2011 Feb", "2012 Mar", NA)
labels_to_dates_month(x, name = "x", open_ok = TRUE)

x <- c(NA, "2011 Feb", "2012 Mar", NA)
labels_to_dates_month(x, name = "x", open_ok = FALSE)
}
\seealso{
\code{\link{labels_to_dates_year}},
\code{\link{labels_to_dates_quarter}}
}
\keyword{internal}
