% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_labels_age_group.R
\name{make_labels_age_group}
\alias{make_labels_age_group}
\alias{make_labels_age_group_quarter}
\alias{make_labels_age_group_month}
\title{Make age labels}
\usage{
make_labels_age_group(breaks, open_first = FALSE, open_last = TRUE,
  include_na = FALSE)

make_labels_age_group_quarter(break_min = 0, break_max = 400,
  open_first = FALSE, open_last = TRUE, include_na = FALSE)

make_labels_age_group_month(break_min = 0, break_max = 1200,
  open_first = FALSE, open_last = TRUE, include_na = FALSE)
}
\description{
Make age labels

Make labels for age groups measured in years
}
