% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_labels_age.R
\name{make_labels_age_month}
\alias{make_labels_age_month}
\title{Make labels for age groups measured in months}
\usage{
make_labels_age_month(
  break_min = 0,
  break_max = 1200,
  open_last = TRUE,
  include_na = FALSE
)
}
\arguments{
\item{break_min}{An integer. The lower limit (in months) of
the youngest age group. Defaults to \code{0}.}

\item{break_max}{An integer. If \code{open_last} is \code{TRUE},
the lower limit (in months) of the oldest age group;
otherwise the upper limit of the oldest age group.
Defaults to \code{1200}.}

\item{open_last}{Whether to append an open-ended
age group to the end of the labels.
Defaults to \code{TRUE}.}

\item{include_na}{Whether to append an \code{NA} to
the end of the labels. Defaults to \code{FALSE}.}
}
\value{
A character vector.
}
\description{
Make labels for age groups with lengths measured in
months.
}
\details{
A person's age group, measured in months, is the
number of months that have elapsed since that person was
born. For instance, the age in months of a person
born on 10 February 2010 is as follows:
\tabular{ll}{
  \strong{Date}  \tab \strong{Age} \cr
  10 February 2010 - 9 March 2010 \tab 0m \cr
  10 March 2010 - 9 April 2010 \tab 1m \cr
  10 April 2010 - 9 May 2010 \tab 2m \cr
  10 May 2010 - 9 June 2010 \tab 3m \cr
  10 June 2010 - 9 July 2010 \tab 4m \cr
  etc \tab etc
}

The lowest and highest age groups are defined by
the \code{break_min} and \code{break_max} arguments,
along with the \code{open_last} argument. The youngest age group
consists of everyone aged at least \code{break_min}
months. When \code{open_last} is \code{TRUE} (the default)
the oldest age group consists of everyone aged at least
\code{break_max} months, with no upper limit. When
\code{open_last} is \code{FALSE}, the oldest age group
consists of everyone aged at least \code{break_max - 1} months
but less than \code{break_max} months.

When \code{include_na} is \code{TRUE}, an \code{NA}
is added to the end of the labels. This can be useful
when dealing with dates that include \code{NA}s.
}
\examples{
make_labels_age_month(break_max = 10)
make_labels_age_month(break_max = 10,
                            open_last = FALSE)
make_labels_age_month(break_min = 5,
                            break_max = 10,
                            open_last = FALSE)
make_labels_age_month(break_max = 10,
                            include_na = TRUE)
}
\seealso{
To make labels for periods measured in months, use
\code{\link{make_labels_period_month}}. There is
no \code{make_labels_cohort_month} function. To construct
labels for month cohorts, just use \code{make_labels_period_month}.
To make labels for age groups measured in years,
use function \code{\link{make_labels_age}}, and
to make labels for age groups measured in quarters,
use function \code{\link{make_labels_age_quarter}}.
}
