% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_labels_age.R
\name{make_labels_age_quarter}
\alias{make_labels_age_quarter}
\title{Make labels for age groups measured in quarters}
\usage{
make_labels_age_quarter(
  break_min = 0,
  break_max = 400,
  open_last = TRUE,
  include_na = FALSE
)
}
\arguments{
\item{break_min}{An integer. The lower limit (in quarters) of
the youngest age group. Defaults to \code{0}.}

\item{break_max}{An integer. If \code{open_last} is \code{TRUE},
the lower limit (in quarters) of the oldest age group;
otherwise the upper limit of the oldest age group.
Defaults to \code{400}.}

\item{open_last}{Whether to append an open-ended
age group to the end of the labels.
Defaults to \code{TRUE}.}

\item{include_na}{Whether to append an \code{NA} to
the end of the labels. Defaults to \code{FALSE}.}
}
\value{
A character vector.
}
\description{
Make labels for age groups with lengths measured in
quarters (ie three months).
}
\details{
A person's age group, measured in quarters, is the
number of quarters that have elapsed since that person was
born. For instance, the age in quarters of a person
born on 10 February 2010 is as follows:
\tabular{ll}{
  \strong{Date}  \tab \strong{Age} \cr
  10 February 2010 - 9 May 2010 \tab 0q \cr
  10 May 2010 - 9 August 2010 \tab 1q \cr
  10 August 2010 - 9 November 2010 \tab 2q \cr
  10 November 2010 - 9 February 2011 \tab 3q \cr
  10 February 2011 - 9 May 2011 \tab 4q \cr
  etc \tab etc
}

The lowest and highest age groups are defined by
the \code{break_min} and \code{break_max} arguments,
along with the \code{open_last} argument. The youngest age group
consists of everyone aged at least \code{break_min}
quarters. When \code{open_last} is \code{TRUE} (the default)
the oldest age group consists of everyone aged at least
\code{break_max} quarters, with no upper limit. When
\code{open_last} is \code{FALSE}, the oldest age group
consists of everyone aged at least \code{break_max - 1} quarters
but less than \code{break_max} quarters.

When \code{include_na} is \code{TRUE}, an \code{NA}
is added to the end of the labels. This can be useful
when dealing with dates that include \code{NA}s.
}
\examples{
make_labels_age_quarter(break_max = 10)
make_labels_age_quarter(break_max = 10,
                              open_last = FALSE)
make_labels_age_quarter(break_min = 5,
                              break_max = 10,
                              open_last = FALSE)
make_labels_age_quarter(break_max = 10,
                              include_na = TRUE)
}
\seealso{
To make labels for periods measured in quarters, use
\code{\link{make_labels_period_quarter}}. There is
no \code{make_labels_cohort_quarter} function. To construct
labels for quarter cohorts, just use \code{make_labels_period_quarter}.
To make labels for age groups measured in years,
use function \code{\link{make_labels_age}}, and
to make labels for age groups measured in months,
use function \code{\link{make_labels_age_month}}.
}
