% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_labels.R
\name{make_labels_categories}
\alias{make_labels_categories}
\alias{make_labels_triangles}
\alias{make_labels_directions}
\alias{make_labels_quantiles}
\title{Make Categories labels}
\usage{
make_labels_categories(x)

make_labels_triangles(x)

make_labels_directions(x)

make_labels_quantiles(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector
}
\description{
Make labels with the format expected for a \code{"Categories"}
dimlabels object (as defined in the demarray package), or for
the \code{"Triangles"}, \code{"Directions"},
and \code{"Quantiles"} subclasses of \code{"Categories"}.
}
\details{
These functions would not normally be called directly
by end users.

\code{x} is a character vector, with no duplicates or blanks.
With \code{make_labels_triangles}, the elements must be
\code{"Lower"}, \code{"Upper"}, or \code{NA}.
With \code{make_labels_directions}, the elements must be
\code{"In"}, \code{"Out"}, or \code{NA}.
With \code{make_labels_quantiles}, the elements must be
a number between 0 and 100, followed by a "%" sign,
or be an \code{NA}.
}
\examples{
make_labels_categories(x = c("Nigeria", "Ethiopia", NA, "South Africa"))
make_labels_triangles(x = c("Upper", NA, "Lower"))
make_labels_directions(x = c("In", NA, "Out"))
make_labels_quantiles(x = c("97.5\%", NA, "2.5\%", "50\%"))
}
\keyword{internal}
