% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_labels.R
\name{make_labels_datepoints}
\alias{make_labels_datepoints}
\title{Make DatePoints labels}
\usage{
make_labels_datepoints(x)
}
\arguments{
\item{x}{A vector of class \code{\link[base]{Date}}.}
}
\value{
A character vector
}
\description{
Make labels with the format expected for a \code{"DatePoints"}
dimlabels object (as defined in the demarray package.)
}
\details{
This function would not normally be called directly
by end users.

\code{x} is a dates vector with no duplicates.
}
\examples{
x <- as.Date(c("2020-07-01", "2015-07-01", NA))
make_labels_datepoints(x)
}
\keyword{internal}
