% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_labels.R
\name{make_labels_dateranges}
\alias{make_labels_dateranges}
\title{Make DateRanges labels}
\usage{
make_labels_dateranges(x)
}
\arguments{
\item{x}{A list of Date vectors.}
}
\value{
A character vector
}
\description{
Make labels with the format expected for an
object of class \code{"DateRanges"}
(as defined in the demarray package).
This function would not normally be called directly
by end users.
}
\details{
\code{x} is a list of Date vectors of length 2.
If elements of a vector are non-NA, then the second
element must be later than the first.

The elements of \code{x} are converted into
labels as follows:

\tabular{ll}{
  Element \tab Label \cr
  \code{as.Date(c("2020-01-15", "2020-02-11"))} \tab \code{"[2020-01-15, 2020-02-10]"} \cr
  \code{as.Date(c(NA, "2020-01-01"))} \tab \code{"(-Inf, 2019-12-31]"} \cr
  \code{as.Date(c("2020-01-01", NA))} \tab \code{"[2020-01-01, Inf)"} \cr
  \code{as.Date(c(NA, NA))} \tab \code{NA} \cr
}
}
\examples{
x <- list(as.Date(c("2020-10-15", "2021-02-28")),
          as.Date(c(NA, "2020-01-31")),
          as.Date(c(NA_character_, NA_character_)),
          as.Date(c("2025-01-01", NA)),
          as.Date(c("2023-01-01", "2023-01-02")))
make_labels_dateranges(x)
}
\seealso{
\code{\link{make_labels_categories}},
\code{\link{make_labels_triangles}},
\code{\link{make_labels_directions}},
\code{\link{make_labels_quantiles}},
\code{\link{make_labels_integers}},
\code{\link{make_labels_intervals}},
\code{\link{make_labels_quantities}},
\code{\link{make_labels_quarters}},
\code{\link{make_labels_months}},
\code{\link{make_labels_datepoints}}
}
\keyword{internal}
