% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_labels_enum.R
\name{make_labels_enum}
\alias{make_labels_enum}
\title{Make labels for general numeric quantities}
\usage{
make_labels_enum(
  breaks,
  open_first = FALSE,
  open_last = FALSE,
  include_na = FALSE
)
}
\arguments{
\item{breaks}{A integer vector, or a vector that can
be coerced to integer via function \code{\link[base]{as.integer}}.}

\item{open_first}{Whether to append an open-ended
interval to the start of the labels.
Defaults to \code{FALSE}.}

\item{open_last}{Whether to append an open-ended
interval to the end of the labels.
Defaults to \code{FALSE}.}

\item{include_na}{Whether to append an \code{NA} to
the end of the labels. Defaults to \code{FALSE}.}
}
\value{
A character vector.
}
\description{
Make labels for numeric quantities, broadly defined.
The quantities may be combined to form intervals,
including intervals that have no lower limit or
no upper limit. Negative values are allowed,
but fractions are not.
}
\details{
The labels for intervals follow the same conventions
as age groups. Single-value intervals are represented
by single integers. Multiple-value intervals are
represented by intervals that start with the
lower value and end with the upper value minus one.
See below for examples.

When \code{open_first} is \code{TRUE}, an 'open' interval
with no lower limit is appended to the start of the
labels. When \code{open_last} is \code{TRUE}, an open interval
with no upper limit is appended to the end of the
labels.

When \code{include_na} is \code{TRUE}, an \code{NA}
is added to the end of the labels. This can be useful
when dealing with data that include \code{NA}s.

\code{breaks} can only have length 0 if
\code{open_first} and \code{open_last} are both
\code{FALSE}. \code{breaks} can have only have
length 1 if one or both of \code{open_first} and
\code{open_last} is \code{TRUE}.
}
\examples{
make_labels_enum(breaks = c(10, 100, 1000))
make_labels_enum(breaks = c(10, 100, 1000),
                open_last = TRUE)
make_labels_enum(breaks = seq(50, 100, 10))
make_labels_enum(breaks = 0:10,
                open_last = TRUE)
make_labels_enum(breaks = c(-1000, 0, 1000),
                open_first = TRUE,
                open_last = TRUE,
                include_na = TRUE)
}
\seealso{
To make labels for age groups, periods,
or cohorts, it is generally better to use
use specialist functions such as \code{\link{make_labels_age}}
or \code{\link{make_labels_period}}.
}
