% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_labels.R
\name{make_labels_integers}
\alias{make_labels_integers}
\title{Make Integers labels}
\usage{
make_labels_integers(x)
}
\arguments{
\item{x}{An integer vector.}
}
\value{
A character vector
}
\description{
Make labels with the format expected for a \code{"Integers"}
dimlabels object (as defined in the demarray package.)
}
\details{
This function would not normally be called directly
by end users.

\code{x} is an integer vector with no duplicates.
}
\examples{
x <- c(20L, 1:3, -2L)
make_labels_integers(x)
}
\keyword{internal}
