% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_labels.R
\name{make_labels_intervals}
\alias{make_labels_intervals}
\title{Make Intervals labels}
\usage{
make_labels_intervals(x)
}
\arguments{
\item{x}{A list.}
}
\value{
A character vector
}
\description{
Make labels with the format expected for a \code{"Intervals"}
dimlabels object (as defined in the demarray package.)
}
\details{
This function would not normally be called directly
by end users.

\code{x} is a list of integer vectors of
length 2. When both elements in an item are
non-NA, the second element must be greater than
the first. Intervals must not overlap.
}
\examples{
x <- list(c(2000L, 2010L),
          c(NA, 2000L),
          c(NA_integer_, NA__integer_),
          c(2030L, NA),
          c(2020L, 2030L))
make_labels_intervals(x)
}
\keyword{internal}
