% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_labels.R
\name{make_labels_months}
\alias{make_labels_months}
\title{Make Months labels}
\usage{
make_labels_months(x)
}
\arguments{
\item{x}{A list.}
}
\value{
A character vector
}
\description{
Make labels with the format expected for a \code{"Months"}
dimlabels object (as defined in the demarray package.)
}
\details{
This function would not normally be called directly
by end users.

\code{x} is a list of date vectors of
length 2. When both elements in an item are
non-NA, the second element must be one month
greater than the first. Months must not overlap.
}
\examples{
x <- list(as.Date(c("2020-10-01", "2020-11-01")),
          as.Date(c(NA, "2020-01-01")),
          as.Date(c(NA_character_, NA__character_)),
          as.Date(c("2025-01-01", NA)),
          as.Date(c("2020-01-01", "2020-02-01")))
make_labels_months(x)
}
\keyword{internal}
