% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_labels_period.R
\name{make_labels_period}
\alias{make_labels_period}
\alias{make_labels_period_year}
\alias{make_labels_period_quarter}
\alias{make_labels_period_month}
\title{Make period labels}
\usage{
make_labels_period_year(breaks, open_left = FALSE, open_right = FALSE,
  year_to = TRUE, include_na = FALSE)

make_labels_period_quarter(min_break, max_break, open_left = FALSE,
  open_right = FALSE, include_na = FALSE)

make_labels_period_month(min_break, max_break, open_left = FALSE,
  open_right = FALSE, include_na = FALSE)
}
\description{
Make period labels
}
