% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_labels.R
\name{make_labels_quantiles}
\alias{make_labels_quantiles}
\title{Make Quantiles labels}
\usage{
make_labels_quantiles(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector
}
\description{
Make labels with the format expected for an
object of class \code{"Quantiles"}
(as defined in the demarray package).
This function would not normally be called directly
by end users.
}
\details{
\code{x} is a character vector, with no duplicates.
The elements of \code{x} must be
a number between 0 and 100 (inclusive),
followed by a "\%" sign,
or be an \code{NA}.
}
\examples{
make_labels_quantiles(x = c("97.5\%", NA, "2.5\%", "50\%"))
}
\seealso{
\code{\link{make_labels_categories}},
\code{\link{make_labels_triangles}},
\code{\link{make_labels_directions}},
\code{\link{make_labels_integers}},
\code{\link{make_labels_intervals}},
\code{\link{make_labels_quantities}},
\code{\link{make_labels_quarters}},
\code{\link{make_labels_months}},
\code{\link{make_labels_dateranges}},
\code{\link{make_labels_datepoints}}
}
\keyword{internal}
