% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_labels.R
\name{make_labels_quantities}
\alias{make_labels_quantities}
\title{Make Quantities labels}
\usage{
make_labels_quantities(x)
}
\arguments{
\item{x}{A list.}
}
\value{
A character vector
}
\description{
Make labels with the format expected for a \code{"Quantities"}
dimlabels object (as defined in the demarray package.)
}
\details{
This function would not normally be called directly
by end users.

\code{x} is a list of integer vectors of
length 2. When both elements in an item are
non-NA, the second element must be greater than
the first. Quantities must not overlap.
}
\examples{
x <- list(c(0L, 4L),
          c(NA, -1L),
          c(NA_integer_, NA__integer_),
          c(20L, NA),
          c(10L, 19L))
make_labels_quantities(x)
}
\keyword{internal}
