% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_labels.R
\name{make_labels_quarters}
\alias{make_labels_quarters}
\title{Make Quarters labels}
\usage{
make_labels_quarters(x)

make_labels_quarters(x)
}
\arguments{
\item{x}{A list.}
}
\value{
A character vector

A character vector
}
\description{
Make labels with the format expected for a \code{"Quarters"}
dimlabels object (as defined in the demarray package.)

Make labels with the format expected for a \code{"DateRanges"}
dimlabels object (as defined in the demarray package.)
}
\details{
This function would not normally be called directly
by end users.

\code{x} is a list of date vectors of
length 2. When both elements in an item are
non-NA, the second element must be one quarter
later than the first. Quarters must not overlap.

This function would not normally be called directly
by end users.

\code{x} is a list of date vectors of
length 2. When both elements in an item are
non-NA, the second element must later than
the first. Ranges must not overlap.
}
\examples{
x <- list(as.Date(c("2020-10-01", "2021-01-01")),
          as.Date(c(NA, "2020-01-01")),
          as.Date(c(NA_character_, NA__character_)),
          as.Date(c("2025-01-01", NA)),
          as.Date(c("2020-01-01", "2020-04-01")))
make_labels_quarters(x)
x <- list(as.Date(c("2020-10-15", "2021-02-28")),
          as.Date(c(NA, "2020-01-31")),
          as.Date(c(NA_character_, NA__character_)),
          as.Date(c("2025-01-01", NA)),
          as.Date(c("2023-01-01", "2023-01-01")))
make_labels_dateranges(x)
}
\keyword{internal}
