% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_labels.R
\name{make_labels_triangles}
\alias{make_labels_triangles}
\title{Make Triangles labels}
\usage{
make_labels_triangles(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector
}
\description{
Make labels with the format expected for an
object of class \code{"Triangles"}
(as defined in the demarray package).
This function would not normally be called directly
by end users.
}
\details{
\code{x} is a character vector, with no duplicates.
The elements of \code{x} for must belong
to \code{c("Lower", "Upper", NA)}.
}
\examples{
make_labels_triangles(x = c("Upper", NA, "Lower"))
}
\seealso{
\code{\link{make_labels_categories}},
\code{\link{make_labels_directions}},
\code{\link{make_labels_quantiles}},
\code{\link{make_labels_integers}},
\code{\link{make_labels_intervals}},
\code{\link{make_labels_quantities}},
\code{\link{make_labels_quarters}},
\code{\link{make_labels_months}},
\code{\link{make_labels_dateranges}},
\code{\link{make_labels_datepoints}}
}
\keyword{internal}
