% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzdeaths.R
\docType{data}
\name{nzdeaths}
\alias{nzdeaths}
\title{Deaths in New Zealand}
\format{
A data frame with 138 rows and four columns:
\describe{
  \item{age}{Age group.}
  \item{sex}{Female or Male.}
  \item{year}{Year to June (ie 1 July to 30 June.)}
  \item{count}{Number of deaths.}
}
}
\source{
Downloaded from the "Deaths by age and sex (Annual-Jun)" table in the Infoshare online database on the Statistics New Zealand website on 30 June 2021.
}
\usage{
nzdeaths
}
\description{
Counts of deaths age group and sex,
in the years to June 2019 and 2020.
}
\keyword{datasets}
