% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_date_to_age_group.R
\name{plot_date_to_age_births}
\alias{plot_date_to_age_births}
\title{Depict the intervals created by
function 'date_to_age_births'}
\usage{
plot_date_to_age_births(
  date,
  dob,
  break_min = 15,
  break_max = 50,
  width = 5,
  recode_up = FALSE,
  recode_down = FALSE,
  show_months = FALSE
)
}
\arguments{
\item{date}{Dates when births being measured occur.}

\item{dob}{Dates of birth of monthers.}

\item{break_min}{An integer or \code{NULL}.
Defaults to 15.}

\item{break_max}{An integer or \code{NULL}.
Defaults to 50.}

\item{width}{The width in years of the age intervals.
A positive integer defaulting to 5.}

\item{recode_up}{If \code{TRUE}, births to women
aged less than \code{break_min} are treated as occurring to
women in the lowest repoductive age group.}

\item{recode_down}{If \code{TRUE}, births to women
aged \code{break_max} or more are treated as
occurring to women in the highest reproductive
age group.}

\item{show_months}{Whether to include vertical
lines showing boundaries between months.
Defaults to \code{FALSE}.}
}
\description{
Create plot illustrating the inputs and outputs
of function \code{\link{date_to_age_births}}.
}
\details{
This function is for learning about the
labelling conventions used in the \strong{dem} packages,
and about their implementation in package
\strong{demprep}. It would not normally be used
during actual data analysis.
}
\examples{
plot_date_to_age_births(date = c("2024-03-27", "2022-11-09"),
                              dob = c("2001-03-21", "2000-07-13"))
plot_date_to_age_births(date = c("2024-03-27", "2022-11-09"),
                              dob = c("2001-03-21", "2000-07-13"),
                              width = 10,
                              break_min = 20)

## allow youngest and oldest age groups to be
## set by the data
plot_date_to_age_births(date = c("2052-01-02", "2019-09-22", "2022-10-08"),
                              dob = c("2000-01-01", "2001-03-17", "2010-07-05"),
                              break_min = NULL,
                              break_max = NULL)

## recode ages outside the expected range
plot_date_to_age_births(date = c("2052-01-02", "2019-09-22", "2022-10-08"),
                              dob = c("2000-01-01", "2001-03-17", "2010-07-05"),
                              recode_up = TRUE,
                              recode_down = TRUE)
}
