% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_date_to_age.R
\name{plot_date_to_age_custom}
\alias{plot_date_to_age_custom}
\title{Depict the intervals created by
function 'date_to_age_custom'}
\usage{
plot_date_to_age_custom(
  date,
  dob,
  breaks = NULL,
  open_last = TRUE,
  show_months = FALSE
)
}
\arguments{
\item{date}{Dates of events or measurements.
A vector of class \code{\link[base]{Date}},
or a vector that can be coerced to class
\code{Date} using function \code{\link[base]{as.Date}}.}

\item{dob}{Dates of birth.
A vector of class \code{\link[base]{Date}},
or a vector that can be coerced to class
\code{Date} using function \code{\link[base]{as.Date}}.}

\item{breaks}{A vector of strictly increasing integer values.}

\item{open_last}{Whether the final age group
has no upper limit. Defaults to \code{TRUE}.}

\item{show_months}{Whether to include vertical
lines showing boundaries between months.
Defaults to \code{FALSE}.}
}
\description{
Create a plot illustrating how function
\code{\link{date_to_age_custom}} works.
}
\details{
\code{plot_date_to_age_custom} is typically used for
learning or documentation, rather than for
actual data analysis.
}
\examples{
plot_date_to_age_custom(date = c("2024-03-27",
                                 "2022-11-09"),
                        dob = c("2001-03-21",
                                "2010-07-13"),
                        breaks = c(0, 15, 60))
plot_date_to_age_custom(date = c("2024-03-27",
                                 "2022-11-09"),
                        dob = c("2001-03-21",
                                "2010-07-13"),
                        breaks = c(10, 40, 65))

## alternative specifications for oldest age group
plot_date_to_age_custom(date = c("2024-03-27",
                                 "2022-11-09"),
                        dob = c("2001-03-21",
                                "2010-07-13"),
                        breaks = c(0, 15, 50))
plot_date_to_age_custom(date = c("2024-03-27",
                                 "2022-11-09"),
                        dob = c("2001-03-21",
                                "2010-07-13"),
                        breaks = c(0, 15, 50),
                        open_last = FALSE)
}
