% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_date_to_age_group.R
\name{plot_date_to_age_group_month}
\alias{plot_date_to_age_group_month}
\title{Depict the intervals created by
function 'date_to_age_group_month'}
\usage{
plot_date_to_age_group_month(date, dob, break_max = 1200, open_last = TRUE)
}
\arguments{
\item{date}{Dates of events or measurements.}

\item{dob}{Dates of birth.}

\item{break_max}{An integer or \code{NULL}.
Defaults to 1200.}

\item{open_last}{Whether the final age group
has no upper limit. Defaults to \code{TRUE}.}
}
\description{
Create plot illustrating the inputs and outputs
of function \code{\link{date_to_age_group_month}}.
}
\details{
This function is for learning about the
labelling conventions used in the \strong{dem} packages,
and about their implementation in package
\strong{demprep}. It would not normally be used
during actual data analysis.
}
\examples{
plot_date_to_age_group_month(date = c("2004-03-27", "2002-11-09"),
                             dob = c("2001-03-21", "2000-07-13"))

## alternative specifications for oldest age group
plot_date_to_age_group_month(date = "2019-09-22",
                             dob = "1910-01-01")
plot_date_to_age_group_month(date = "2019-09-22",
                             dob = "1910-01-01",
                             break_max = 320)
plot_date_to_age_group_month(date = "2019-09-22",
                             dob = "1910-01-01",
                             break_max = NULL)
plot_date_to_age_group_month(date = "2019-09-22",
                             dob = "1910-01-01",
                             break_max = NULL,
                             open_last = FALSE)
}
