% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_date_to_age_group.R
\name{plot_date_to_age_group_multi}
\alias{plot_date_to_age_group_multi}
\title{Depict the intervals created by
function 'date_to_age_group_multi'}
\usage{
plot_date_to_age_group_multi(
  date,
  dob,
  width = 5,
  break_max = 100,
  open_last = TRUE
)
}
\arguments{
\item{date}{Dates of events or measurements.
A vector of class \code{\link[base]{Date}},
or a vector that can be coerced to class
\code{Date} using function \code{\link[base]{as.Date}}.}

\item{dob}{Dates of birth.
A vector of class \code{\link[base]{Date}},
or a vector that can be coerced to class
\code{Date} using function \code{\link[base]{as.Date}}.}

\item{width}{The width in years of the age intervals.
A positive integer. Defaults to 5.}

\item{break_max}{An integer or \code{NULL}.
Defaults to 100.}

\item{open_last}{Whether the final age group
has no upper limit. Defaults to \code{TRUE}.}
}
\description{
Create plot illustrating the inputs and outputs
of function \code{\link{date_to_age_group_multi}}.
}
\details{
This function is for learning about the
labelling conventions used in the \strong{dem} packages,
and about their implementation in package
\strong{demprep}. It would not normally be used
during actual data analysis.
}
\examples{
plot_date_to_age_group_multi(date = c("2024-03-27", "2022-11-09"),
                             dob = c("2001-03-21", "2000-07-13"))

## alternative values for 'width'
plot_date_to_age_group_multi(date = c("2024-03-27", "2022-11-09"),
                             dob = c("2001-03-21", "2000-07-13"),
                             width = 10)

## alternative specifications for oldest age group
plot_date_to_age_group_multi(date = "2019-09-22",
                             dob = "1910-01-01",
                              width = 20)
plot_date_to_age_group_multi(date = "2019-09-22",
                             dob = "1910-01-01",
                             width = 20,
                             break_max = 80)
}
