% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_date_to_age_group.R
\name{plot_date_to_age_group_year}
\alias{plot_date_to_age_group_year}
\title{Depict the intervals created by
function 'date_to_age_group_year'}
\usage{
plot_date_to_age_group_year(date, dob, break_max = 100, open_last = TRUE)
}
\arguments{
\item{date}{Dates of events or measurements.
A vector of class \code{\link[base]{Date}},
or a vector that can be coerced to class
\code{Date} using function \code{\link[base]{as.Date}}.}

\item{dob}{Dates of birth.
A vector of class \code{\link[base]{Date}},
or a vector that can be coerced to class
\code{Date} using function \code{\link[base]{as.Date}}.}

\item{break_max}{An integer or \code{NULL}.
Defaults to 100.}

\item{open_last}{Whether the final age group
has no upper limit. Defaults to \code{TRUE}.}
}
\value{
If \code{as_factor} is \code{TRUE}, then the return
value is a factor; otherwise it is a character vector.
The length of the return value equals the length
of \code{date} or the length of \code{dob}, whichever
is greater.
}
\description{
Depict the intervals created by
function 'date_to_age_group_year'
}
\examples{
plot_date_to_age_group_year(date = c("2004-03-27", "2002-11-09"),
                            dob = c("2001-03-21", "2000-07-13"),
                            break_max = 5)
plot_date_to_age_group_year(date = c("2004-03-27", "2002-11-09"),
                            dob = c("2001-03-21", "2000-07-13"),
                            break_max = 5,
                            open_last = FALSE)

## replicate date of birth
plot_date_to_age_group_year(date = c("2024-03-27", "2022-11-09"),
                       dob = "2011-05-18")

## return non-factor
plot_date_to_age_group_year(date = c("2024-03-27", "2022-11-09"),
                       dob = "2011-05-18")

## alternative specifications for oldest age group
plot_date_to_age_group_year(date = "2019-09-22",
                       dob = "1910-01-01")
plot_date_to_age_group_year(date = "2019-09-22",
                       dob = "1910-01-01",
                       break_max = 80)
plot_date_to_age_group_year(date = "2019-09-22",
                       dob = "1910-01-01",
                       break_max = NULL)
plot_date_to_age_group_year(date = "2019-09-22",
                       dob = "1910-01-01",
                       break_max = NULL,
                       open_last = FALSE)
}
\seealso{
\code{\link{date_to_age_group_year}}
}
