% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_date_to_age.R
\name{plot_date_to_age_lifetab}
\alias{plot_date_to_age_lifetab}
\title{Depict the intervals created by
function 'date_to_age_lifetab'}
\usage{
plot_date_to_age_lifetab(date, dob, break_max = 100, show_months = FALSE)
}
\arguments{
\item{date}{Date of death.}

\item{dob}{Dates of birth.
A vector of class \code{\link[base]{Date}},
or a vector that can be coerced to class
\code{Date} using function \code{\link[base]{as.Date}}.}

\item{break_max}{An integer or \code{NULL}.
Defaults to 100.}

\item{show_months}{Whether to include vertical
lines showing boundaries between months.
Defaults to \code{FALSE}.}
}
\description{
Create a plot illustrating how function
\code{\link{date_to_age_lifetab}} works.
}
\details{
\code{plot_date_to_age_lifetab} is typically used for
learning or documentation, rather than for
actual data analysis.
}
\examples{
plot_date_to_age_lifetab(date = c("2024-03-27",
                                  "2022-11-09"),
                         dob = c("2001-03-21",
                                 "2004-07-13"),
                         break_max = 30)
}
