% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_date_to_age.R
\name{plot_date_to_age_multi}
\alias{plot_date_to_age_multi}
\title{Depict the intervals created by
function 'date_to_age_multi'}
\usage{
plot_date_to_age_multi(
  date,
  dob,
  width = 5,
  break_min = 0,
  break_max = 100,
  open_last = TRUE,
  show_months = FALSE
)
}
\arguments{
\item{date}{Dates of events or measurements.
A vector of class \code{\link[base]{Date}},
or a vector that can be coerced to class
\code{Date} using function \code{\link[base]{as.Date}}.}

\item{dob}{Dates of birth.
A vector of class \code{\link[base]{Date}},
or a vector that can be coerced to class
\code{Date} using function \code{\link[base]{as.Date}}.}

\item{width}{The width in years of the age intervals.
A positive integer. Defaults to 5.}

\item{break_min}{An integer or \code{NULL}.
Defaults to 0.}

\item{break_max}{An integer or \code{NULL}.
Defaults to 100.}

\item{open_last}{Whether the final age group
has no upper limit. Defaults to \code{TRUE}.}

\item{show_months}{Whether to include vertical
lines showing boundaries between months.
Defaults to \code{FALSE}.}
}
\description{
Create a plot illustrating how function
\code{\link{date_to_age_multi}} works.
}
\details{
\code{plot_date_to_age_multi} is typically used for
learning or documentation, rather than for
actual data analysis.
}
\examples{
plot_date_to_age_multi(date = c("2024-03-27",
                                "2022-11-09"),
                       dob = c("2001-03-21",
                               "2013-07-13"),
                       break_max = 35)

## alternative values for 'width' and 'break_max'
plot_date_to_age_multi(date = c("2024-03-27",
                                "2022-11-09"),
                       dob = c("2001-03-21",
                               "2013-07-13"),
                       width = 10,
                       break_max = 40)

## oldest age group closed
plot_date_to_age_multi(date = c("2024-03-27",
                                "2022-11-09"),
                       dob = c("2001-03-21",
                               "2013-07-13"),
                       width = 10,
                       break_max = 40,
                       open_last = FALSE)
}
\keyword{interal}
