% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_date_to_cohort.R
\name{plot_date_to_cohort_custom}
\alias{plot_date_to_cohort_custom}
\title{Depict the intervals created by
function 'date_to_cohort_custom'}
\usage{
plot_date_to_cohort_custom(date, breaks, open_first = TRUE)
}
\arguments{
\item{date}{Dates of events or measurements.}

\item{breaks}{Dates defining starts and ends of cohorts.}

\item{open_first}{Whether the first cohort
has no lower limit. If \code{break_min} is non-\code{NULL}
and \code{label_year_start} is \code{TRUE}, then
then \code{open_first} defaults to \code{TRUE};
otherwise it defaults to \code{FALSE}.}
}
\description{
Create plot illustrating how function
\code{\link{date_to_cohort_custom}} works.
\code{plot_date_to_cohort_custom} is designed for
learning or documentation, rather than for
actual data analysis.
}
\examples{
## cohorts start on 1 January
plot_date_to_cohort_custom(date = c("2024-03-27",
                                    "2018-11-09",
                                    "2021-03-02"),
                           breaks = c("2000-01-01",
                                      "2019-01-01",
                                      "2026-01-01"))

## cohorts start on 1 March
plot_date_to_cohort_custom(date = c("2024-03-27",
                                    "2018-11-09",
                                    "2021-03-02"),
                           breaks = c("2000-03-01",
                                      "2019-03-01",
                                      "2026-03-01"))
}
