% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_date_to_cohort.R
\name{plot_date_to_cohort_month}
\alias{plot_date_to_cohort_month}
\title{Depict the intervals created by
function 'date_to_cohort_month'}
\usage{
plot_date_to_cohort_month(date, break_min = NULL, open_first = NULL)
}
\arguments{
\item{date}{Dates of events or measurements.}
}
\description{
Create plot illustrating how function
\code{\link{date_to_cohort_month}} works.
}
\details{
\code{plot_date_to_cohort_month} is typically used for
learning or documentation, rather than for
actual data analysis.
}
\examples{
plot_date_to_cohort_month(date = c("2024-03-27",
                                   "2023-08-13",
                                   "2023-11-09"))
plot_date_to_cohort_month(date = c("2024-03-27",
                                   "2023-08-13",
                                   "2023-11-09"),
                          break_min = "2023-10-01")
}
\keyword{internal}
