% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_date_to_cohort.R
\name{plot_date_to_cohort_multi}
\alias{plot_date_to_cohort_multi}
\title{Depict the intervals created by
function 'date_to_cohort_multi'}
\usage{
plot_date_to_cohort_multi(
  date,
  width = 5,
  origin = 2000,
  month_start = "Jan",
  break_min = NULL,
  open_first = NULL
)
}
\arguments{
\item{date}{Dates of events or measurements.}

\item{width}{The length, in whole years, of the cohorts.
Defaults to 5.}

\item{origin}{An integer. Defaults to 2000.}

\item{month_start}{An element of \code{\link[base]{month.name}},
or \code{\link[base]{month.abb}}. The cohort starts on
the first day of this month.}

\item{break_min}{The start date of the first cohort,
or \code{NULL} (the default.)
If non-\code{NULL}, \code{break_min} can be a single value of
class \code{\link[base]{Date}}, or a value that
can be coerced to class \code{Date}
via function \code{\link[base]{as.Date}};
in either case, it must be the first day of
a month.}

\item{open_first}{Whether the first cohort
has no lower limit. If \code{break_min} is non-\code{NULL},
then \code{open_first} defaults to \code{TRUE};
otherwise it defaults to \code{FALSE}.}
}
\description{
Create plot illustrating how function
\code{\link{date_to_cohort_multi}} works.
}
\details{
\code{plot_date_to_cohort_multi} is typically used for
learning or documentation, rather than for
actual data analysis.
}
\examples{
plot_date_to_cohort_multi(date = c("2024-03-27",
                                   "2018-11-09",
                                   "2021-03-02"))

## width is 10
plot_date_to_cohort_multi(date = c("2024-03-27",
                                   "2018-11-09",
                                   "2021-03-02"),
                          width = 5)

## origin is 2001
plot_date_to_cohort_multi(date = c("2024-03-27",
                                   "2018-11-09",
                                   "2021-03-02"),
                          origin = 2001)

## starts on 1 July rather than 1 January
plot_date_to_cohort_multi(date = c("2024-03-27",
                                   "2018-11-09",
                                   "2021-03-02"),
                          origin = 2001,
                          month_start = "Jul")

## first age group closed
plot_date_to_cohort_multi(date = c("2024-03-27",
                                   "2018-11-09",
                                   "2021-03-02"),
                          open_first = FALSE)

## first age group closed, set value for 'break_min'
plot_date_to_cohort_multi(date = c("2024-03-27",
                                   "2018-11-09",
                                   "2021-03-02"),
                          break_min = "2005-01-01",
                          open_first = FALSE)
}
\keyword{interal}
