% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_date_to_cohort.R
\name{plot_date_to_cohort_quarter}
\alias{plot_date_to_cohort_quarter}
\title{Depict the intervals created by
function 'date_to_cohort_quarter'}
\usage{
plot_date_to_cohort_quarter(date)
}
\arguments{
\item{date}{Dates of events or measurements.}
}
\description{
Create plot illustrating how function
\code{\link{date_to_cohort_quarter}} works.
\code{plot_date_to_cohort_quarter} is typically used for
learning or documentation, rather than for
actual data analysis.
}
\examples{
plot_date_to_cohort_quarter(date = c("2024-03-27",
                                     "2022-05-13",
                                     "2022-11-09"))
}
\keyword{internal}
