% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_date_to_cohort.R
\name{plot_date_to_cohort_year}
\alias{plot_date_to_cohort_year}
\title{Depict the intervals created by
function 'date_to_cohort_year'}
\usage{
plot_date_to_cohort_year(date, month_start = "Jan", label_year_start = TRUE)
}
\arguments{
\item{date}{Dates of events or measurements.}

\item{month_start}{An element of \code{\link[base]{month.name}},
or \code{\link[base]{month.abb}}. The cohort starts on
the first day of this month.}

\item{label_year_start}{Whether to label a cohort
by the calendar year at the beginning of the cohort
or the calendar year at the end. Not needed for cohorts
that start on 1 January. Defaults to \code{TRUE}.}
}
\description{
Create a plot illustrating how function
\code{\link{date_to_cohort_year}} works.
}
\details{
\code{plot_date_to_cohort_year} is typically used for
learning or documentation, rather than for
actual data analysis.
}
\examples{
plot_date_to_cohort_year(date = c("2024-03-27",
                                  "2022-11-09"))

## starts on 1 July rather than 1 January
plot_date_to_cohort_year(date = c("2024-03-27",
                                  "2022-11-09"),
                         month_start = "Jul")

## starts on 1 July rather than 1 January,
## and uses calendar year at end rather than
## calendar year at the beginning
plot_date_to_cohort_year(date = c("2024-03-27",
                                  "2022-11-09"),
                         month_start = "Jul",
                         label_year_start = FALSE)

## specify oldest cohort, with 'open_first'
## at default value of 'TRUE', so first cohort
## has no lower limit
}
\keyword{internal}
