% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_date_to_period.R
\name{plot_date_to_period_custom}
\alias{plot_date_to_period_custom}
\title{Depict the intervals created by
function 'date_to_period_custom'}
\usage{
plot_date_to_period_custom(date, breaks)
}
\arguments{
\item{date}{Dates of events or measurements.}

\item{breaks}{Dates defining starts and ends of periods.}
}
\description{
Create plot illustrating the inputs and outputs
of function \code{\link{date_to_period_custom}}.
}
\details{
This function is for learning about the
labelling conventions used in the \strong{dem} packages,
and about their implementation in package
\strong{demprep}. It would not normally be used
during actual data analysis.
}
\examples{
## periods start on 1 January
plot_date_to_period_custom(date = c("2024-03-27",
                                    "2018-11-09",
                                    "2021-03-02"),
                           breaks = c("2000-01-01",
                                      "2019-01-01",
                                      "2026-01-01"))

## periods start on 1 March
plot_date_to_period_custom(date = c("2024-03-27",
                                    "2018-11-09",
                                    "2021-03-02"),
                           breaks = c("2000-03-01",
                                      "2019-03-01",
                                      "2026-03-01"))
}
