% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_date_to_period.R
\name{plot_date_to_period_quarter}
\alias{plot_date_to_period_quarter}
\title{Depict the intervals created by
function 'date_to_period_quarter'}
\usage{
plot_date_to_period_quarter(date)
}
\arguments{
\item{date}{Dates of events or measurements.}
}
\description{
Depict the intervals created by
function 'date_to_period_quarter'
}
\examples{
plot_date_to_period_quarter(date = c("2024-03-27",
                                     "2020-01-03",
                                     "2022-11-09"))
}
\seealso{
\code{\link{date_to_period_quarter}}
}
