% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_date_to_period.R
\name{plot_date_to_period_year}
\alias{plot_date_to_period_year}
\title{Depict the intervals created by
function 'date_to_period_year'}
\usage{
plot_date_to_period_year(date, month_start = "Jan", label_year_start = TRUE)
}
\arguments{
\item{date}{Dates of events or measurements.}

\item{month_start}{An element of \code{\link[base]{month.name}},
or \code{\link[base]{month.abb}}. Each period starts on
the first day of this month.}

\item{label_year_start}{Whether to label a period
by the calendar year at the beginning of the period
or the calendar year at the end. Not needed for periods
that start on 1 January. Defaults to \code{TRUE}.}
}
\description{
Create plot illustrating how function
\code{\link{date_to_period_year}} works.
}
\details{
\code{plot_date_to_period_year} is typically used for
learning or documentation, rather than for
actual data analysis.
}
\examples{
plot_date_to_period_year(date = c("2024-03-27",
                                  "2022-11-09"))

## July to June
plot_date_to_period_year(date = c("2024-03-27",
                                  "2022-11-09"),
                         month_start = "Jul")

## July to June, using the calendar year at
## the end for the label
plot_date_to_period_year(date = c("2024-03-27",
                                  "2022-11-09"),
                         month_start = "Jul",
                         label_year_start = FALSE)
}
\keyword{internal}
