% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_date_to_triangle.R
\name{plot_date_to_triangle_births}
\alias{plot_date_to_triangle_births}
\title{Depict the Lexis triangles created by
function 'date_to_triangle_births'}
\usage{
plot_date_to_triangle_births(
  date,
  dob,
  width = 5,
  break_min = 15,
  break_max = 50,
  recode_up = FALSE,
  recode_down = FALSE,
  origin = 2000,
  month_start = "Jan",
  show_months = FALSE
)
}
\arguments{
\item{date}{Dates when births being measured occur.}

\item{dob}{Dates of birth of parents.}

\item{width}{The width in years of the periods and
age groups. A positive integer defaulting to 5.}

\item{break_min}{An integer or \code{NULL}.
Defaults to 15.}

\item{break_max}{An integer or \code{NULL}.
Defaults to 50.}

\item{recode_up}{If \code{TRUE}, births to people
aged less than \code{break_min} are treated as occurring to
people in the lowest repoductive age group.}

\item{recode_down}{If \code{TRUE}, births to people
aged \code{break_max} or more are treated as
occurring to people in the highest reproductive
age group.}

\item{origin}{An integer. Defaults to 2000.}

\item{month_start}{An element of \code{\link[base]{month.name}},
or \code{\link[base]{month.abb}}. The period starts on
the first day of this month.}

\item{show_months}{Whether to include vertical
lines showing boundaries between months.
Defaults to \code{FALSE}.}
}
\description{
Create plot illustrating how function
\code{\link{date_to_triangle_births}} works.
}
\details{
\code{plot_date_to_triangle_births} is typically used for
instruction or documentation, rather than for
actual data analysis.
}
\examples{
plot_date_to_triangle_births(date = c("2020-01-27",
                                      "2022-11-09"),
                             dob = c("2001-03-21",
                                     "2000-07-13"))

## alternative values for 'width'
plot_date_to_triangle_births(date = c("2024-01-27",
                                      "2022-11-09"),
                             dob = c("2001-03-21",
                                     "2000-07-13"),
                             width = 10,
                             break_min = 20)
plot_date_to_triangle_births(date = c("2024-01-27",
                                      "2022-11-09"),
                             dob = c("2001-03-21",
                                     "2000-07-13"),
                             width = 1,
                             break_max = 45)
}
\keyword{internal}
