% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_date_to_triangle.R
\name{plot_date_to_triangle_month}
\alias{plot_date_to_triangle_month}
\title{Depict the Lexis triangles created by
function 'date_to_triangle_month'}
\usage{
plot_date_to_triangle_month(date, dob)
}
\arguments{
\item{date}{Dates of events.}

\item{dob}{Dates of birth.}
}
\description{
Create plot illustrating how function
\code{\link{date_to_triangle_month}} works.
}
\details{
\code{plot_date_to_triangle_month} is typically used for
instruction or documentation, rather than for
actual data analysis.
}
\examples{
plot_date_to_triangle_month(date = c("2024-03-27",
                                     "2024-11-09"),
                            dob = c("2023-10-05",
                                    "2024-03-11"))
}
\keyword{internal}
