% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_date_to_triangle.R
\name{plot_date_to_triangle_multi}
\alias{plot_date_to_triangle_multi}
\title{Depict the Lexis triangles created by
function 'date_to_triangle_multi'}
\usage{
plot_date_to_triangle_multi(
  date,
  dob,
  width = 5,
  break_max = 100,
  open_last = TRUE,
  origin = 2000,
  month_start = "Jan",
  label_year_start = TRUE,
  show_months = FALSE
)
}
\arguments{
\item{date}{Dates of events.}

\item{dob}{Dates of birth.}

\item{width}{The width in years of the periods and
age groups. A positive integer defaulting to 5.}

\item{break_max}{An integer or \code{NULL}.
Defaults to 100.}

\item{open_last}{Whether the final age group
has no upper limit. Defaults to \code{TRUE}.}

\item{origin}{An integer. Defaults to 2000.}

\item{month_start}{An element of \code{\link[base]{month.name}},
or \code{\link[base]{month.abb}}. The period starts on
the first day of this month.}

\item{label_year_start}{Whether to label a period
by the calendar year at the beginning of the period
or the calendar year at the end. Not needed for periods
that start on 1 January. Defaults to \code{FALSE}.}

\item{show_months}{Whether to include vertical
lines showing boundaries between months.
Defaults to \code{FALSE}.}
}
\description{
Create plot illustrating how function
\code{\link{date_to_triangle_multi}} works.
\code{plot_date_to_triangle_multi} is designed for
instruction or documentation, rather than for
actual data analysis.
}
\examples{
plot_date_to_triangle_multi(date = c("2027-03-27",
                                     "2022-11-09"),
                            dob = "2010-05-12")

## width is 10
plot_date_to_triangle_multi(date = c("2027-03-27",
                                     "2022-11-09"),
                            dob = "2010-05-12",
                            width = 10)


## period starts on 1 July rather than 1 January
plot_date_to_triangle_multi(date = c("2027-03-27",
                                     "2022-11-09"),
                            dob = "2010-05-12",
                            month_start = "Jul")

## open age group starts at 10 years
plot_date_to_triangle_multi(date = c("2027-03-27",
                                     "2022-11-09"),
                            dob = "2003-05-12",
                            break_max = 10)
}
