% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_date_to_triangle.R
\name{plot_date_to_triangle_year}
\alias{plot_date_to_triangle_year}
\title{Depict the Lexis triangles created by
function 'date_to_triangle_year'}
\usage{
plot_date_to_triangle_year(
  date,
  dob,
  break_max = 100,
  open_last = TRUE,
  month_start = "Jan",
  label_year_start = TRUE,
  show_months = FALSE
)
}
\arguments{
\item{date}{Dates of events.}

\item{dob}{Dates of birth.}

\item{break_max}{An integer or \code{NULL}.
Defaults to 100.}

\item{open_last}{Whether the final age group
has no upper limit. Defaults to \code{TRUE}.}

\item{month_start}{An element of \code{\link[base]{month.name}},
or \code{\link[base]{month.abb}}. The period starts on
the first day of this month.}

\item{label_year_start}{Whether to label a period
by the calendar year at the beginning of the period
or the calendar year at the end. Not needed for periods
that start on 1 January. Defaults to \code{FALSE}.}

\item{show_months}{Whether to include vertical
lines showing boundaries between months.
Defaults to \code{FALSE}.}
}
\description{
Create plot illustrating how function
\code{\link{date_to_triangle_year}} works.
}
\details{
\code{plot_date_to_triangle_year} is typically used for
instruction or documentation, rather than for
actual data analysis.
}
\examples{
plot_date_to_triangle_year(date = c("2024-03-27",
                                    "2022-11-09"),
                           dob = "2020-01-01")

## July to June
plot_date_to_triangle_year(date = c("2024-03-27",
                                    "2022-11-09"),
                           dob = "2020-01-01",
                           month_start = "Jul")
}
