% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_age_group.R
\name{date_to_age_group_lifetab}
\alias{date_to_age_group_lifetab}
\title{Convert dates to age groups used in abridged life table}
\usage{
date_to_age_group_lifetab(date, dob, break_max = 100, as_factor = TRUE)
}
\arguments{
\item{date}{Date of death.}

\item{dob}{Dates of birth.
A vector of class \code{\link[base]{Date}},
or a vector that can be coerced to class
\code{Date} using function \code{\link[base]{as.Date}}.}

\item{break_max}{An integer or \code{NULL}.
Defaults to 100.}

\item{as_factor}{Whether the return value is a factor.
Defaults to \code{TRUE}.}
}
\value{
If \code{as_factor} is \code{TRUE}, then the return
value is a factor; otherwise it is a character vector.
The length of the return value equals the length
of \code{date} or the length of \code{dob}, whichever
is greater.
}
\description{
Given dates of death and dates of birth,
allocate the deaths to age groups. These
age groups are the ones typically used in
"abridged" (ie not single-year) life tables: "0", "1-4",
and "5-9", "10-14", "10-14", and so on up to the
highest age group.
}
\details{
\code{date} and \code{dob} must have the same length,
unless one of them has length 1, in which case the
length-1 argument is recycled.

\code{break_max} is used to specify
the oldest age group, which is always open
on the right.
If \code{break_max} is \code{NULL}, the oldest
age group is derived from the data.

When \code{as_factor} is \code{TRUE} the levels of
the factor include all intermediate age groups,
including age groups that not appear in the data.
}
\examples{
date_to_age_group_lifetab(date = c("2024-03-27", "2022-11-09"),
                          dob = c("2001-03-21", "2000-07-13"))

## replicate date of birth
date_to_age_group_lifetab(date = c("2024-03-27", "2022-11-09"),
                          dob = "2011-05-18")

## return non-factor
date_to_age_group_lifetab(date = c("2024-03-27", "2022-11-09"),
                          dob = "2011-05-18",
                          as_factor = FALSE)

## alternative specifications for oldest age group
date_to_age_group_lifetab(date = "2019-09-22",
                          dob = "1910-01-01")
date_to_age_group_lifetab(date = "2019-09-22",
                          dob = "1910-01-01",
                          break_max = 80)
}
\seealso{
Other functions for creating age groups are
\code{\link{date_to_age_group_year}},
\code{\link{date_to_age_group_multi}},
\code{\link{date_to_age_group_fert}},
\code{\link{date_to_age_group_custom}},
\code{\link{date_to_age_group_quarter}},
and \code{\link{date_to_age_group_month}}.
See \code{\link{make_labels_age_group}} for the rules
on constructing labels for age groups.
}
