% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_age_group.R
\name{date_to_age_multi}
\alias{date_to_age_multi}
\title{Convert dates to multi-year age groups}
\usage{
date_to_age_multi(
  date,
  dob,
  width = 5,
  break_max = 100,
  open_last = TRUE,
  as_factor = TRUE
)
}
\arguments{
\item{date}{Dates of events or measurements.
A vector of class \code{\link[base]{Date}},
or a vector that can be coerced to class
\code{Date} using function \code{\link[base]{as.Date}}.}

\item{dob}{Dates of birth.
A vector of class \code{\link[base]{Date}},
or a vector that can be coerced to class
\code{Date} using function \code{\link[base]{as.Date}}.}

\item{width}{The width in years of the age intervals.
A positive integer. Defaults to 5.}

\item{break_max}{An integer or \code{NULL}.
Defaults to 100.}

\item{open_last}{Whether the final age group
has no upper limit. Defaults to \code{TRUE}.}

\item{as_factor}{Whether the return value is a factor.
Defaults to \code{TRUE}.}
}
\value{
If \code{as_factor} is \code{TRUE}, then the
return value is a factor; otherwise it is a character vector.
The length of the return value equals the length
of \code{date} or the length of \code{dob}, whichever
is greater.
}
\description{
Given dates when events occurred or measurements were made,
together with dates of birth, allocate the events or
measurements to age groups. These
age groups all have the same width, except possibly
the final age group. The width is measured in years
and must be an integer.
}
\details{
\code{date} and \code{dob} must have the same length,
unless one of them has length 1, in which case the
length-1 argument is recycled.

\code{break_max} and \code{open_last} are used to specify
the oldest age group.
If \code{break_max} is non-\code{NULL} and
\code{open_last} is \code{TRUE}, the oldest
age group is \code{[break_max, Inf)} years. If
\code{break_max} is non-\code{NULL} and 
\code{open_last} is \code{FALSE}, the oldest age
group is \code{[break_max-width, break_max)} years.
If \code{break_max} is \code{NULL}, the oldest
age group is derived from the data.

When \code{as_factor} is \code{TRUE} the levels of
the factor include all intermediate age groups,
including age groups that not appear in the data.
}
\examples{
date_to_age_multi(date = c("2024-03-27", "2022-11-09"),
                        dob = c("2001-03-21", "2000-07-13"))

## alternative values for 'width'
date_to_age_multi(date = c("2024-03-27", "2022-11-09"),
                        dob = c("2001-03-21", "2000-07-13"),
                        width = 10)
date_to_age_multi(date = c("2024-03-27", "2022-11-09"),
                        dob = c("2001-03-21", "2000-07-13"),
                        width = 1)

## replicate date of birth
date_to_age_multi(date = c("2024-03-27", "2022-11-09"),
                        dob = "2011-05-18")

## return non-factor
date_to_age_multi(date = c("2024-03-27", "2022-11-09"),
                        dob = "2011-05-18",
                        as_factor = FALSE)

## alternative specifications for oldest age group
date_to_age_multi(date = "2019-09-22",
                        dob = "1910-01-01")
date_to_age_multi(date = "2019-09-22",
                        dob = "1910-01-01",
                        break_max = 80)
date_to_age_multi(date = "2019-09-22",
                        dob = "1910-01-01",
                        break_max = NULL)
date_to_age_multi(date = "2019-09-22",
                        dob = "1910-01-01",
                        break_max = NULL,
                        open_last = FALSE)
}
\seealso{
Other functions for creating age groups are
\code{\link{date_to_age_year}},
\code{\link{date_to_age_lifetab}},
\code{\link{date_to_age_births}},
\code{\link{date_to_age_custom}},
\code{\link{date_to_age_quarter}},
and \code{\link{date_to_age_month}}.
Other functions for working with multi-year intervals are
\code{\link{date_to_period_multi}},
\code{\link{date_to_cohort_multi}},
and \code{\link{date_to_triangle_multi}}.
See \code{\link{make_labels_age}} for the rules
on constructing labels for age groups.
Function \code{\link{plot_date_to_age_multi}} depicts
the intervals created by \code{date_to_age_multi}.
}
