% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_age.R
\name{date_to_age_quarter}
\alias{date_to_age_quarter}
\title{Convert dates to one-quarter age groups}
\usage{
date_to_age_quarter(
  date,
  dob,
  break_min = 0,
  break_max = 400,
  open_last = TRUE,
  as_factor = TRUE
)
}
\arguments{
\item{date}{Dates of events.
A vector of class \code{\link[base]{Date}},
or a vector that can be coerced to class
\code{Date} using function \code{\link[base]{as.Date}}.}

\item{dob}{Dates of birth.
A vector of class \code{\link[base]{Date}},
or a vector that can be coerced to class
\code{Date} using function \code{\link[base]{as.Date}}.}

\item{break_min}{An integer or \code{NULL}.
Defaults to 0.}

\item{break_max}{An integer or \code{NULL}.
Defaults to 400.}

\item{open_last}{Whether the final age group
has no upper limit. Defaults to \code{TRUE}.}

\item{as_factor}{Whether the return value is a factor.
Defaults to \code{TRUE}.}
}
\value{
If \code{as_factor} is \code{TRUE}, then the return
value is a factor; otherwise it is a character vector.
The length of the return value equals the length
of \code{date} or the length of \code{dob}, whichever
is greater.
}
\description{
Given the dates when events occurred,
and the dates of birth of the people experiencing the events,
allocate the events to age groups. The resulting
age groups all have widths of one quarter (ie three months),
except possibly the final age group.
}
\details{
A person belongs to age group \code{"x"} if that
person was exactly \code{x} quarters
old at their most recent birthday. For instance, a person
belongs to age group \code{"20q"} if that person had
their 5th birthday (= 20 quarters) two days ago.

\code{date} and \code{dob} must have the same length,
unless one of them has length 1, in which case the
length-1 argument is recycled.

If \code{break_min} or \code{break_max} is set to \code{NULL},
rather than to a specific value, then \code{date_to_age_year}
finds the narrowest range that accommodates the data.

When \code{open_last} is \code{TRUE}, the final age group
is "open", which means that it has no upper limit.
(The definition of an open interval in demography is
is different from the definition of an open interval
in mathematics.)

When \code{as_factor} is \code{TRUE} the levels of
the factor include all intermediate age groups,
including age groups that not appear in the data.
}
\examples{
date_to_age_quarter(date = c("2024-03-27",
                             "2022-11-09"),
                    dob = c("2001-03-21",
                            "2000-07-13"))

## specify highest age group
date_to_age_quarter(date = "2019-09-22",
                    dob = "2010-01-01",
                    break_max = 48)

## let lowest age group be set by the data
date_to_age_quarter(date = "2019-09-22",
                    dob = "2010-01-01",
                    break_min = NULL,
                    break_max = 48)

## make final age group closed
date_to_age_quarter(date = "2019-09-22",
                    dob = "2010-01-01",
                    break_min = NULL,
                    break_max = 48,
                    open_last = FALSE)
}
\seealso{
Other functions for creating age groups are
\code{\link{date_to_age_year}},
\code{\link{date_to_age_multi}},
\code{\link{date_to_age_lifetab}},
\code{\link{date_to_age_births}},
\code{\link{date_to_age_custom}},
and \code{\link{date_to_age_month}}.

Other functions for working with one-quarter intervals are
\code{\link{date_to_period_quarter}},
\code{\link{date_to_cohort_quarter}},
and \code{\link{date_to_triangle_quarter}}.

\code{\link{make_labels_age_quarter}} describes the rules
for constructing labels for quarter age groups.

\code{\link{plot_date_to_age_quarter}} creates plots
that illustrate how \code{date_to_age_quarter}
works.
}
