% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_age.R
\name{date_to_age_year}
\alias{date_to_age_year}
\title{Convert dates to one-year age groups}
\usage{
date_to_age_year(
  date,
  dob,
  break_min = 0,
  break_max = 100,
  open_last = TRUE,
  as_factor = TRUE
)
}
\arguments{
\item{date}{Dates of events.
A vector of class \code{\link[base]{Date}},
or a vector that can be coerced to class
\code{Date} using function \code{\link[base]{as.Date}}.}

\item{dob}{Dates of birth.
A vector of class \code{\link[base]{Date}},
or a vector that can be coerced to class
\code{Date} using function \code{\link[base]{as.Date}}.}

\item{break_min}{An integer or \code{NULL}.
Defaults to 0.}

\item{break_max}{An integer or \code{NULL}.
Defaults to 100.}

\item{open_last}{Whether the final age group
has no upper limit. Defaults to \code{TRUE}.}

\item{as_factor}{Whether the return value is a factor.
Defaults to \code{TRUE}.}
}
\value{
If \code{as_factor} is \code{TRUE}, then the return
value is a factor; otherwise it is a character vector.
The length of the return value equals the length
of \code{date} or the length of \code{dob}, whichever
is greater.
}
\description{
Given the dates when events occurred,
and the dates of birth of the people experiencing the events,
allocate the events to age groups.
All the resulting age groups have widths
of one year, except possibly the final age group.
}
\details{
A person belongs to age group \code{"x"} if that
person was exactly \code{x} years
old at their most recent birthday. For instance, a person
belongs to age group {"5"} if that person had their
5th birthday two days ago, and a person belongs to age
group \code{"0"} if that person had their 0th birthday
(ie was born) three months ago.

\code{date} and \code{dob} must have the same length,
unless one of them has length 1, in which case the
argument with length 1 is recycled.

If \code{break_min} or \code{break_max} is set to \code{NULL},
rather than to a specific value, then \code{date_to_age_year}
finds the narrowest range that accommodates the data.

When \code{open_last} is \code{TRUE}, the final age group
is "open", which means that it has no upper limit.
(The definition of an open interval in demography is
is different from the definition of an open interval
in mathematics.)

When \code{as_factor} is \code{TRUE} the levels of
the factor include all intermediate age groups,
including age groups that not appear in the data.
}
\examples{
date_to_age_year(date = c("2024-03-27",
                          "2022-11-09"),
                 dob = c("2001-03-21",
                         "2000-07-13"))

## replicate date of birth
date_to_age_year(date = c("2024-03-27",
                          "2022-11-09"),
                 dob = "2011-05-18")

## specify alternative values for the
## lowest and highest breaks
date_to_age_year(date = c("2024-03-27",
                          "2022-11-09"),
                 dob = c("2001-03-21",
                         "2000-07-13"),
                 break_min = 20,
                 break_max = 30)

## allow lowest age group to be set by the data
date_to_age_year(date = c("2024-03-27",
                          "2022-11-09"),
                 dob = c("2001-03-21",
                         "2000-07-13"),
                 break_min = NULL,
                 break_max = 30)

## highest age group is closed rather
## than open
date_to_age_year(date = c("2024-03-27",
                          "2022-11-09"),
                 dob = c("2001-03-21",
                         "2000-07-13"),
                 break_min = NULL,
                 break_max = 30,
                 open_last = FALSE)

## return non-factor
date_to_age_year(date = c("2024-03-27",
                          "2022-11-09"),
                 dob = "2011-05-18",
                 as_factor = FALSE)
}
\seealso{
Other functions for creating age groups are
\code{\link{date_to_age_multi}},
\code{\link{date_to_age_lifetab}},
\code{\link{date_to_age_births}},
\code{\link{date_to_age_custom}},
\code{\link{date_to_age_quarter}},
and \code{\link{date_to_age_month}}.

Other functions for working with one-year intervals are
\code{\link{date_to_period_year}},
\code{\link{date_to_cohort_year}},
and \code{\link{date_to_triangle_year}}.

\code{\link{make_labels_age}} describes the rules
for constructing labels for age groups.

\code{\link{plot_date_to_age_year}} creates plots
that illustrate how \code{\link{date_to_age_year}}
works.
}
