% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_cohort.R
\name{date_to_cohort}
\alias{date_to_cohort}
\alias{date_to_cohort_year}
\alias{date_to_cohort_multi}
\alias{date_to_cohort_quarter}
\alias{date_to_cohort_month}
\title{Converting dates to cohorts}
\usage{
date_to_cohort_year(date, first_month = "Jan", year_to = TRUE,
  break_min = NULL, open_left = TRUE, as_factor = TRUE)

date_to_cohort_multi(date, width = 5, origin = 2000,
  first_month = "Jan", break_min = NULL, open_left = TRUE,
  as_factor = TRUE)

date_to_cohort_quarter(date, break_min = NULL, open_left = TRUE,
  as_factor = TRUE)

date_to_cohort_month(date, break_min = NULL, open_left = TRUE,
  as_factor = TRUE)
}
\description{
Converting dates to cohorts
}
