% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_cohort.R
\name{date_to_cohort_custom}
\alias{date_to_cohort_custom}
\title{Convert dates to customized cohorts}
\usage{
date_to_cohort_custom(date, breaks, open_first = TRUE)
}
\arguments{
\item{date}{Dates of events defining cohorts.
A vector of class \code{\link[base]{Date}},
or a vector that can be coerced to class \code{Date}
via function \code{\link[base]{as.Date}}.}

\item{breaks}{Dates defining starts and ends of cohorts.
A vector of class \code{\link[base]{Date}},
or a vector that can be coerced to class \code{Date}
via function \code{\link[base]{as.Date}}.}

\item{open_first}{Whether the first cohort
has no lower limit. Defaults to \code{TRUE}.}
}
\value{
A factor with the same length as \code{date}.
}
\description{
Identify cohorts, based on dates of births or other events.
The cohorts have varying widths. All widths
are measured in whole years.
}
\details{
The interface for \code{date_to_cohort_custom} is the similar
to that of \code{\link{date_to_period_custom}}, except that
\code{date_to_cohort_custom} also has arguments \code{break_min}
and \code{open_first}.

\code{breaks} is
used to define the points where each cohort starts and finishes.
The dates in \code{breaks} must all be the first day of
the same month of the year. For instance,
\code{breaks} could consist of
the values \code{"2010-01-01"} and \code{"2017-01-01"},
but not \code{"2010-01-01"} and \code{"2017-01-02"},
or \code{"2010-01-01"} and \code{"2017-02-01"}.

The definition of the first cohort depends on the
value of \code{open_first}. If \code{open_first} is
is \code{TRUE} (the default), then the first cohort consists of
all dates before the first break. Setting
\code{open_first} to \code{TRUE} is essentially
the same as having no upper limit on the oldest age group.
If \code{open_first} is \code{FALSE}, then the first cohort
is defined by the first two dates.

The return value is a factor. The levels of this
factor contain all intermediate cohorts,
including cohorts that do not appear in the data.

If \code{date} contains \code{NA}, then the
levels of the factor created by \code{date_to_cohort_custom}
also contain \code{NA}.
}
\examples{
## cohorts start on 1 January
date_to_cohort_custom(date = c("2024-03-27",
                               "2018-11-09",
                               "2020-05-13",
                               "2021-03-02"),
                      breaks = c("2020-01-01",
                                 "2021-01-01",
                                 "2026-01-01"))

## cohorts start on 1 March
date_to_cohort_custom(date = c("2024-03-27",
                               "2018-11-09",
                               "2020-05-13",
                               "2021-03-02"),
                      breaks = c("2020-03-01",
                                 "2021-03-01",
                                 "2026-03-01"))

## first cohort is closed
date_to_cohort_custom(date = c("2024-03-27",
                               "2018-11-09",
                               "2020-05-13",
                               "2021-03-02"),
                      breaks = c("2018-01-01",
                                 "2021-01-01",
                                 "2026-01-01"),
                      open_first = FALSE)
}
\seealso{
Other functions for creating cohorts are
\code{\link{date_to_cohort_year}},
\code{\link{date_to_cohort_multi}},
\code{\link{date_to_cohort_quarter}},
and \code{\link{date_to_cohort_month}}.

Other functions for working with customised intervals are
\code{\link{date_to_age_custom}},
and \code{\link{date_to_period_custom}}.

See \code{\link{make_labels_period}} for the rules
on constructing labels for periods and cohorts.

\code{\link{plot_date_to_cohort_custom}} creates plots
that illustrate how \code{date_to_cohort_custom}
works.
}
