% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_cohort.R
\name{date_to_cohort_month}
\alias{date_to_cohort_month}
\title{Convert dates to one-month cohorts}
\usage{
date_to_cohort_month(date)
}
\arguments{
\item{date}{Dates of events defining cohorts (typically births).
A vector of class \code{\link[base]{Date}},
or a vector that can be coerced to class \code{Date}
via function \code{\link[base]{as.Date}}.}
}
\value{
A character vector with the same length as \code{date}.
}
\description{
Identify cohorts, based on dates of births or other events.
The cohorts have widths of one month.
}
\examples{
date_to_cohort_month(date = c("2024-03-27",
                              "2023-05-01",
                              "2022-11-09"))
}
\seealso{
The output from \code{date_to_cohort_month}
is often processed further using \code{\link{format_cohort_month}}.

Other functions for creating cohorts from dates are
\code{\link{date_to_cohort_year}} and
and \code{\link{date_to_cohort_quarter}}.

Other functions for creating one-month units
from dates are
\code{\link{date_to_age_month}},
\code{\link{date_to_period_month}},
and \code{\link{date_to_triangle_month}}.
The interface for \code{date_to_cohort_month} is identical
to that of \code{\link{date_to_period_month}}.
}
