% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_cohort.R
\name{date_to_cohort_quarter}
\alias{date_to_cohort_quarter}
\title{Convert dates to quarter cohorts}
\usage{
date_to_cohort_quarter(date, break_min = NULL, open_first = NULL,
  as_factor = TRUE)
}
\arguments{
\item{date}{Dates of events defining cohorts.
A vector of class \code{\link[base]{Date}},
or a vector that can be coerced to class \code{Date}
via function \code{\link[base]{as.Date}}.}

\item{break_min}{The start date of the first cohort,
or \code{NULL} (the default.)
If non-\code{NULL}, \code{break_min} can be a single value of
class \code{\link[base]{Date}}, or a value that
can be coerced to class \code{Date}
via function \code{\link[base]{as.Date}};
in either case, it must be the first day of
a quarter.}

\item{open_first}{Whether the first cohort
has no lower limit. If \code{break_min} is \code{NULL},
then \code{open_first} is ignored. If \code{break_min} is
non-\code{NULL}, then \code{open_first} defaults to
\code{TRUE}.}

\item{as_factor}{Whether the return value is a factor.
Defaults to \code{TRUE}.}
}
\value{
If \code{as_factor} is \code{TRUE}, then the return
value is a factor; otherwise it is a character vector.
The return value has the same length as \code{date}.
}
\description{
Allocate dates of events
to cohorts of length one quarter, ie three months.
The events are typically, but not always, births.
Q1 (the first quarter) starts on 1 January
and ends on 31 March; Q2 starts on 1 April and ends on
30 June; Q3 starts on 1 July and ends on 30 September;
Q4 starts on 1 October and ends on 31 December.
}
\details{
The interface for \code{date_to_cohort_quarter} is the similar
to that of \code{\link{date_to_period_quarter}}, except that
\code{date_to_cohort_quarter} also has arguments \code{break_min}
and \code{open_first}.

Supplying a date for \code{break_min} defines
the first cohort. It is similar
to supplying a value for \code{break_max} in
\code{\link{date_to_age_group_multi}}. If
\code{open_first} is \code{TRUE}, then \code{break_min}
forms the upper limit for the first cohort, and there is no lower
limit. If \code{open_first} is \code{FALSE}, then
\code{break_min} forms the lower limit for the first cohort.

When \code{as_factor} is \code{TRUE} the levels of
the factor include all intermediate cohorts,
including cohorts that not appear in the data.
}
\examples{
date_to_cohort_quarter(date = c("2024-03-27",
                                "2022-11-09"))

## Specify oldest cohort, with open_first
## at default value of TRUE
date_to_cohort_quarter(date = c("2024-03-27",
                                "2019-08-22",
                                "2022-11-09"),
                       break_min = "2020-01-01")

## Specify oldest cohort, with open_first = FALSE
date_to_cohort_quarter(date = c("2024-03-27",
                                "2019-08-22",
                                "2022-11-09"),
                    break_min = "2015-07-01",
                    open_first = FALSE)

## return non-factor
date_to_cohort_quarter(date = c("2024-03-27", "2022-11-09"),
                    as_factor = FALSE)
}
\seealso{
Other functions for creating cohorts are
\code{\link{date_to_cohort_year}},
\code{\link{date_to_cohort_multi}},
\code{\link{date_to_cohort_custom}},
and \code{\link{date_to_cohort_month}}.
Other functions for working with one-quarter intervals are
\code{\link{date_to_age_group_quarter}},
\code{\link{date_to_period_quarter}},
and \code{\link{date_to_triangle_quarter}}.
See \code{\link{make_labels_period_quarter}} for the rules
on constructing labels for periods and cohorts.
}
