% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_cohort.R
\name{date_to_cohort_year}
\alias{date_to_cohort_year}
\title{Convert dates to one-year cohorts}
\usage{
date_to_cohort_year(
  date,
  month_start = "Jan",
  label_year_start = TRUE,
  break_min = NULL,
  open_first = NULL
)
}
\arguments{
\item{date}{Dates of events defining cohorts.
A vector of class \code{\link[base]{Date}},
or a vector that can be coerced to class \code{Date}
via function \code{\link[base]{as.Date}}.}

\item{month_start}{An element of \code{\link[base]{month.name}},
or \code{\link[base]{month.abb}}. Cohorts start on
the first day of this month.}

\item{label_year_start}{Logical. Whether to label a cohort
by the calendar year at the beginning of the cohort
or the calendar year at the end. Not needed for cohorts
that start on 1 January. Defaults to \code{TRUE}.}

\item{break_min}{The start date of the first cohort,
or \code{NULL} (the default.)
If non-\code{NULL}, \code{break_min} can be a single value of
class \code{\link[base]{Date}}, or a value that
can be coerced to class \code{Date}
via function \code{\link[base]{as.Date}};
in either case, it must be the first day of
a month.}

\item{open_first}{Whether the first cohort
has no lower limit. If \code{break_min} is non-\code{NULL}
and \code{label_year_start} is \code{TRUE},
then \code{open_first} defaults to \code{TRUE};
otherwise it defaults to \code{FALSE}.}
}
\value{
A factor with the same length as \code{date}.
}
\description{
Identify cohorts, based on dates of births or other events.
The cohorts have widths of one year.
}
\details{
The interface for \code{date_to_cohort_year} is similar
to that of \code{\link{date_to_period_year}}, except that
\code{date_to_cohort_year} also has arguments \code{break_min}
and \code{open_first}.

\code{break_min} sets the limits of the first cohort,
and is similar to \code{break_max} with function
\code{\link{date_to_age_year}}. If no value for \code{break_min}
is supplied, then the limits of the first cohort are
determined by the data.

If \code{open_first} is \code{TRUE}, then the first cohort
has no lower limit, which is similar to having \code{open_last}
equal to \code{TRUE}  with function
\code{\link{date_to_age_year}}

If a value for \code{break_min} is supplied, then
cohorts start on the day and month specified by \code{break_min};
otherwise they start on first day of \code{month_start}.
Cohorts end one-year-minus-one-day later.
The default value for \code{month_start} is \code{"Jan"},
so when \code{break_min} is not specified,
cohorts by default start on 1 January and
end on 31 December. \code{month_start} can be a
full month name or an abbreviation.

If a cohort starts on 1 January, then the first day and last day
of the cohort belong to the same calendar year.
But if a cohort starts on any other day, then
the first day belongs to one calendar
year and the last day belongs to the next calendar year.
For instance, if a cohort extends from
1 July 2000 to 30 June 2001, then the first
day belongs to the year 2000, and the last
day belongs to the year 2001. Some authors
use the first year to label such cohorts,
and others use the last year.
For instance, if a cohort extends from
1 July 2000 to 30 June 2001, some authors
label the cohort \code{"2000"}, and
others label it \code{"2001"}. Function
\code{date_to_cohort_year} by default uses
the start year. To use the end year, set
\code{label_year_start} to \code{FALSE}.
If \code{open_first} is \code{TRUE}, then
\code{label_year_start} must be \code{TRUE}.

The return value is a factor. The levels of this
factor contain all intermediate cohorts,
including cohorts that do not appear in the data.

If \code{date} contains \code{NA}, then the
levels of the factor created by \code{date_to_cohort_year}
also contain \code{NA}.
}
\examples{
date_to_cohort_year(date = c("2024-03-27", "2022-11-09"))

## starts on 1 July rather than 1 January
date_to_cohort_year(date = c("2024-03-27", "2022-11-09"),
                    month_start = "Jul")

## periods starts on 1 July, rather than 1 January, and uses
## calendar year at end, rather than beginning, for the label
date_to_cohort_year(date = c("2024-03-27", "2022-11-09"),
                    month_start = "Jul",
                    label_year_start = FALSE)

## Specify oldest cohort, with open_first
## at default value of TRUE
date_to_cohort_year(date = c("2024-03-27",
                             "2019-08-22",
                             "2022-11-09"),
                    break_min = "2020-01-01")

## Specify oldest cohort, with open_first = FALSE
date_to_cohort_year(date = c("2024-03-27",
                             "2019-08-22",
                             "2022-11-09"),
                    break_min = "2015-07-01",
                    open_first = FALSE)
}
\seealso{
Other functions for creating
cohorts from dates are
\code{\link{date_to_cohort_multi}},
\code{\link{date_to_cohort_custom}},
\code{\link{date_to_cohort_quarter}},
and \code{\link{date_to_cohort_month}}.

Other functions for creating one-year
intervals from dates are
\code{\link{date_to_age_year}},
\code{\link{date_to_period_year}},
and \code{\link{date_to_triangle_year}}.

Existing labels for one-year cohorts
can be processed using \code{\link{clean_age}}
and \code{\link{format_cohort_year}}.
}
