% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_cohort.R
\name{date_to_cohort_year}
\alias{date_to_cohort_year}
\title{Convert dates to one-year cohorts}
\usage{
date_to_cohort_year(
  date,
  month_start = "Jan",
  label_year_start = TRUE,
  break_min = NULL,
  open_first = NULL,
  as_factor = TRUE
)
}
\arguments{
\item{date}{Dates of events defining cohorts.
A vector of class \code{\link[base]{Date}},
or a vector that can be coerced to class \code{Date}
via function \code{\link[base]{as.Date}}.}

\item{month_start}{An element of \code{\link[base]{month.name}},
or \code{\link[base]{month.abb}}. Cohorts start on
the first day of this month.}

\item{label_year_start}{Logical. Whether to label a cohort
by the calendar year at the beginning of the cohort
or the calendar year at the end. Not needed for cohorts
that start on 1 January. Defaults to \code{TRUE}.}

\item{break_min}{The start date of the first cohort,
or \code{NULL} (the default.)
If non-\code{NULL}, \code{break_min} can be a single value of
class \code{\link[base]{Date}}, or a value that
can be coerced to class \code{Date}
via function \code{\link[base]{as.Date}};
in either case, it must be the first day of
a month.}

\item{open_first}{Whether the first cohort
has no lower limit. If \code{break_min} is non-\code{NULL}
and \code{label_year_start} is \code{TRUE}, then
then \code{open_first} defaults to \code{TRUE};
otherwise it defaults to \code{FALSE}.}

\item{as_factor}{Whether the return value is a factor.
Defaults to \code{TRUE}.}
}
\value{
If \code{as_factor} is \code{TRUE}, then the return
value is a factor; otherwise it is a character vector.
The return value has the same length as \code{date}.
}
\description{
Allocate dates of events to cohorts with widths of one year.
The cohorts by default align with calendar years.
The events are typically, but not always, births.
}
\details{
The interface for \code{date_to_cohort_year} is similar
to that of \code{\link{date_to_period_year}}, except that
\code{date_to_cohort_year} also has arguments \code{break_min}
and \code{open_first}.

Supplying a date for \code{break_min} defines
the first cohort, and is similar to
to supplying a value for \code{break_max} in
\code{\link{date_to_age_year}}. If
\code{open_first} is \code{TRUE}, then \code{break_min}
forms the upper limit for the first cohort, and there is no lower
limit. If \code{open_first} is \code{FALSE}, then
\code{break_min} forms the lower limit for the first cohort.

If a value for \code{break_min} is supplied, then
cohorts start on the day and month specified by \code{break_min};
otherwise they start on first day of \code{month_start}.
Cohorts end one-year-minus-one-day later.
The default value for \code{month_start} is \code{"Jan"},
so when \code{break_min} is not specified,
cohorts by default start on 1 January and
end on 31 December. \code{month_start} can be a
full month name or an abbreviation.

If a cohort starts on 1 January, then the first day and last day
of the cohort belong to the same calendar year.
But if a cohort starts on any other day, then
the first day belongs to one calendar
year and the last day belongs to the next calendar year.
For instance, if a cohort extends from
1 July 2000 to 30 June 2001, then the first
day belongs to the year 2000, and the last
day belongs to the year 2001. Some authors
use the first year to label such cohorts,
and others use the last year.
For instance, if a cohort extends from
1 July 2000 to 30 June 2001, some authors
label the cohort \code{"2000"}, and
others label it \code{"2001"}. Function
\code{date_to_cohort_year} by default uses
the start year. To use the end year, set
\code{label_year_start} to \code{FALSE}.

If \code{open_first} is \code{TRUE}, then
\code{label_year_start} must also be \code{TRUE}.

When \code{as_factor} is \code{TRUE} the levels of
the factor include all intermediate cohorts,
including cohorts that not appear in the data.
}
\examples{
date_to_cohort_year(date = c("2024-03-27", "2022-11-09"))

## July to June
date_to_cohort_year(date = c("2024-03-27", "2022-11-09"),
                    month_start = "Jul")

## July to June, using the calendar year at
## the end for the label
date_to_cohort_year(date = c("2024-03-27", "2022-11-09"),
                    month_start = "Jul",
                    label_year_start = FALSE)

## Specify oldest cohort, with open_first
## at default value of TRUE
date_to_cohort_year(date = c("2024-03-27",
                             "2019-08-22",
                             "2022-11-09"),
                    break_min = "2020-01-01")

## Specify oldest cohort, with open_first = FALSE
date_to_cohort_year(date = c("2024-03-27",
                             "2019-08-22",
                             "2022-11-09"),
                    break_min = "2015-07-01",
                    open_first = FALSE)

## return non-factor
date_to_cohort_year(date = c("2024-03-27", "2022-11-09"),
                    as_factor = FALSE)
}
\seealso{
Other functions for creating cohorts are
\code{\link{date_to_cohort_multi}},
\code{\link{date_to_cohort_custom}},
\code{\link{date_to_cohort_quarter}},
and \code{\link{date_to_cohort_month}}.
Other functions for working with one-year intervals are
\code{\link{date_to_age_year}},
\code{\link{date_to_period_year}},
and \code{\link{date_to_triangle_year}}.
See \code{\link{make_labels_period}} for the rules
on constructing labels for periods and cohorts.
}
