% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_period.R
\name{date_to_period_year}
\alias{date_to_period_year}
\alias{date_to_period_month}
\alias{date_to_period_quarter}
\title{\code{year_min} and \code{year_max} are for when 'date' does not
cover the entire range that is wanted}
\usage{
date_to_period_year(date, year_min = -Inf, year_max = Inf,
  first_month = "Jan", year_to = TRUE, as_factor = TRUE)

date_to_period_month(date, as_factor = TRUE)

date_to_period_quarter(date, as_factor = TRUE)
}
\description{
\code{year_min} and \code{year_max} are for when 'date' does not
cover the entire range that is wanted
}
