% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_period.R
\name{date_to_period}
\alias{date_to_period}
\alias{date_to_period_year}
\alias{date_to_period_multi}
\alias{date_to_period_quarter}
\alias{date_to_period_month}
\title{Converting dates to periods
\code{break_min} and \code{break_max} are for when 'date' does not
cover the entire range that is wanted}
\usage{
date_to_period_year(date, first_month = "Jan", year_to = TRUE,
  as_factor = TRUE)

date_to_period_multi(date, width = 5, origin = 2000,
  first_month = "Jan", as_factor = TRUE)

date_to_period_quarter(date, as_factor = TRUE)

date_to_period_month(date, as_factor = TRUE)
}
\description{
Converting dates to periods
\code{break_min} and \code{break_max} are for when 'date' does not
cover the entire range that is wanted
}
