% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_period.R
\name{date_to_period_multi}
\alias{date_to_period_multi}
\title{Convert dates to multi-year periods}
\usage{
date_to_period_multi(
  date,
  width = 5,
  origin = 2000,
  month_start = "Jan",
  as_factor = TRUE
)
}
\arguments{
\item{date}{Dates of events or measurements.}

\item{width}{The length, in whole years, of the periods.
Defaults to 5.}

\item{origin}{An integer. Defaults to 2000.}

\item{month_start}{An element of \code{\link[base]{month.name}},
or \code{\link[base]{month.abb}}. The period starts on
the first day of this month.}

\item{as_factor}{Whether the return value is a factor.
Defaults to \code{TRUE}.}
}
\value{
If \code{as_factor} is \code{TRUE}, then the return
value is a factor; otherwise it is a character vector.
The return value has the same length as \code{date}.
}
\description{
Allocate dates to multi-year periods. The periods all have
the same width, which by default is 5 years.
}
\details{
\code{date} is a vector of class \code{\link[base]{Date}},
or can be coerced to class \code{Date}
via function \code{\link[base]{as.Date}}.

Periods are \code{width} years long.
They start on the first day of \code{month_start},
and end \code{width}-years-minus-one-day later.
The default value for \code{month_start} is \code{"Jan"},
so periods by default start on 1 January and
end on 31 December. \code{month_start} can be a
full month name or an abbreviation.

The location of the periods can be shifted
by using different values for \code{origin}.

When \code{as_factor} is \code{TRUE} the levels of
the factor include all intermediate periods,
including periods that not appear in the data.
}
\examples{
date_to_period_multi(date = c("2024-03-27",
                              "2018-11-09",
                              "2021-03-02"))

## width is 10
date_to_period_multi(date = c("2024-03-27",
                              "2018-11-09",
                              "2021-03-02"),
                     width = 5)

## origin is 2001
date_to_period_multi(date = c("2024-03-27",
                              "2018-11-09",
                              "2021-03-02"),
                     origin = 2001)

## July to June
date_to_period_multi(date = c("2024-03-27",
                              "2018-11-09",
                              "2021-03-02"),
                     origin = 2001,
                     month_start = "Jul")

## return non-factor
date_to_period_multi(date = c("2024-03-27",
                              "2018-11-09",
                              "2021-03-02"),
                     as_factor = FALSE)
}
\seealso{
Other functions for creating periods are
\code{\link{date_to_period_year}},
\code{\link{date_to_period_custom}},
\code{\link{date_to_period_quarter}},
and \code{\link{date_to_period_month}}.
Other functions for working with multi-year intervals are
\code{\link{date_to_age_multi}},
\code{\link{date_to_cohort_multi}},
and \code{\link{date_to_triangle_multi}}.
See \code{\link{make_labels_period}} for the rules
on constructing labels for periods. Function
\code{\link{plot_date_to_period_multi}} depicts
the intervals created by \code{date_to_period_multi}.
}
