% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_period.R
\name{date_to_period_quarter}
\alias{date_to_period_quarter}
\title{Convert dates to quarter (three-month) periods}
\usage{
date_to_period_quarter(date)
}
\arguments{
\item{date}{Dates of events or measurements.
A vector of class \code{\link[base]{Date}},
or a vector than can be coerced to class \code{Date}
via function \code{\link[base]{as.Date}}.}
}
\value{
A character vector with the same length as \code{date}.
}
\description{
Allocate dates to periods. The
periods have widths of one quarter.
Quarters are defined as follows:
\tabular{lll}{
  \strong{Quarter} \tab \strong{Start} \tab \strong{End} \cr
  Q1 \tab 1 January \tab 31 March \cr
  Q2 \tab 1 April \tab 30 June \cr
  Q3 \tab 1 July \tab 30 September \cr
  Q4 \tab 1 October \tab 31 December
}
}
\examples{
date_to_period_quarter(date = c("2024-03-27",
                                "2020-01-03",
                                "2022-11-09"))
}
\seealso{
The output from \code{date_to_period_quarter}
is often processed further using \code{\link{format_period_quarter}}.

Other functions for creating periods from dates are
\code{\link{date_to_period_year}} and
and \code{\link{date_to_period_month}}.

Other functions for creating one-quarter units
from dates are
\code{\link{date_to_age_quarter}},
\code{\link{date_to_cohort_quarter}},
and \code{\link{date_to_triangle_quarter}}.
The interface for \code{date_to_period_quarter} is identical
to that of \code{\link{date_to_cohort_quarter}}.
}
