% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_period.R
\name{date_to_period_year}
\alias{date_to_period_year}
\title{Convert dates to one-year periods}
\usage{
date_to_period_year(
  date,
  month_start = "Jan",
  label_year_start = TRUE,
  as_factor = TRUE
)
}
\arguments{
\item{date}{Dates of events or measurements.}

\item{month_start}{An element of \code{\link[base]{month.name}},
or \code{\link[base]{month.abb}}. Each period starts on
the first day of this month.}

\item{label_year_start}{Whether to label a period
by the calendar year at the beginning of the period
or the calendar year at the end. Not needed for periods
that start on 1 January. Defaults to \code{TRUE}.}

\item{as_factor}{Whether the return value is a factor.
Defaults to \code{TRUE}.}
}
\value{
If \code{as_factor} is \code{TRUE}, then the return
value is a factor; otherwise it is a character vector.
The return value has the same length as \code{date}.
}
\description{
Allocate dates to one-year periods. The one-year periods are,
by default, calendar years.
}
\details{
\code{date} is a vector of class \code{\link[base]{Date}},
or can be coerced to class \code{Date}
via function \code{\link[base]{as.Date}}.

Periods start on the first day of \code{month_start},
and end one-year-minus-one-day later.
The default value for \code{month_start} is \code{"Jan"},
so periods by default start on 1 January and
end on 31 December. \code{month_start} can be a
full month name or an abbreviation.

If a period starts on 1 January, then the first day and last day
of the period belong to the same calendar year.
But if a period starts on any other day, then
the first day belongs to one calendar
year and the last day belongs to the next calendar year.
For instance, if a period extends from
1 July 2000 to 30 June 2001, then the first
day belongs to the year 2000, and the last
day belongs to the year 2001. Some authors
use the first year to label such periods,
and others use the last year.
For instance, if a period extends from
1 July 2000 to 30 June 2001, some authors
label the period \code{"2000"}, and
others label it \code{"2001"}. Function
\code{date_to_period_year} by default uses
the start year. To use the end year, set
\code{label_year_start} to \code{FALSE}.

When \code{as_factor} is \code{TRUE} the levels of
the factor include all intermediate periods,
including periods that not appear in the data.
}
\examples{
date_to_period_year(date = c("2024-03-27", "2022-11-09"))

## period starts on 1 July, not 1 January
date_to_period_year(date = c("2024-03-27", "2022-11-09"),
                    month_start = "Jul")

## period starts on 1 July, using the calendar year at
## the end, rather than beginning, for the label
date_to_period_year(date = c("2024-03-27", "2022-11-09"),
                    month_start = "Jul",
                    label_year_start = FALSE)

## return non-factor
date_to_period_year(date = c("2024-03-27", "2022-11-09"),
                    as_factor = FALSE)
}
\seealso{
Other functions for creating periods are
\code{\link{date_to_period_multi}},
\code{\link{date_to_period_custom}},
\code{\link{date_to_period_quarter}},
and \code{\link{date_to_period_month}}.

Other functions for working with one-year intervals are
\code{\link{date_to_age_year}},
\code{\link{date_to_cohort_year}},
and \code{\link{date_to_triangle_year}}.

\code{\link{make_labels_period}} describes the rules
for constructing labels for periods.

\code{\link{plot_date_to_period_year}} creates plots
that illustrate how \code{date_to_period_year}
works.
}
