% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_triangle.R
\name{date_to_triangle}
\alias{date_to_triangle}
\alias{date_to_triangle_year}
\alias{date_to_triangle_fert}
\alias{date_to_triangle_multi}
\alias{date_to_triangle_quarter}
\alias{date_to_triangle_month}
\title{Use dates and dates of birth to calculate Lexis triangles}
\usage{
date_to_triangle_year(date, dob, break_max = 100, open_right = TRUE,
  first_month = "Jan", as_factor = TRUE)

date_to_triangle_fert(date, dob, width = 5, break_min = 15,
  break_max = 50, recode_up = FALSE, recode_down = FALSE,
  origin = 2000, first_month = "Jan", as_factor = TRUE)

date_to_triangle_multi(date, dob, width = 5, break_max = 100,
  open_right = TRUE, origin = 2000, first_month = "Jan",
  as_factor = TRUE)

date_to_triangle_quarter(date, dob, break_max = 400, open_right = TRUE,
  as_factor = TRUE)

date_to_triangle_month(date, dob, break_max = 1200, open_right = TRUE,
  as_factor = TRUE)
}
\description{
Use dates and dates of birth to calculate Lexis triangles
}
