% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_triangle.R
\name{date_to_triangle_multi}
\alias{date_to_triangle_multi}
\title{Convert dates to multi-year Lexis triangles}
\usage{
date_to_triangle_multi(date, dob, width = 5, break_max = 100,
  open_right = TRUE, origin = 2000, month_start = "Jan",
  as_factor = TRUE)
}
\arguments{
\item{date}{Dates of events or measurements.}

\item{dob}{Dates of birth.}

\item{width}{The width in years of the periods and
age groups. A positive integer defaulting to 5.}

\item{break_max}{An integer or \code{NULL}.
Defaults to 100.}

\item{open_right}{Whether the final age group
has no upper limit. Defaults to \code{TRUE}.}

\item{origin}{An integer. Defaults to 2000.}

\item{month_start}{An element of \code{\link[base]{month.name}},
or \code{\link[base]{month.abb}}. The period starts on
the first day of this month.}

\item{as_factor}{Whether the return value is a factor.
Defaults to \code{TRUE}.}
}
\value{
If \code{as_factor} is \code{TRUE}, then the return
value is a factor; otherwise it is a character vector.
The return value has the same length as \code{date}.
}
\description{
Given dates when events occurred, together with dates of birth,
allocate the events to Lexis triangles.
All the Lexis triangles have the same width,
which by default is 5 years, though the final upper triangle
typically has no age limit.
}
\details{
An event occurring during period \code{[t, t+n)} to
a person in age group \code{[a, a+n)} belongs to an
upper Lexis triangle if the person was already in age group
\code{[a, a+n)} at time \code{t}, and belongs to a
lower Lexis triangle if the person joins age group
\code{[a, a+n)} during period \code{[t, t+n)}.

Information on the Lexis triangle of an event, in addition
to the period and age group, allows that event to
be allocated to a birth cohort. The ability to allocate
events to birth cohorts is essential for demographic accounting.

\code{date} and \code{dob} are both vectors of class
\code{\link[base]{Date}}, or vectors that can be coerced to class
\code{Date} via function \code{\link[base]{as.Date}}.

\code{date} and \code{dob} must have the same length,
unless one of them has length 1, in which case the
length-1 argument is recycled.

\code{break_max} and \code{open_right} are used to specify
the oldest age group.
When \code{break_max} is non-\code{NULL} and
\code{open_right} is \code{TRUE}, the oldest
age group is \code{[break_max, Inf)} years. When
\code{break_max} is non-\code{NULL} and 
\code{open_right} is \code{FALSE}, the oldest age
group is \code{[break_max-width, break_max)} years.

If \code{break_max} is \code{NULL}, \code{date_to_triangle_multi}
derives a value, based on the highest age in the data,
and the value for \code{open_right}.

Periods start on the first day of \code{month_start},
and end \code{width}-years-minus-one-day later.
The default value for \code{month_start} is \code{"Jan"},
so periods by default start on 1 January and
end on 31 December. \code{month_start} can be a
full month name or an abbreviation.

The location of the periods can be shifted
by using different values for \code{origin}.

The allocation of events to Lexis triangles becomes
tricky when the event and the date of birth share the same
month and day of the month (eg the event occurs on September
the 12th and the person's birthday is on
September the 12th), and the difference in years
is a multiple of \code{width}. Always allocating such
events to the lower triangle, or always allocating
them to the upper triangle, would lead to a slight
imbalance between upper and lower triangles. Instead,
if the dates match and it is first day of the month,
the event is allocated to \code{"Lower"}; if the dates match and
it is the second day of the month, the event is allocated to
\code{"Upper"}; if the dates match and it is the third day
of the month, the event is allocated to \code{"Lower"}; and so on.
See below for an example.

When \code{as_factor} is \code{TRUE}, the levels of
the factor includes both \code{"Lower"} and
\code{"Upper"}, even when they do not both appear
in the data.
}
\examples{
date_to_triangle_multi(date = c("2027-03-27",
                                "2022-11-09"),
                       dob = "2010-05-12")

## width is 10
date_to_triangle_multi(date = c("2027-03-27",
                                "2022-11-09"),
                       dob = "2010-05-12",
                       width = 10)


## July to June
date_to_triangle_multi(date = c("2027-03-27",
                                "2022-11-09"),
                       dob = "2010-05-12",
                       month_start = "Jul")

## open age group starts at 10 years
date_to_triangle_multi(date = c("2027-03-27",
                                "2022-11-09"),
                       dob = "2003-05-12",
                       break_max = 10)

## events and births occur on same month
## and day of month, and years differ by
## a multiple of width
date_to_triangle_multi(date = c("2020-03-01",
                                "2020-03-02",
                                "2020-03-03",
                                "2020-03-04",
                                "2020-03-05"),
                      dob = c("2010-03-01",
                              "2010-03-02",
                              "2010-03-03",
                              "2010-03-04",
                              "2010-03-05"))

## return non-factor
date_to_triangle_multi(date = c("2024-03-27",
                                "2022-11-09"),
                       dob = "2012-03-01",
                       as_factor = FALSE)
}
\seealso{
Other functions for creating Lexis triangles are
\code{\link{date_to_triangle_year}},
\code{\link{date_to_triangle_fert}},
\code{\link{date_to_triangle_quarter}},
and \code{\link{date_to_triangle_month}}.
\code{date_to_triangle_multi} is typically used in combination with
\code{\link{date_to_age_group_multi}} and
\code{\link{date_to_period_multi}}.
}
