% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_triangle.R
\name{date_to_triangle_quarter}
\alias{date_to_triangle_quarter}
\title{Convert dates to one-quarter (three-month) Lexis triangles}
\usage{
date_to_triangle_quarter(date, dob)
}
\arguments{
\item{date}{Dates of events.
A vector of class \code{\link[base]{Date}},
or a vector that can be coerced to class
\code{Date} using function \code{\link[base]{as.Date}}.}

\item{dob}{Dates of birth.
A vector of class \code{\link[base]{Date}},
or a vector that can be coerced to class
\code{Date} using function \code{\link[base]{as.Date}}.}
}
\value{
A character vector with the same length as \code{date}.
}
\description{
Use dates of events and dates of birth to create
one-quarter (three-month) Lexis triangles.
}
\details{
See \code{vignette("demprep")} for the definition
of Lexis triangles.

\code{date} and \code{dob} must have the same length,
unless one of them has length 1, in which case the
length-1 argument is recycled.
}
\examples{
date_to_triangle_quarter(date = c("2024-03-27",
                                  "2022-11-09"),
                         dob = "2020-01-01")
}
\seealso{
The output from \code{date_to_triangle_quarter}
is often processed further using function
\code{\link{format_triangle_quarter}}.

Other functions for creating
triangles from dates are
\code{\link{date_to_triangle_year}}
and \code{\link{date_to_triangle_month}}.

Other functions for creating one-quarter
units from dates are
\code{\link{date_to_age_quarter}}.
\code{\link{date_to_period_quarter}}, and
\code{\link{date_to_cohort_quarter}}.
}
