% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_triangle.R
\name{date_to_triangle_quarter}
\alias{date_to_triangle_quarter}
\title{Convert dates to quarter Lexis triangles}
\usage{
date_to_triangle_quarter(date, dob, break_max = 400, open_last = TRUE,
  as_factor = TRUE)
}
\arguments{
\item{date}{Dates of events or measurements.}

\item{dob}{Dates of birth.}

\item{break_max}{An integer or \code{NULL}.
Defaults to 400.}

\item{open_last}{Whether the final age group
has no upper limit. Defaults to \code{TRUE}.}

\item{as_factor}{Whether the return value is a factor.
Defaults to \code{TRUE}.}
}
\value{
If \code{as_factor} is \code{TRUE}, then the return
value is a factor; otherwise it is a character vector.
The return value has the same length as \code{date}.
}
\description{
Given dates when events occurred, together with dates of birth,
allocate the events to Lexis triangles.
All the Lexis triangles have widths
of one quarter (ie three months), though the final upper triangle
typically has no age limit.
}
\details{
An event occurring during period \code{[t, t+1)} to
a person in age group \code{[a, a+1)} belongs to an
upper Lexis triangle if the person was already in age group
\code{[a, a+1)} at time \code{t}, and belongs to a
lower Lexis triangle if the person joins age group
\code{[a, a+1)} during period \code{[t, t+1)}.

Information on the Lexis triangle of an event, in addition
to the period and age group, allows that event to
be allocated to a birth cohort. The ability to allocate
events to birth cohorts is essential for demographic accounting.

\code{date} and \code{dob} are both vectors of class
\code{\link[base]{Date}}, or vectors that can be coerced to class
\code{Date} via function \code{\link[base]{as.Date}}.

\code{date} and \code{dob} must have the same length,
unless one of them has length 1, in which case the
length-1 argument is recycled.

\code{break_max} and \code{open_last} are used to specify
the oldest age group.
When \code{break_max} is non-\code{NULL} and
\code{open_last} is \code{TRUE}, the oldest
age group is \code{[break_max, Inf)} years. When
\code{break_max} is non-\code{NULL} and 
\code{open_last} is \code{FALSE}, the oldest age
group is \code{[break_max-1, break_max)} years.

If \code{break_max} is \code{NULL}, \code{date_to_triangle_quarter}
derives a value, based on the highest age in the data,
and the value for \code{open_last}.

The allocation of events to Lexis triangles becomes
tricky when the event and the date of birth share the same
day of the month and are a multiple of three months apart.
Always allocating such
events to the lower triangle, or always allocating
them to the upper triangle, would lead to a slight
imbalance between upper and lower triangles. Instead,
if the dates match and it is first day of the month,
the event is allocated to \code{"Lower"}; if the dates match and
it is the second day of the month, the event is allocated to
\code{"Upper"}; if the dates match and it is the third day
of the month, the event is allocated to \code{"Lower"}; and so on.
See below for an example.

When \code{as_factor} is \code{TRUE}, the levels of
the factor includes both \code{"Lower"} and
\code{"Upper"}, even when they do not both appear
in the data.
}
\examples{
date_to_triangle_quarter(date = c("2024-03-27",
                                  "2022-11-09"),
                         dob = "2020-01-01")


## open age group starts at 40 quarters
date_to_triangle_quarter(date = c("2017-03-27",
                                  "2024-03-27"),
                         dob = "2010-01-01",
                         break_max = 40)

## events and births occur on same day of month
## and are a multiple of three months apart
date_to_triangle_quarter(date = c("2020-04-01",
                                  "2020-04-02",
                                  "2020-04-03",
                                  "2020-04-04",
                                  "2020-04-05"),
                         dob = c("2019-01-01",
                                 "2019-01-02",
                                 "2019-01-03",
                                 "2019-01-04",
                                 "2019-01-05"))

## return non-factor
date_to_triangle_quarter(date = c("2024-03-27",
                                  "2022-11-09"),
                         dob = "2012-03-01",
                         as_factor = FALSE)
}
\seealso{
Other functions for creating Lexis triangles are
\code{\link{date_to_triangle_year}},
\code{\link{date_to_triangle_multi}},
\code{\link{date_to_triangle_fert}},
and \code{\link{date_to_triangle_month}}.
\code{date_to_triangle_quarter} is typically used in combination with
\code{\link{date_to_age_group_quarter}}
and \code{\link{date_to_period_quarter}}.
}
