% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_triangle.R
\name{date_to_triangle_quarter}
\alias{date_to_triangle_quarter}
\title{Convert dates to quarter-length Lexis triangles}
\usage{
date_to_triangle_quarter(date, dob, break_max = 400, open_last = TRUE)
}
\arguments{
\item{date}{Dates of events.}

\item{dob}{Dates of birth.}

\item{break_max}{An integer or \code{NULL}.
Defaults to 400.}

\item{open_last}{Whether the final age group
has no upper limit. Defaults to \code{TRUE}.}
}
\value{
A factor with the same length as \code{date}.
}
\description{
Based on an age-time plan where age groups and periods both have
widths of one quarter (ie three months),
use dates of events and dates of birth
to allocate events to Lexis triangles.
}
\details{
The allocation of an event to a Lexis triangle depends
on the timing of the event and on the timing of changes
in age group. A person moves up into a new age group
once during each period. An event is allocated
to an upper Lexis triangle if the event occurs
before the move to the new age group.
An event is allocated to a lower Lexis triangle
if the event occurs with or after the move to the
new age group.

Consider, for instance, events occurring to a person
who was born on 3 March. \code{date_to_triangle_quarter}
allocates events to Lexis triangles as follows:
\tabular{ll}{
  \emph{Date of event} \tab \emph{Lexis triangle} \cr
  3 March to 31 March \tab \code{"Lower"} \cr
  1 April to 2 June \tab \code{"Upper"} \cr
  3 June to 31 September \tab \code{"Lower"} \cr
  1 October to 2 December \tab \code{"Upper"} \cr
  3 December to 31 December \tab \code{"Lower"} \cr
  etc \tab etc
}

\code{date} and \code{dob} are both vectors of class
\code{\link[base]{Date}}, or vectors that can be coerced to class
\code{Date} via function \code{\link[base]{as.Date}}.

\code{date} and \code{dob} must have the same length,
unless one of them has length 1, in which case the
length-1 argument is recycled.

\code{break_max} and \code{open_last} are used to specify
the oldest age group.
When \code{break_max} is non-\code{NULL} and
\code{open_last} is \code{TRUE}, the oldest
age group is \code{[break_max, Inf)} years. When
\code{break_max} is non-\code{NULL} and 
\code{open_last} is \code{FALSE}, the oldest age
group is \code{[break_max-1, break_max)} years.

If \code{break_max} is \code{NULL}, then
\code{date_to_triangle_quarter}
derives a value for \code{break_max}
based on the highest age in the data,
and the value for \code{open_last}.

The return value is a factor.
The levels of the factor include \code{"Lower"} and
\code{"Upper"}, even when \code{"Lower"}
and \code{"Upper"} do not both appear
in the data.
}
\examples{
date_to_triangle_quarter(date = c("2024-03-27",
                                  "2022-11-09"),
                         dob = "2020-01-01")


## open age group starts at 40 quarters
date_to_triangle_quarter(date = c("2017-03-27",
                                  "2024-03-27"),
                         dob = "2010-01-01",
                         break_max = 40)
}
\seealso{
Other functions for creating Lexis triangles are
\code{\link{date_to_triangle_year}},
\code{\link{date_to_triangle_multi}},
\code{\link{date_to_triangle_births}},
and \code{\link{date_to_triangle_month}}.

\code{date_to_triangle_quarter} is typically used in combination with
\code{\link{date_to_age_quarter}}
and \code{\link{date_to_period_quarter}}.
}
