% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demprep.R
\docType{package}
\name{demprep}
\alias{demprep}
\title{demprep: Tools for preparing demographic data}
\description{
General-purpose tools for preparing data
for demographic analysis. Typically, the
hardest part of preparing
data for demographic analysis is defining
age and time labels, so \code{demprep} focuses
particularly on these.
}
\section{Groups of functions}{


\strong{Calculating ages, cohorts, and periods from dates}

\tabular{lll}{
  \strong{age} \tab \strong{time} \tab \strong{cohort} \cr
  \code{\link{date_to_age_year}} \tab \code{\link{date_to_period_year}} \tab \code{\link{date_to_cohort_year}} \cr
  \code{\link{date_to_age_quarter}} \tab \code{\link{date_to_period_quarter}} \tab \code{\link{date_to_cohort_quarter}} \cr
  \code{\link{date_to_age_month}} \tab \code{\link{date_to_period_month}} \tab \code{\link{date_to_cohort_month}} \cr
}

\tabular{l}{
  \code{\link{date_to_triangle_year}} \cr
  \code{\link{date_to_triangle_quarter}} \cr
  \code{\link{date_to_triangle_month}} \cr
}

\code{\link{impute_date}}, \code{\link{impute_dob}}


\strong{Putting existing labels into the expected format}

\code{\link{clean_age}}, \code{\link{clean_cohort}}, \code{\link{clean_period}}

\code{\link{clean_age_df}}, \code{\link{clean_cohort_df}}, \code{\link{clean_period_df}}

\code{\link{is_valid_age}}, \code{\link{is_valid_cohort}}, \code{\link{is_valid_period}}


\strong{Generating a complete set of labels}

\tabular{llll}{
  Width \tab \strong{age} \tab \strong{time} \tab \strong{cohort} \cr
  Single year \tab \code{\link{format_age_year}} \tab \code{\link{format_period_year}} \tab \code{\link{format_cohort_year}} \cr
  Multi year, constant \tab \code{\link{format_age_multi}} \tab \code{\link{format_period_multi}} \tab \code{\link{format_cohort_multi}} \cr
  Multi year, varying \tab \code{\link{format_age_custom}} \tab \code{\link{format_period_custom}} \tab \code{\link{format_cohort_custom}} \cr
  Single quarter \tab \code{\link{format_age_quarter}} \tab \code{\link{format_period_quarter}} \tab \code{\link{format_cohort_quarter}} \cr
  Single month \tab \code{\link{format_age_month}} \tab \code{\link{format_period_month}} \tab \code{\link{format_cohort_month}} \cr
  Life table \tab \code{\link{format_age_lifetab}} \tab  \tab \cr
  Births \tab \code{\link{format_age_births}} \tab  \tab
}


\strong{Converting labels that use explicit dates}

\code{\link{as_date_range_year}}, \code{\link{as_date_range_quarter}}, \code{\link{as_date_range_month}} 

- note that can also be used to encode other periods/cohorts, eg weekly or daily data; irregular lenths.

\strong{Tabulating data}

Once the cross-classifying variables are all in the required format,
the data can be cross-tabulated into a multi-way
\code{\link[base]{array}}.

If (i) the data is at the individual level, but does not have
survey weights, or (ii) the data has already been aggregated,
then cross-tabulation can be carried out using
function \code{\link{dtabs}}.

If the data is at the individual level and does have survey weights,
then it may be better to use function \code{\link{dtabs_survey}}.
}

\section{Typical workflows}{


\strong{Individual-level data that uses dates of events}

\strong{Individual-level data that uses incomplete labels}

\strong{"Life table" age groups}

\strong{Births by age of monther}

\strong{Monthly data}
}

