% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_age.R
\name{format_age_custom}
\alias{format_age_custom}
\title{Create consistent, complete customized age groups}
\usage{
format_age_custom(x, breaks = NULL, open_last = TRUE)
}
\arguments{
\item{x}{A vector of age group labels.}

\item{breaks}{A vector of strictly increasing integer values.}

\item{open_last}{Whether the final age group
has no upper limit. Defaults to \code{TRUE}.}
}
\value{
A factor with length equal to \code{x}.
}
\description{
Given a vector of age group labels, create a \code{\link[base]{factor}}
that contains levels for all age groups, as
defined by the \code{breaks} argument. The labels may include an
open age group, ie an age group with no upper limit.
}
\details{
\code{format_age_custom} is the most flexible
of the \code{format_age} functions
in that the age groups can have any combination of widths.

\code{breaks} is used to specify the points at which
each age group starts and finishes, and \code{open_last}
is used to specify whether the final age group
has an upper limit. If \code{breaks} has
length \code{n}, then the
final age group is defined as follows:

\tabular{ll}{
    \strong{\code{open_last}} \tab \strong{Final age group} \cr
    \code{TRUE} \tab [\code{breaks[n]}, \code{Inf}) \cr
     \code{FALSE} \tab [\code{breaks[n-1]}, \code{breaks[n]})
}


If \code{x} contains \code{NA}, then the
levels of the factor created by \code{format_age_custom}
also contain \code{NA}.
}
\examples{
format_age_custom(x = c(22, 11, 85),
                  breaks = c(0, 15, 45, 70))
format_age_custom(x = c("90+", "19-40", "22", NA),
                  breaks = c(0, 15, 60))
format_age_custom(x = c("50-59", "19-40", "31"),
                  breaks = c(15, 45, 60),
                  open_last = FALSE)
}
\seealso{
Other functions for creating age groups are
\itemize{
  \item \code{\link{format_age_year}}
  \item \code{\link{format_age_multi}}
  \item \code{\link{format_age_lifetab}}
  \item \code{\link{format_age_births}}
  \item \code{\link{format_age_quarter}}
  \item \code{\link{format_age_month}}
}

\code{\link{date_to_age_year}} calculates
ages from dates.
}
