% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_age.R
\name{format_age_month}
\alias{format_age_month}
\title{Create consistent, complete one-month age groups}
\usage{
format_age_month(x, break_min = 0, break_max = 1200, open_last = TRUE)
}
\arguments{
\item{x}{A vector of age group labels.}

\item{break_min}{An integer or \code{NULL}.
Defaults to 0.}

\item{break_max}{An integer or \code{NULL}.
Defaults to 1200.}

\item{open_last}{Whether the final age group
has no upper limit. Defaults to \code{TRUE}.}
}
\value{
A factor with the same length as \code{x}.
}
\description{
Given a vector of age group labels, create a \code{\link[base]{factor}}
that contains levels for all ages between \code{break_min}
and \code{break_max}. The labels may contain an open age group,
ie an age group with no upper limit. Apart from the
open age group, all the age groups have a width of one month.
}
\details{
Even when an age
group between \code{break_min} and \code{break_max}
is not included in \code{x}, \code{factor_month}
still creates a level for it.

If \code{break_min} or \code{break_max} is set to \code{NULL},
rather than to a specific value, then \code{format_age_year}
finds the narrowest range that accommodates the data.

If \code{x} contains \code{NA}, then the
levels of the factor created by \code{format_age_month}
also contain \code{NA}.
}
\examples{
format_age_month(x = c(22, 0, 300))

format_age_month(x = c("3", NA, "12", "1400+"))

## specify highest age group
format_age_month(x = c("3", "12", "1400+"),
                 break_max = 24)

## let lowest age group be determined by the data
format_age_month(x = c("3", "12", "1400+"),
                 break_min = NULL,
                 break_max = 24)

## make final age group closed
format_age_month(x = c("3", "12"),
                 break_min = NULL,
                 break_max = 24,
                 open_last = FALSE)
}
\seealso{
Other functions for creating age groups are
\itemize{
  \item \code{\link{format_age_year}}
  \item \code{\link{format_age_multi}}
  \item \code{\link{format_age_lifetab}}
  \item \code{\link{format_age_births}}
  \item \code{\link{format_age_custom}}
  \item \code{\link{format_age_quarter}}
}

\code{\link{date_to_age_month}} calculates
ages from dates.
}
