% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_age.R
\name{format_age_multi}
\alias{format_age_multi}
\title{Put age group labels into the format required
for multi-year age groups}
\usage{
format_age_multi(
  x,
  width = 5,
  break_min = 0,
  break_max = 100,
  open_last = TRUE
)
}
\arguments{
\item{x}{A vector of age group labels.}

\item{width}{The width in years of the age
groups to be created. A positive integer.
Defaults to 5.}

\item{break_min}{An integer or \code{NULL}.
Defaults to 0.}

\item{break_max}{An integer or \code{NULL}.
Defaults to 100.}

\item{open_last}{Whether the final age group
has no upper limit. Defaults to \code{TRUE}.}
}
\value{
A factor with the same length as
\code{x}.
}
\description{
Given a vector of age group labels, create a factor
that contains levels for all ages between \code{break_min}
and \code{break_max}, and that may contain an open age group
(ie an age group with no upper limit.)
Apart from the open age group, the age groups
produced by \code{format_age_multi} all
have the width specified by \code{width}.
If, for instance, \code{break_min} is \code{0},
\code{break_max} is \code{100}, \code{width} is \code{5},
and \code{open_last} is \code{TRUE} (the defaults),
then \code{format_age_multi} creates
a factor with levels \code{"0-4"}, \code{"5-9"},
\dots, \code{"95-99"}, \code{"100+"}. Even when an age
group between \code{break_min} and \code{break_max}
is not included in \code{x}, \code{factor_age_year}
still creates a level for it.
}
\details{
If \code{break_min} or \code{break_max} is set to \code{NULL},
rather than to a specific value, then \code{format_age_year}
finds the narrowest range that accommodates the data.

All age groups in \code{x} must fall within the intervals
determined by \code{break_min} and \code{width},
except for age groups above \code{break_max}.

If \code{x} contains \code{NA}, then the
levels of the factor created by \code{format_age_multi}
also contain \code{NA}.
}
\examples{
format_age_multi(x = c(22, 11, 99, NA))

format_age_multi(x = c("10-14", "3", "100+", "77-79"))

format_age_multi(x = c("10-14", "3", "100+", "77-79"),
                 width = 10)

## age groups wider than the 'width' argument
## are allowed, provided they are above 'break_max'
format_age_multi(x = c("15", "30-39", "45-59"),
                 width = 10,
                 break_max = 40)

## allow 'break_min' and 'break_max' to be
## determined by the data
format_age_multi(x = c("22", "7", "30"),
                 break_min = NULL,
                 break_max = NULL)

## allow 'break_max' to be determined
## by the data, which includes an
## open age group
format_age_multi(x = c("17", "10+"),
                 break_max = NULL)

## oldest age group is closed
format_age_multi(x = c("10", "3", "77"),
                 open_last = FALSE)
}
\seealso{
Other functions for reformating
age group labels are 
\code{\link{format_age_year}},
\code{\link{format_age_lifetab}},
\code{\link{format_age_births}},
\code{\link{format_age_custom}},
\code{\link{format_age_quarter}},
and \code{\link{format_age_month}}.

\code{\link{date_to_age_multi}} creates
multi-year age groups from dates.
}
