% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_age.R
\name{format_age_quarter}
\alias{format_age_quarter}
\title{Put age group labels into format required
for one-quarter age groups}
\usage{
format_age_quarter(x, break_min = 0, break_max = 400, open_last = TRUE)
}
\arguments{
\item{x}{A vector of age group labels.}

\item{break_min}{An integer or \code{NULL}.
Defaults to 0.}

\item{break_max}{An integer or \code{NULL}.
Defaults to 400.}

\item{open_last}{Whether the final age group
has no upper limit. Defaults to \code{TRUE}.}
}
\value{
A factor with the same length as \code{x}.
}
\description{
Given a vector of age group labels, create a factor
that contains levels for all ages between \code{break_min}
and \code{break_max}, and that may contain an open age group
(ie an age group with no upper limit.)
If, for instance, \code{break_min} is \code{0},
\code{break_max} is \code{400}, and \code{open_last} is
\code{TRUE} (the defaults), then \code{format_age_year}
creates a factor with levels \code{"0"}, \code{"1"},
\dots, \code{"399"}, \code{"400+"}. Even when an age
group between \code{break_min} and \code{break_max}
is not included in \code{x}, \code{factor_age_quarter}
still creates a level for it.
}
\details{
A person belongs to age group \code{"x"} if that
person was exactly \code{x} quarters
old at their most recent birthday. For instance, a person
belongs to age group \code{"20"} if that person had
their 5th birthday (= 20 quarters) two days ago.

If \code{break_min} or \code{break_max} is set to \code{NULL},
rather than to a specific value, then \code{format_age_year}
finds the narrowest range that accommodates the data.

All age groups in \code{x} must be single-quarter age groups,
except above \code{break_max}, where open age groups
are allowed.
}
\examples{
format_age_quarter(x = c("10", "22", "500+"))

## specify highest age group
format_age_quarter(x = c("10", "22", "500+"),
                   break_max = 48)

## let lowest age group be determined by the data
format_age_quarter(x = c("10", "22", "500+"),
                   break_min = NULL,
                   break_max = 48)

## make final age group closed
format_age_quarter(x = c("10", "22"),
                   break_min = NULL,
                   break_max = 48,
                   open_last = FALSE)
}
\seealso{
Other functions for creating age groups are
\code{\link{format_age_year}},
\code{\link{format_age_multi}},
\code{\link{format_age_lifetab}},
\code{\link{format_age_births}},
\code{\link{format_age_custom}},
and \code{\link{format_age_month}}.

\code{\link{make_labels_age_quarter}} describes the rules
for constructing labels for quarter age groups.
}
