% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_age.R
\name{format_age_year}
\alias{format_age_year}
\title{Put age group labels into the format required
for single-year age groups}
\usage{
format_age_year(x, break_min = 0, break_max = 100, open_last = TRUE)
}
\arguments{
\item{x}{A vector of age group labels.}

\item{break_min}{An integer or \code{NULL}.
Defaults to 0.}

\item{break_max}{An integer or \code{NULL}.
Defaults to 100.}

\item{open_last}{Whether the final age group
has no upper limit. Defaults to \code{TRUE}.}
}
\value{
A factor with the same length as
\code{x}.
}
\description{
Given a vector of age group labels, create a factor
that contains levels for all ages between \code{break_min}
and \code{break_max}, and that may contain an open age group
(ie an age group with no upper limit.)
If, for instance, \code{break_min} is \code{0},
\code{break_max} is \code{100}, and \code{open_last} is
\code{TRUE} (the defaults), then \code{format_age_year}
creates a factor with levels \code{"0"}, \code{"1"},
\dots, \code{"99"}, \code{"100+"}. Even when an age
group between \code{break_min} and \code{break_max}
is not included in the data \code{x}, \code{factor_age_year}
still creates a level for it.
}
\details{
If \code{break_min} or \code{break_max} are \code{NULL},
then \code{factor_age_year} supplies values for them, 
based on the data. \code{break_min} is set to the lowest value
in the data, and \code{break_max} is set to the highest value,
except when the data contains an open age group,
in which case \code{break_max} is set to the lower limit
of the open age group.

All age groups in \code{x} must be single-year age groups,
except for age groups above \code{break_max}.
}
\examples{
format_age_year(x = c("10", "3", "100+", "77"))

## multi-year age groups are allowed,
## provided they are above 'break_max'
format_age_year(x = c("15", "30-39", "35+"),
                break_max = 30)

## allow 'break_min' and 'break_max' to be
## determined by the data
format_age_year(x = c("22", "7", "30"),
                break_min = NULL,
                break_max = NULL)

## allow 'break_max' to be determined
## by the data, which includes an
## open age group
format_age_year(x = c("17", "10+"),
                break_max = NULL)

## oldest age group is closed
format_age_year(x = c("10", "3", "77"),
                open_last = FALSE)
}
\seealso{
Other functions for reformating
age group labels are 
\code{\link{format_age_multi}},
\code{\link{format_age_lifetab}},
\code{\link{format_age_births}},
\code{\link{format_age_custom}},
\code{\link{format_age_quarter}},
and \code{\link{format_age_month}}.

Other functions for reformatting one-year
labels are \code{\link{format_period_year}} and
\code{\link{format_cohort_year}},

\code{\link{date_to_age_year}} creates
one-year age groups from dates.

\code{\link{make_labels_age}} describes the rules
for constructing labels for age groups.
}
