% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_cohort.R
\name{format_cohort_multi}
\alias{format_cohort_multi}
\title{Put cohort labels into the format required
for multi-year cohorts}
\usage{
format_cohort_multi(
  x,
  width = 5,
  origin = 2000,
  break_min = NULL,
  open_first = NULL
)
}
\arguments{
\item{x}{A vector of cohort labels.}

\item{width}{The length, in whole years, of the cohorts.
Defaults to 5.}

\item{origin}{An integer. Defaults to 2000.}

\item{break_min}{The lower limit of the cohort,
or \code{NULL} (the default).}

\item{open_first}{Whether the first cohort
has no lower limit.}
}
\value{
A factor with the same length as
\code{x}.
}
\description{
Given a vector of cohort labels, create a
\code{\link[base]{factor}} where the levels
contain a complete set of cohorts. 
If \code{open_first} is \code{TRUE}, then the earliest
cohort has no lower limit. (This is equivalent
to an open age group with no upper limit.)
}
\details{
The elements of \code{x} are typically multi-year
intervals such as \code{"1950-1960"},
\code{"2020-2025"}, or intervals that are
open on the left, eg \code{"<2000"}
or \code{<"1960"}. However, \code{x} can
also contain single-year labels, such as
\code{"2018"} or \code{"1974"}.

\code{open_first} defaults to \code{TRUE}
if a value for \code{break_min} is supplied,
or if any intervals in \code{x} is open,
and to \code{FALSE} otherwise.

If \code{x} contains \code{NA}, then the
levels of the factor created by \code{format_cohort_year}
also contain \code{NA}.
}
\examples{
format_cohort_multi(x = c("2000", "2005-2010", NA, "1995-1999"))

## 'break_min' is higher than the minimum of 'x'
format_cohort_multi(x = c("2000", "2005-2010", NA, "1995-1999"),
                    break_min = 2005)

## 'break_min' is lower then the minimum of 'x'
format_cohort_multi(x = c("2000", "2005-2010", NA, "1995-1999"),
                    break_min = 1990)

## 'break_min' supplied, but 'open_first' is FALSE
format_cohort_multi(x = c("2000", "2005-2010", NA, "1995-1999"),
                    break_min = 1990,
                    open_first = FALSE)

## non-default value for 'width'
format_cohort_multi(x = c("2000-2001", "2005-2010", NA, "1995-1999"),
                    width = 10)

## non-default value for 'origin', to shift labels by one year
format_cohort_multi(x = c("2000-2001", "2005-2010", NA, "1995-1999"),
                    width = 10,
                    origin = 2001)
}
\seealso{
Other functions for reformating
cohort labels are 
\code{\link{format_cohort_year}},
\code{\link{format_cohort_custom}},
\code{\link{format_cohort_quarter}},
and \code{\link{format_cohort_month}}.

\code{\link{date_to_cohort_multi}} creates
multi-year cohorts from dates.

\code{\link{make_labels_cohort}} describes the rules
for constructing labels for cohorts.
}
