% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_cohort.R
\name{format_cohort_multi}
\alias{format_cohort_multi}
\title{Put cohort labels into the format required
for multi-year cohorts}
\usage{
format_cohort_multi(
  x,
  width = 5,
  origin = 2000,
  break_min = NULL,
  open_first = NULL,
  month_start = "Jan",
  label_year_start = TRUE
)
}
\arguments{
\item{x}{A vector of cohort labels.}

\item{width}{The width, in whole years, of the cohorts
to be created. Defaults to 5.}

\item{origin}{An integer. Defaults to 2000.}

\item{break_min}{An integer or \code{NULL}
(the default). If an integer, it is the
year in which the oldest cohort begins.}

\item{open_first}{Whether the oldest cohort
has no lower limit.}

\item{month_start}{An element of \code{\link[base]{month.name}},
or \code{\link[base]{month.abb}}. Cohorts start on
the first day of this month.}

\item{label_year_start}{Logical. Whether to label a cohort
by the calendar year at the beginning of the cohort
or the calendar year at the end. Defaults to \code{TRUE}.}
}
\value{
A factor with the same length as
\code{x}.
}
\description{
Given a vector of cohort labels, create a
\code{\link[base]{factor}}  containing
levels for the earliest and latest cohorts
in \code{x}, and for all cohorts in between.
For instance, if the earliest cohort in \code{x}
is \code{"1990-1995"}, and the latest is \code{"2005-2010"},
then \code{format_cohort_multi} creates a factor
with levels \code{"1990-1995"}, \code{"1995-2000"},
\code{"2000-2005"}, and \code{"2005-2010"}.
All cohorts, with the possible exception of
a first "open" cohort, have the same width,
which is controlled by the \code{width} argument.
}
\details{
If \code{open_first} is \code{TRUE}, then the earliest
cohort has no lower limit. (This is equivalent
to an open age group with no upper limit.)

The elements of \code{x} can be
single-year labels such as \code{"2020"},
multi-year labels such as \code{"1950-1960"},
and intervals that are open on the left,
such as \code{"<2000"}.

As discussed in \code{\link{date_to_cohort_year}},
single-year labels such as \code{"2000"} are ambiguous.
Correctly aligning single-year and multi-year cohorts
requires knowing which month the single-year cohorts start on,
and whether single-year cohorts are labelled according
to the calendar year at the start of the cohort or the end.
\code{format_cohort_multi} assumes that
cohorts start in January, and that single-year cohorts
are labelled by calendar year at the start. These
settings can be changed using \code{month_start}
and \code{label_year_start}.

The multi-year labels produced
by \code{format_cohort_multi} are less ambiguous,
in that they always show a pair of years: the calendar
year at the start of the cohort and the calendar year at
the-end-of-the-cohort-plus-one-day. For instance, if a cohort
starts on 1 January 2000 and ends on 31 December 2000,
then the end of the cohort plus one day is 1 January 2001,
and the label is \code{"2000-2001"}.

\code{open_first} defaults to \code{TRUE}
if a value for \code{break_min} is supplied,
or if any intervals in \code{x} are open,
and to \code{FALSE} otherwise.

If \code{x} contains \code{NA}, then the
levels of the factor created by \code{format_cohort_multi}
also contain \code{NA}.
}
\examples{
format_cohort_multi(x = c(2000, 2005, NA, 2004))

format_cohort_multi(x = c("2000", "2005-2010", NA, "1995-1999"))

## contains open interval
format_cohort_multi(x = c("2000", "2005-2010", NA, "<1995"))

## changing the interpretation of the labels results in the
## reclassification of cohort "2000"
format_cohort_multi(x = c(2000, 2005, NA, 2004),
                    month_start = "Jul",
                    label_year_start = FALSE)
format_cohort_multi(x = c("2000", "2005-2010", NA, "1995-1999"),
                    month_start = "Jul",
                    label_year_start = FALSE)

## 'break_min' is higher than the minimum of 'x'
format_cohort_multi(x = c("2000", "2005-2010", NA, "1995-1999"),
                    break_min = 2005)

## 'break_min' is lower then the minimum of 'x'
format_cohort_multi(x = c("2000", "2005-2010", NA, "1995-1999"),
                    break_min = 1990)

## 'break_min' supplied, but 'open_first' is FALSE
format_cohort_multi(x = c("2000", "2005-2010", NA, "1995-1999"),
                    break_min = 1990,
                    open_first = FALSE)

## non-default value for 'width'
format_cohort_multi(x = c("2000", "2005-2010", NA, "1995-1999"),
                    width = 10)

## non-default value for 'origin', to shift labels by one year
format_cohort_multi(x = c("2000", "2005-2010", NA, "1995-1999"),
                    width = 10,
                    origin = 2001)
}
\seealso{
Other functions for reformating
cohort labels are 
\code{\link{format_cohort_year}},
\code{\link{format_cohort_custom}},
\code{\link{format_cohort_quarter}},
and \code{\link{format_cohort_month}}.

\code{\link{date_to_cohort_year}} calculates
cohorts from dates.
}
