% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_cohort.R
\name{format_cohort_year}
\alias{format_cohort_year}
\title{Put cohort labels into the format required
for single-year cohorts}
\usage{
format_cohort_year(x, break_min = NULL, open_first = NULL)
}
\arguments{
\item{x}{A vector of cohort labels.}

\item{break_min}{The lower limit of the cohort,
or \code{NULL} (the default).}

\item{open_first}{Whether the first cohort
has no lower limit.}
}
\value{
A factor with the same length as
\code{x}.
}
\description{
Given a vector of cohort labels, create a
\code{\link[base]{factor}} where the levels
contain a complete set of cohorts. 
If \code{open_first} is \code{TRUE}, then the earliest
cohort has no lower limit. (This is equivalent
to an open age group with no upper limit.)
}
\details{
The elements of \code{x} are typically single-year
labels, such as \code{"2001"} or \code{"2055"}, or
labels for cohorts
that are open on the left, such as \code{"<2000"}
or \code{<"1960"}. However, when \code{open_first}
is \code{TRUE}, \code{x} can also include
multi-year labels such as \code{"1950-1960"},
provided that the intervals are less
than \code{break_min} (and hence are absorbed
into the open age group).

\code{open_first} defaults to \code{TRUE}
if a value for \code{break_min} is supplied,
or if any intervals in \code{x} is open,
and to \code{FALSE} otherwise.

If \code{x} contains \code{NA}, then the
levels of the factor created by \code{format_cohort_year}
also contain \code{NA}.
}
\examples{
format_cohort_year(x = c("2000", "2005", NA, "2004"))

## 'x' contains an open interval, so
## 'open_first' defaults to TRUE
format_cohort_year(x = c("<2000", "2005", NA, "2004"))

## specify 'break_min', which makes also
## 'open_first' default to TRUE
format_cohort_year(x = c("2000", "2005", NA, "2004"),
                   break_min = 1998)

## set 'break_min' above minimum value of 'x'
format_cohort_year(x = c("2000", "2005", NA, "2004"),
                   break_min = 2001)

## multi-year labels are allowed if they are earlier than 'break_min'
format_cohort_year(x = c("1950-1980", "2005", NA, "2004"),
                   break_min = 2001)

## set 'open_first' to FALSE 
format_cohort_year(x = c("2000", "2005", NA, "2004"),
                   break_min = 1998,
                   open_first = FALSE)

## if no value for 'break_min' is supplied, and one or
## more of the labels refers to an open cohort,
## then the levels start with the most
## recent open cohort
format_cohort_year(x = c("2000", "2005", "<2002", "2004", "<2001"))
}
\seealso{
Other functions for reformatting
cohort labels are 
\code{\link{format_cohort_multi}},
\code{\link{format_cohort_custom}},
\code{\link{format_cohort_quarter}},
and \code{\link{format_cohort_month}}.

\code{\link{date_to_cohort_year}} creates
single-year cohorts from dates.

\code{\link{make_labels_cohort}} describes the rules
for constructing labels for cohorts.
}
