% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_period.R
\name{format_period_custom}
\alias{format_period_custom}
\title{Put period labels into the format required
for customized periods}
\usage{
format_period_custom(x, breaks)
}
\arguments{
\item{x}{A vector of character labels.}

\item{breaks}{A vector of strictly increasing integer values.}
}
\value{
A factor with the same length as
\code{x}.
}
\description{
Given a vector of period labels, create a
\code{\link[base]{factor}}
that contains levels for all periods
defined by \code{breaks}. \code{format_period_custom}
is the most flexible
of the \code{format_period} functions
in that the periods can have any combination of widths,
though the widths must be defined in whole numbers of years.
}
\details{
The elements of \code{x} are typically multi-year
intervals such as \code{"1950-1960"},
\code{"2020-2025"}. However, \code{x} can
also contain single-year labels, such as
\code{"2018"} or \code{"1974"}.

If \code{x} contains \code{NA}, then the
levels of the factor created by \code{format_period_year}
also contain \code{NA}.
}
\examples{
format_period_custom(x = c("2000-2001", "2005-2010", "1999"),
                     breaks = c(1990, 2000, 2020))

format_period_custom(x = c("2000-2001", "2005-2010", "1999"),
                     breaks = c(1995, 2005, 2010, 2020))
}
\seealso{
Other functions for reformating
period labels are 
\code{\link{format_period_year}},
\code{\link{format_period_multi}},
\code{\link{format_period_quarter}},
and \code{\link{format_period_month}}.

\code{\link{date_to_period_custom}} creates
customized periods from dates.

\code{\link{make_labels_period}} describes the rules
for constructing labels for periods.
}
