% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_period.R
\name{format_period_multi}
\alias{format_period_multi}
\title{Create consistent, complete multi-year periods}
\usage{
format_period_multi(
  x,
  width = 5,
  origin = 2000,
  month_start = "Jan",
  label_year_start = TRUE
)
}
\arguments{
\item{x}{A vector of period labels.}

\item{width}{The width, in whole years, of the periods
to be created. Defaults to 5.}

\item{origin}{An integer. Defaults to 2000.}

\item{month_start}{An element of \code{\link[base]{month.name}},
or \code{\link[base]{month.abb}}. Periods start on
the first day of this month.}

\item{label_year_start}{Logical. Whether to label a period
by the calendar year at the beginning of the period
or the calendar year at the end. Defaults to \code{TRUE}.}
}
\value{
A factor with the same length as
\code{x}.
}
\description{
Given a vector of period labels, create a
\code{\link[base]{factor}}  containing
levels for the earliest and latest periods
in \code{x}, and for all periods in between.
All periods have the same width,
which is controlled by the \code{width} argument.
}
\details{
The elements of \code{x} are typically multi-year
labels such as \code{"1950-1960"} or
\code{"2020-2025"} or single-year
labels such as \code{"2000"} or \code{"2025"}.
\code{x} cannot contain open intervals
such as \code{"<2020"}.

As discussed in \code{\link{date_to_period_year}},
single-year labels such as \code{"2000"} are ambiguous.
Correctly aligning single-year and multi-year periods
requires knowing which month the single-year periods start on,
which is controlled by the \code{month_start}
argument, and whether single-year periods
are labelled according 
to the calendar year at the start or end of the period,
which is controlled by the \code{label_year_start}
argument.

The location of the periods can be shifted
by using different values for \code{origin}.

If \code{x} contains \code{NA}, then the
levels of the factor created by \code{format_period_multi}
also contain \code{NA}.
}
\examples{
format_period_multi(x = c("2000-2001", "2005-2010", NA, "1996-1998"))

## use non-default value for 'width'
format_period_multi(x = c("2000-2001", "2005-2010", NA, "1996-1998"),
                    width = 10)

## use non-default value for 'origin' to shift periods
format_period_multi(x = c("2000-2001", "2005-2010", NA, "1996-1998"),
                    width = 10,
                    origin = 2001)

## Labels include single-year periods. Use default assumption
## that single-year periods start in January and/or
## are labelled by the calendar year at the start of
## the period
format_period_multi(x = c("2000-2001", "2005", NA, "1996-1998"))

## Change to assumption that single-year periods are
## start in July, and are labelled by the calendar
## year at the end of the period
format_period_multi(x = c("2000-2001", "2005", NA, "1996-1998"),
                    month_start = "Jul",
                    label_year_start = FALSE)

}
\seealso{
Other functions for reformating
period labels are
\itemize{
  \item \code{\link{format_period_year}}
  \item \code{\link{format_period_custom}}
  \item \code{\link{format_period_quarter}}
  \item \code{\link{format_period_month}}
}

\code{\link{date_to_period_year}} creates
 periods from dates.
}
