% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_period.R
\name{format_period_multi}
\alias{format_period_multi}
\title{Put period labels into the format required
for multi-year periods}
\usage{
format_period_multi(x, width = 5, origin = 2000)
}
\arguments{
\item{x}{A vector of period labels.}

\item{width}{The length, in whole years, of the periods.
Defaults to 5.}

\item{origin}{An integer. Defaults to 2000.}
}
\value{
A factor with the same length as
\code{x}.
}
\description{
Given a vector of period labels, create a
\code{\link[base]{factor}}  containing
levels for the earliest and latest periods
in \code{x}, and for all periods in between.
For instance, if the earliest period in \code{x}
is \code{"1990-1995"}, and the latest is \code{"2005-2010"},
then \code{format_period_multi} creates a factor
with levels \code{"1990-1995"}, \code{"1995-2000"},
\code{"2000-2005"},and \code{"2005-2010"}.
All periods in the return value have the same length,
which is controlled by the \code{width} parameter.
}
\details{
The elements of \code{x} must be multi-year
intervals such as \code{"1950-1960"},
\code{"2020-2025"}.

The location of the periods can be shifted
by using different values for \code{origin}.

If \code{x} contains \code{NA}, then the
levels of the factor created by \code{format_period_multi}
also contain \code{NA}.
}
\examples{
format_period_multi(x = c("2000-2001", "2005-2010", NA, "1996-1998"))

format_period_multi(x = c("2000-2001", "2005-2010", NA, "1996-1998"),
                    width = 10)

format_period_multi(x = c("2000-2001", "2005-2010", NA, "1996-1998"),
                    width = 10,
                    origin = 2001)
}
\seealso{
Other functions for reformating
period labels are 
\code{\link{format_period_year}},
\code{\link{format_period_custom}},
\code{\link{format_period_quarter}},
and \code{\link{format_period_month}}.

\code{\link{date_to_period_multi}} creates
multi-year periods from dates.

\code{\link{make_labels_period}} describes the rules
for constructing labels for periods.
}
