% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_triangle.R
\name{format_triangle_births}
\alias{format_triangle_births}
\title{Format labels for Lexis triangles
used when tabulating births}
\usage{
format_triangle_births(
  x,
  age,
  period,
  width = 5,
  break_min = 15,
  break_max = 50,
  recode_up = FALSE,
  recode_down = FALSE,
  month_start = "Jan",
  label_year_start = TRUE,
  origin = 2000
)
}
\arguments{
\item{x}{A vector of Lexis triangle labels.}

\item{age}{A vector of age groups, the same length
as \code{x}.}

\item{period}{A vector of periods, the same length
as \code{x}.}

\item{width}{The width, in whole years, of the
triangles to be created. Defaults to 5.}

\item{break_min}{An integer or \code{NULL}.
Defaults to 15.}

\item{break_max}{An integer or \code{NULL}.
Defaults to 50.}

\item{recode_up}{If \code{TRUE}, births to parents
aged less than \code{break_min} are treated as occurring to
people in the youngest repoductive age group.}

\item{recode_down}{If \code{TRUE}, births to parents
aged \code{break_max} or more are treated as
occurring to people in the oldest reproductive
age group.}

\item{month_start}{An element of \code{\link[base]{month.name}},
or \code{\link[base]{month.abb}}. Periods start on
the first day of this month.}

\item{label_year_start}{Logical. Whether a single-year
period in \code{x} is labelled using the calendar year
at the beginning of the period
or the calendar year at the end.
Defaults to \code{TRUE}.}

\item{origin}{An integer. Defaults to 2000.}
}
\value{
A factor with the same length as \code{x}.
}
\description{
Format labels for Lexis triangles to be used with
age groups and periods for tabulating births.
These age groups and periods must all have the same
length, which is set by the \code{width} parameter.
}
\details{
\code{age} and \code{period} define the
age groups and periods to which the
Lexis triangles
in \code{x} belong. These age groups and periods
can be narrower than \code{width}.
Age groups and periods can be single-year
(\code{"23"}, \code{"2023"}) or
multi-year (\code{"20-24"}, \code{"2020-2025"}).

\code{break_min} and \code{break_max} specify
the range of ages over which reproduction
is assumed to occur, and \code{recode_up} and
\code{recode_down} control the way that reported
ages outside this range are handled. See
\code{\link{format_age_births}} for details.

The values for \code{width}, \code{break_min},
\code{break_max}, and \code{origin} together define a new system
of Lexis triangles. \code{format_triangle_births}
calculates where the triangles defined by
\code{x}, \code{age}, and \code{period} fall within
this new system. For instance, if an upper triangle
defined by \code{x}, \code{age}, and \code{period}
falls entirely within a lower triangle in the new
system, then \code{format_triangle_births}
returns \code{"Lower"}.

If \code{recode_up} is \code{TRUE} and
an age is recoded upwards to fall within
the youngest reproductive age group,
then the corresponding Lexis triangle is set to
\code{"Lower"}. If \code{recode_down} is \code{FALSE}
and an age is recoded downwards to
fall within the lowest reproductive age group, then
the corresponding Lexis triangle is set to \code{"Upper"}.
}
\examples{
## we construct 'x', 'age', and 'period'
## from dates information ourselves before
## calling 'format_triangle_multi'
date_birth <- c("2024-03-27", "2022-11-09")
dob_mother <- "2000-01-01"
x <- date_to_triangle_year(date = date_birth,
                           dob = dob_mother,
                           month_start = "Jul")
age <- date_to_age_year(date = date_birth,
                        dob = dob_mother)
period <- date_to_period_year(date = date_birth,
                              month_start = "Jul")
format_triangle_births(x = x,
                       age = age,
                       period = period)

## someone else has constructed
## 'x', 'age', and 'period' from
## dates information
x_processed <- c("Lower", "Upper", "Upper")
age_processed <- c("20", "30-34", "25")
period_processed <- c("2002", "2015-2020", "2011")
format_triangle_births(x = x_processed,
                       age = age_processed,
                       period = period_processed)

## recode up and down
x <- c("Upper", "Lower", "Upper")
age <- c("10", "50-54", "25")
period <- c("2002", "2015-2020", "2011")
format_triangle_births(x = x,
                       age = age,
                       period = period,
                       recode_up = TRUE,
                       recode_down = TRUE)
}
\seealso{
Other functions for creating Lexis triangles are
\code{\link{format_triangle_year}},
\code{\link{format_triangle_multi}},
\code{\link{format_triangle_quarter}},
and \code{\link{format_triangle_month}}.

\code{\link{date_to_triangle_year}} creates
Lexis triangles from dates.
}
