% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_triangle.R
\name{format_triangle_month}
\alias{format_triangle_month}
\title{Format labels for quarter Lexis triangles}
\usage{
format_triangle_month(x, age, break_max = 1200, open_last = TRUE)
}
\arguments{
\item{x}{A vector of Lexis triangle labels.}

\item{age}{A vector of age groups, the same length
as \code{x}.}

\item{break_max}{An integer or \code{NULL}.
Defaults to 400.}

\item{open_last}{Whether the final age group
has no upper limit. Defaults to \code{TRUE}.}
}
\value{
A factor with the same length as
\code{x}.
}
\description{
Format labels for one-month
Lexis triangles to be used with one-month
age groups and periods.
}
\details{
\code{age} gives the age group to which each triangle
in \code{x} belongs. All age groups in \code{age}
must have a width of one month,
except for any open age groups.

\code{open_last} determines whether the
allocation of triangles needs to
account for an open age group, and \code{break_max}
specifies the cut-off for the open age group.
See \code{\link{format_age_month}} for a description
of how \code{open_last} and \code{break_max}
control age groups.

When \code{break_max} is \code{NULL},
the return value from \code{format_triangle_year}
is identical to \code{x}. When \code{break_max}
is non-\code{NULL}, the return value is as follows.

\tabular{lll}{
  \code{x} \tab \code{age} \tab return value \cr
  \code{"Lower"} \tab \code{<= break_max} \tab \code{"Lower"} \cr
  \code{"Lower"} \tab \code{> break_max} \tab \code{"Upper"} \cr
  \code{"Lower"} \tab \code{NA} \tab \code{NA} \cr
  \code{"Upper"} \tab \code{<= break_max} \tab \code{"Upper"} \cr
  \code{"Upper"} \tab \code{> break_max} \tab \code{"Upper"} \cr
  \code{"Upper"} \tab \code{NA} \tab \code{"Upper"} \cr
  \code{NA} \tab \code{<= break_max} \tab \code{NA} \cr
  \code{NA} \tab \code{> break_max} \tab \code{"Upper"} \cr
  \code{NA} \tab \code{NA} \tab \code{NA} \cr
}
}
\examples{
## we construct 'x' and 'age' from
## dates information ourselves before
## calling 'format_triangle_month'
date_original <- c("2024-03-27", "2022-11-09")
dob_original <- "2020-01-01"
x <- date_to_triangle_month(date = date_original,
                            dob = dob_original)
age <- date_to_age_month(date = date_original,
                         dob = dob_original)
format_triangle_month(x = x,
                      age = age)

## someone else has constructed
## 'x' and 'age' from
## dates information
x_processed <- c("Lower", "Lower", "Lower")
age_processed <- c("10", "12+", "5")
format_triangle_month(x = x_processed,
                      age = age_processed,
                      break_max = NULL)

## alternative value for 'break_max'
format_triangle_month(x = x_processed,
                      age = age_processed,
                      break_max = NULL)
}
\seealso{
Other functions for reformating
triangle labels are 
\code{\link{format_triangle_year}},
\code{\link{format_triangle_multi}},
\code{\link{format_triangle_births}},
and \code{\link{format_triangle_quarter}}.

\code{\link{date_to_triangle_month}} creates
one-month Lexis triangles from dates.
}
