% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_triangle.R
\name{format_triangle_multi}
\alias{format_triangle_multi}
\title{Format labels for multi-year Lexis triangles}
\usage{
format_triangle_multi(
  x,
  age,
  period,
  width = 5,
  break_max = 100,
  open_last = TRUE,
  month_start = "Jan",
  label_year_start = TRUE,
  origin = 2000
)
}
\arguments{
\item{x}{A vector of Lexis triangle labels.}

\item{age}{A vector of age groups, the same length
as \code{x}.}

\item{period}{A vector of periods, the same length
as \code{x}.}

\item{break_max}{An integer or \code{NULL}.
Defaults to 100.}

\item{open_last}{Whether the final age group
has no upper limit. Defaults to \code{TRUE}.}
}
\value{
A factor with the same length as
\code{x}.
}
\description{
Create labels for multi-year Lexis triangles to
be used with multi-year age groups and periods.
The age groups and periods must all have the same
length.
}
\details{
\code{age} and \code{period} define the
age groups and periods to which the
Lexis triangles
in \code{x} belong. Age groups can be single-year
(eg \code{"23"}), multi-year (eg \code{"20-24"})
or open (eg \code{"100+"}).
Periods can be single-year
(eg \code{"2023"}), multi-year (eg \code{"2020-2025"}).

\code{format_triangle_multi} returns different labels
from \code{x} in two situations: (i) when the intervals
in \code{age} and \code{period} are narrower than
\code{width}, so that aggregation is required, and
(ii) when the age groups are older than
\code{break_max + width}, in which case lower
triangles are recoded to upper.

\code{open_last} determines whether the
triangles need to account for an
open age group, and \code{break_max}
specifies the cut-off for the open age group.
See \code{\link{format_age_multi}} for a description
of how \code{open_last} and \code{break_max}
determine age groups.

\code{x} and \code{period} must be based on the same
starting month, e.g. if \code{x} uses years that
start in July and end in June,
then \code{period} must do so too. If
\code{x} was created using function
\code{\link{date_to_triangle_year}}
and \code{period} was created using function
\code{\link{date_to_period_year}},
then both should have used the
same value for \code{month_start}.
If \code{x} and \code{period} were not
calculated from raw dates data, 
then it may be necessary to check the
documentation for \code{x} and \code{period}
to see which
months of the year were used.
}
\examples{
## we construct 'x', 'age', and 'period'
## from dates information ourselves before
## calling 'format_triangle_multi'
x <- date_to_triangle_multi(date = c("2024-03-27",
                                     "2022-11-09"),
                            dob = "2020-01-01",
                            month_start = "Jul")
age <- date_to_age_multi(date = c("2024-03-27",
                                  "2022-11-09"),
                         dob = "2020-01-01")
period <- date_to_period_multi(date = c("2024-03-27",
                                        "2022-11-09"))
                               month_start = "Jul") ## same as above
format_triangle_multi(x = x,
                      age = age,
                      period = period)
format_triangle_multi(x = x,
                      age = age,
                      period = period,
                      break_max = 10)

## someone else has constructed
## 'x', 'age', and 'period' from
## dates information
x <- c("Lower", "Lower", "Lower")
age <- c("10", "15+", "5")
period <- c(2002, 2015, 2011)
format_triangle_multi(x = x,
                      age = age,
                      period = period,
                      break_max = 10)
}
\seealso{
Other functions for reformating
triangle labels are 
\code{\link{format_triangle_year}},
\code{\link{format_triangle_births}},
\code{\link{format_triangle_quarter}},
and \code{\link{format_triangle_month}}.

\code{\link{date_to_triangle_multi}} creates
multi-year Lexis triangles from dates.
}
