% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_labels_period.R
\name{make_labels_period}
\alias{make_labels_period}
\title{Make labels for periods measured in years}
\usage{
make_labels_period(
  breaks,
  open_first = FALSE,
  open_last = FALSE,
  label_year_start = TRUE,
  include_na = FALSE
)
}
\arguments{
\item{breaks}{A vector of class \code{\link[base]{Date}},
or a vector that can be coerced to class \code{Date}
via function \code{\link[base]{as.Date}}.}

\item{open_first}{Whether to append an open-ended
period to the start of the labels.
Defaults to \code{FALSE}.}

\item{open_last}{Whether to append an open-ended
period to the end of the labels.
Defaults to \code{FALSE}.}

\item{label_year_start}{Whether to label a period
by the calendar year at the beginning of the period.
Not needed for multi-year periods, or for single-year
periods that start on 1 January.
Defaults to \code{TRUE}.}

\item{include_na}{Whether to append an \code{NA} to
the end of the labels. Defaults to \code{FALSE}.}
}
\value{
A character vector.
}
\description{
Make labels for periods with lengths measured in whole years.
The labels follow standard demographic conventions, though
these conventions are sometimes ambiguous.
}
\details{
Periods are defined by the dates specified in \code{breaks}.
A period ordinarily starts at one date in \code{breaks} and ends
the day before the next date. For instance, if \code{breaks} is
\code{c("2000-01-01", "2001-01-01", "2002-01-01")},
one period begins on 2000-01-01 and ends on 2000-12-31,
and the other starts on 2001-01-01 and ends on 2001-12-31.
However, when \code{open_first} is \code{TRUE}, an 'open' period
extending indefinitely into the past is appended to the start
of the series, and when \code{open_last} is \code{TRUE},
an open period extending indefinitely into the future is
appended to the end. By default, \code{open_first}
and \code{open_last} are both \code{FALSE}.

If all periods have widths of one year,
then the labels consist of a single years, eg
\code{"2020", "2021", "2020"}. Otherwise, the
labels consist of start years and end years
separated by dashes, eg \code{"2020-2025", "2025-2030", "2030-2035"}.

Unfortunately, the practice of using a single year to label
one-year periods can be ambiguous.
If a period starts on 2020-01-01 and ends on 2020-12-31,
it is clear that the single-year
label should be \code{"2020"}. However, if a period
starts on 2020-07-01 and ends on 2021-06-30, it is
not obvious how the label should be constructed.
Some authorities use the calendar year at the \emph{start} of
the period, so that the label is
\code{"2020"}. Other authorities use the calendar
year at the \emph{end} of the period, so that the label
is \code{"2021"}. (In the latter case,
periods are often called "years to", eg "years to June".)

\code{make_labels_period} by default uses the start year
to make single-year labels. To use the end year,
set \code{label_year_start} to \code{FALSE}.

If \code{label_year_start} is \code{FALSE}, then
\code{open_right} must also be \code{FALSE}.

When \code{include_na} is \code{TRUE}, an \code{NA}
is added to the end of the labels. This can be useful
when dealing with dates that include \code{NA}s.

\code{breaks} can only have length 0 if \code{open_first}
and \code{open_last} are both \code{FALSE}. \code{breaks} can
have only have length 1 if at least one of \code{open_first}
and \code{open_last} is TRUE.
}
\examples{
## 5-year periods
make_labels_period(breaks = c("2005-01-01",
                              "2010-01-01",
                              "2015-01-01",
                              "2020-01-01"))

## one-year periods, using default setting where
## labels use calender year at start 
make_labels_period(breaks = c("2005-07-01",
                              "2006-07-01",
                              "2007-07-01"))

## this time with label using
## calendar year at end
make_labels_period(breaks = c("2005-07-01",
                              "2006-07-01",
                              "2007-07-01"),
                   label_year_start = FALSE)

## add an open period at the beginning...
make_labels_period(breaks = c("2005-01-01",
                              "2010-01-01",
                              "2015-01-01",
                              "2020-01-01"),
                   open_first = TRUE)

## ...and at the end
make_labels_period(breaks = c("2005-01-01",
                              "2010-01-01",
                              "2015-01-01",
                              "2020-01-01"),
                   open_first = TRUE,
                   open_last = TRUE)

## add an 'NA' label
make_labels_period(breaks = c("2005-07-01",
                              "2006-07-01",
                              "2007-07-01"),
                   include_na = TRUE)
}
\seealso{
To make labels for age groups, use
\code{\link{make_labels_age_group}}. There is
no \code{make_labels_cohort} function. To construct
labels for cohorts, just use \code{make_labels_period}.
To make labels for periods with widths of one quarter or one month,
use functions \code{\link{make_labels_period_quarter}} or
\code{\link{make_labels_period_month}}.
}
