% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_labels_period.R
\name{make_labels_period}
\alias{make_labels_period}
\title{Make complete labels for periods measured in years}
\usage{
make_labels_period(breaks, label_year_start = TRUE, include_na = FALSE)
}
\arguments{
\item{breaks}{A vector of class \code{\link[base]{Date}},
or a vector that can be coerced to class \code{Date}
via function \code{\link[base]{as.Date}}.}

\item{label_year_start}{Whether to label a period
by the calendar year at the beginning of the period.
Not needed for multi-year periods, or for single-year
periods that start on 1 January.
Defaults to \code{TRUE}.}

\item{include_na}{Whether to append an \code{NA} to
the end of the labels. Defaults to \code{FALSE}.}
}
\value{
A character vector.
}
\description{
Make labels for periods with lengths measured in whole years.
The function would not normally be
called directly by end users.
}
\details{
Periods are defined by the dates specified in \code{breaks}.
A period ordinarily starts at one date in \code{breaks} and ends
the day before the next date. For instance, if \code{breaks} is
\code{c("2000-01-01", "2001-01-01", "2002-01-01")},
one period begins on 2000-01-01 and ends on 2000-12-31,
and the other starts on 2001-01-01 and ends on 2001-12-31.

If all periods have widths of one year,
then the labels consist of a single years, eg
\code{"2020", "2021", "2020"}. Otherwise, the
labels consist of start years and end years
separated by dashes, eg \code{"2020-2025", "2025-2030", "2030-2035"}.

Unfortunately, the practice of using a single year to label
one-year periods can be ambiguous.
If a period starts on 2020-01-01 and ends on 2020-12-31,
it is clear that the single-year
label should be \code{"2020"}. However, if a period
starts on 2020-07-01 and ends on 2021-06-30,
some people use the calendar year at the
\emph{start}, so that the label is
\code{"2020"}, and others use the calendar
year at the \emph{end}, so that the label
is \code{"2021"}. (In the latter case,
periods are often called "years to", eg "years to June".)
\code{make_labels_period} by default uses the start year
to make single-year labels. To use the end year,
set \code{label_year_start} to \code{FALSE}.

When \code{include_na} is \code{TRUE}, an \code{NA}
is added to the end of the labels. This can be useful
when dealing with dates that include \code{NA}s.

If \code{breaks} has length 0 then \code{open_first}
must be \code{FALSE}, and if \code{breaks} has
length 1, then \code{open_first} must be \code{TRUE}.
}
\examples{
## 5-year periods
make_labels_period(breaks = c("2005-01-01",
                              "2010-01-01",
                              "2015-01-01",
                              "2020-01-01"))

## one-year periods, using default setting where
## labels use calender year at start 
make_labels_period(breaks = c("2005-07-01",
                              "2006-07-01",
                              "2007-07-01"))

## this time with label using
## calendar year at end
make_labels_period(breaks = c("2005-07-01",
                              "2006-07-01",
                              "2007-07-01"),
                   label_year_start = FALSE)

## add an 'NA' label
make_labels_period(breaks = c("2005-07-01",
                              "2006-07-01",
                              "2007-07-01"),
                   include_na = TRUE)
}
\seealso{
To make labels for age groups measured in years, use
\code{\link{make_labels_age}}, and to
make labels for cohorts, use \code{\link{make_labels_cohort}}.
To make labels for periods with widths of one quarter or one month,
use functions \code{\link{make_labels_period_quarter}} or
\code{\link{make_labels_period_month}}.
}
\keyword{internal}
