% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_labels_period.R
\name{make_labels_period}
\alias{make_labels_period}
\alias{make_labels_period_quarter}
\alias{make_labels_period_month}
\title{Make period labels}
\usage{
make_labels_period(breaks, open_left = FALSE, open_right = FALSE,
  label_year_start = TRUE, include_na = FALSE)

make_labels_period_quarter(break_min, break_max, open_left = FALSE,
  open_right = FALSE, include_na = FALSE)

make_labels_period_month(break_min, break_max, open_left = FALSE,
  open_right = FALSE, include_na = FALSE)
}
\description{
Make period labels
}
