% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_labels_period.R
\name{make_labels_period_month}
\alias{make_labels_period_month}
\title{Make labels for periods lengths of one month}
\usage{
make_labels_period_month(break_min, break_max, include_na = FALSE)
}
\arguments{
\item{break_min}{A value of class \code{\link[base]{Date}},
or a value that can be coerced to class \code{Date}
via function \code{\link[base]{as.Date}}.}

\item{break_max}{A value of class \code{\link[base]{Date}},
or a value that can be coerced to class \code{Date}
via function \code{\link[base]{as.Date}}.}

\item{include_na}{Whether to append an \code{NA} to
the end of the labels. Defaults to \code{FALSE}.}
}
\value{
A character vector.
}
\description{
Make labels for periods that are all one month long.
}
\details{
\code{break_min} and \code{break_max}
define the lower and upper limits of the periods.
\code{break_min} and \code{break_max} must both
be the start dates of months.

When \code{include_na} is \code{TRUE}, an \code{NA}
is added to the end of the labels. This can be useful
when dealing with data that include \code{NA}s.
}
\examples{
make_labels_period_month(break_min = "2005-01-01",
                         break_max = "2005-08-01")

## add an 'NA' label
make_labels_period_month(break_min = "2005-03-01",
                         break_max = "2006-04-01",
                         include_na = TRUE)
}
\seealso{
To make labels for month age groups, use
\code{\link{make_labels_age_group_month}},
and to make labels for month periods,
use \code{make_labels_period_month}.
To make labels for periods with widths measured in years,
use function \code{\link{make_labels_period}}, and to make
labels for quarter periods, use \code{\link{make_labels_period_quarter}}.
}
