% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_labels.R
\name{make_labels_quantities}
\alias{make_labels_quantities}
\title{Make Quantities labels}
\usage{
make_labels_quantities(x)
}
\arguments{
\item{x}{A list of integer vectors.}
}
\value{
A character vector
}
\description{
Make labels with the format expected for an
object of class \code{"Quantities"}
(as defined in the demarray package).
This function would not normally be called directly
by end users.
}
\details{
\code{x} is a list of integer vectors of length 2.
If elements of a vector are non-NA, then the second
element must be greater than the first.
The vectors cannot overlap.

The elements of \code{x} are converted into
labels as follows:

\tabular{ll}{
  Element \tab Label \cr
  \code{c(100L, 101L)} \tab \code{"100"} \cr
  \code{c(100L, 110L)} \tab \code{"100-109"} \cr
  \code{c(NA_integer_, 100L)} \tab \code{"<100"} \cr
  \code{c(100L, NA_integer_)} \tab \code{"100+"} \cr
  \code{c(NA_integer_, NA_integer_)} \tab \code{NA} \cr
}
}
\examples{
x <- list(c(0L, 1L),
          c(1L, 5L),
          c(NA, 0L),
          c(NA_integer_, NA__integer_),
          c(20L, NA),
          c(10L, 20L))
make_labels_quantities(x)
}
\seealso{
\code{\link{make_labels_categories}},
\code{\link{make_labels_triangles}},
\code{\link{make_labels_directions}},
\code{\link{make_labels_quantiles}},
\code{\link{make_labels_integers}},
\code{\link{make_labels_intervals}},
\code{\link{make_labels_quarters}},
\code{\link{make_labels_months}},
\code{\link{make_labels_dateranges}},
\code{\link{make_labels_datepoints}}
}
\keyword{internal}
