% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_date_to_age.R
\name{plot_date_to_age_month}
\alias{plot_date_to_age_month}
\title{Depict the intervals created by
function 'date_to_age_month'}
\usage{
plot_date_to_age_month(
  date,
  dob,
  break_max = 1200,
  open_last = TRUE,
  show_months = FALSE
)
}
\arguments{
\item{date}{Dates of events or measurements.
A vector of class \code{\link[base]{Date}},
or a vector that can be coerced to class
\code{Date} using function \code{\link[base]{as.Date}}.}

\item{dob}{Dates of birth.
A vector of class \code{\link[base]{Date}},
or a vector that can be coerced to class
\code{Date} using function \code{\link[base]{as.Date}}.}

\item{break_max}{An integer or \code{NULL}.
Defaults to 1200.}

\item{open_last}{Whether the final age group
has no upper limit. Defaults to \code{TRUE}.}

\item{show_months}{Whether to include vertical
lines showing boundaries between months.
Defaults to \code{FALSE}.}
}
\description{
Create plot illustrating the inputs and outputs
of function \code{\link{date_to_age_month}}.
}
\details{
This function is for learning about the
labelling conventions used in the \strong{dem} packages,
and about their implementation in package
\strong{demprep}. It would not normally be used
during actual data analysis.
}
\examples{
plot_date_to_age_month(date = c("2004-03-27", "2002-11-09"),
                             dob = c("2001-03-21", "2000-07-13"))

## alternative specifications for oldest age group
plot_date_to_age_month(date = "2019-09-22",
                             dob = "1910-01-01")
plot_date_to_age_month(date = "2019-09-22",
                             dob = "1910-01-01",
                             break_max = 320)
plot_date_to_age_month(date = "2019-09-22",
                             dob = "1910-01-01",
                             break_max = NULL)
plot_date_to_age_month(date = "2019-09-22",
                             dob = "1910-01-01",
                             break_max = NULL,
                             open_last = FALSE)
}
