% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_date_to_cohort.R
\name{plot_date_to_cohort_quarter}
\alias{plot_date_to_cohort_quarter}
\title{Depict the intervals created by
function 'date_to_cohort_quarter'}
\usage{
plot_date_to_cohort_quarter(date, break_min = NULL, open_first = NULL)
}
\arguments{
\item{date}{Dates of events or measurements.}

\item{break_min}{The start date of the first cohort,
or \code{NULL} (the default.)
If non-\code{NULL}, \code{break_min} can be a single value of
class \code{\link[base]{Date}}, or a value that
can be coerced to class \code{Date}
via function \code{\link[base]{as.Date}};
in either case, it must be the first day of
a month.}

\item{open_first}{Whether the first cohort
has no lower limit. If \code{break_min} is non-\code{NULL}
and \code{label_year_start} is \code{TRUE},
then \code{open_first} defaults to \code{TRUE};
otherwise it defaults to \code{FALSE}.}
}
\description{
Create plot illustrating how function
\code{\link{date_to_cohort_quarter}} works.
\code{plot_date_to_cohort_quarter} is typically used for
learning or documentation, rather than for
actual data analysis.
}
\examples{
plot_date_to_cohort_quarter(date = c("2024-03-27",
                                     "2022-05-13",
                                     "2022-11-09"))
plot_date_to_cohort_quarter(date = c("2024-03-27",
                                     "2022-05-13",
                                     "2022-11-09"),
                            break_min = "2022-01-01")
}
\keyword{interal}
