% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_date_to_cohort.R
\name{plot_date_to_cohort_year}
\alias{plot_date_to_cohort_year}
\title{Depict the intervals created by
function 'date_to_cohort_year'}
\usage{
plot_date_to_cohort_year(
  date,
  month_start = "Jan",
  label_year_start = TRUE,
  break_min = NULL,
  open_first = NULL
)
}
\arguments{
\item{date}{Dates of events or measurements.}

\item{month_start}{An element of \code{\link[base]{month.name}},
or \code{\link[base]{month.abb}}. The cohort starts on
the first day of this month.}

\item{label_year_start}{Whether to label a cohort
by the calendar year at the beginning of the cohort
or the calendar year at the end. Not needed for cohorts
that start on 1 January. Defaults to \code{TRUE}.}

\item{break_min}{The start date of the first cohort,
or \code{NULL} (the default.)
If non-\code{NULL}, \code{break_min} can be a single value of
class \code{\link[base]{Date}}, or a value that
can be coerced to class \code{Date}
via function \code{\link[base]{as.Date}};
in either case, it must be the first day of
a month.}

\item{open_first}{Whether the first cohort
has no lower limit. If \code{break_min} is non-\code{NULL}
and \code{label_year_start} is \code{TRUE}, then
then \code{open_first} defaults to \code{TRUE};
otherwise it defaults to \code{FALSE}.}
}
\description{
Create a plot illustrating how function
\code{\link{date_to_cohort_year}} works.
}
\details{
\code{plot_date_to_cohort_year} is designed for
learning or documentation, rather than for
actual data analysis.
}
\examples{
plot_date_to_cohort_year(date = c("2024-03-27",
                                  "2022-11-09"))

## July to June
plot_date_to_cohort_year(date = c("2024-03-27",
                                  "2022-11-09"),
                         month_start = "Jul")

## July to June, using the calendar year at
## the end for the label
plot_date_to_cohort_year(date = c("2024-03-27",
                                  "2022-11-09"),
                         month_start = "Jul",
                         label_year_start = FALSE)

## Specify oldest cohort, with open_first
## at default value of TRUE
plot_date_to_cohort_year(date = c("2024-03-27",
                                  "2019-08-22",
                                  "2022-11-09"),
                         break_min = "2020-01-01")

## Specify oldest cohort, with open_first = FALSE
plot_date_to_cohort_year(date = c("2024-03-27",
                                  "2019-08-22",
                                  "2022-11-09"),
                         break_min = "2015-07-01",
                         open_first = FALSE)
}
