% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_date_to_period.R
\name{plot_date_to_period_month}
\alias{plot_date_to_period_month}
\title{Depict the intervals created by
function 'date_to_period_month'}
\usage{
plot_date_to_period_month(date)
}
\arguments{
\item{date}{Dates of events or measurements.}
}
\description{
Create plot illustrating how function
\code{\link{date_to_period_month}} works.
}
\details{
\code{plot_date_to_period_month} is typically used for
learning or documentation, rather than for
actual data analysis.
}
\examples{
plot_date_to_period_month(date = c("2021-11-24",
                                   "2022-04-09"))
}
