% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_date_to_period.R
\name{plot_date_to_period_multi}
\alias{plot_date_to_period_multi}
\title{Depict the intervals created by
function 'date_to_period_multi'}
\usage{
plot_date_to_period_multi(date, width = 5, origin = 2000, month_start = "Jan")
}
\arguments{
\item{date}{Dates of events or measurements.}

\item{width}{The length, in whole years, of the periods.
Defaults to 5.}

\item{origin}{An integer. Defaults to 2000.
#'}

\item{month_start}{An element of \code{\link[base]{month.name}},
or \code{\link[base]{month.abb}}. The period starts on
the first day of this month.}
}
\description{
Depict the intervals created by
function 'date_to_period_multi'
}
\examples{
plot_date_to_period_multi(date = c("2024-03-27",
                                   "2018-11-09",
                                   "2021-03-02"))

## width is 10
plot_date_to_period_multi(date = c("2024-03-27",
                                   "2018-11-09",
                                   "2021-03-02"),
                          width = 5)

## origin is 2001
plot_date_to_period_multi(date = c("2024-03-27",
                              "2018-11-09",
                              "2021-03-02"),
                          origin = 2001)

## July to June
plot_date_to_period_multi(date = c("2024-03-27",
                              "2018-11-09",
                              "2021-03-02"),
                          origin = 2001,
                          month_start = "Jul")
}
\seealso{
\code{\link{date_to_period_multi}}
}
