% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_date_to_period.R
\name{plot_date_to_period_quarter}
\alias{plot_date_to_period_quarter}
\title{Depict the intervals created by
function 'date_to_period_quarter'}
\usage{
plot_date_to_period_quarter(date)
}
\arguments{
\item{date}{Dates of events or measurements.}
}
\description{
Create plot illustrating the inputs and outputs
of function \code{\link{date_to_period_quarter}}.
}
\details{
This function is for learning about the
labelling conventions used in the \strong{dem} packages,
and about their implementation in package
\strong{demprep}. It would not normally be used
during actual data analysis.
}
\examples{
plot_date_to_period_quarter(date = c("2024-03-27",
                                     "2020-01-03",
                                     "2022-11-09"))
}
