% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_date_to_period.R
\name{plot_date_to_period_quarter}
\alias{plot_date_to_period_quarter}
\title{Depict the intervals created by
function 'date_to_period_quarter'}
\usage{
plot_date_to_period_quarter(date)
}
\arguments{
\item{date}{Dates of events or measurements.}
}
\description{
Create plot illustrating how function
\code{\link{date_to_period_quarter}} works.
}
\details{
\code{plot_date_to_period_quarter} is typically used for
learning or documentation, rather than for
actual data analysis.
}
\examples{
plot_date_to_period_quarter(date = c("2021-11-24",
                                     "2022-04-09"))
}
\keyword{internal}
