% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_date_to_period.R
\name{plot_date_to_period_year}
\alias{plot_date_to_period_year}
\title{Depict the intervals created by
function 'date_to_period_year'}
\usage{
plot_date_to_period_year(date, month_start = "Jan", label_year_start = TRUE)
}
\arguments{
\item{date}{Dates of events or measurements.}

\item{month_start}{An element of \code{\link[base]{month.name}},
or \code{\link[base]{month.abb}}. The period starts on
the first day of this month.}

\item{label_year_start}{Whether to label a period
by the calendar year at the beginning of the period
or the calendar year at the end. Not needed for periods
that start on 1 January. Defaults to \code{FALSE}.}
}
\description{
Create plot illustrating the inputs and outputs
of function \code{\link{date_to_period_year}}.
}
\details{
This function is for learning about the
labelling conventions used in the \strong{dem} packages,
and about their implementation in package
\strong{demprep}. It would not normally be used
during actual data analysis.
}
\examples{
plot_date_to_period_year(date = c("2024-03-27", "2022-11-09"))

## July to June
plot_date_to_period_year(date = c("2024-03-27", "2022-11-09"),
                         month_start = "Jul")

## July to June, using the calendar year at
## the end for the label
plot_date_to_period_year(date = c("2024-03-27", "2022-11-09"),
                         month_start = "Jul",
                         label_year_start = FALSE)
}
