% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_date_to_triangle.R
\name{plot_date_to_triangle_month}
\alias{plot_date_to_triangle_month}
\title{Depict the Lexis triangles created by
function 'date_to_triangle_month'}
\usage{
plot_date_to_triangle_month(date, dob, break_max = 1200, open_last = TRUE)
}
\arguments{
\item{date}{Dates of events.}

\item{dob}{Dates of birth.}

\item{break_max}{An integer or \code{NULL}.
Defaults to 1200.}

\item{open_last}{Whether the final age group
has no upper limit. Defaults to \code{TRUE}.}
}
\description{
Create plot illustrating how function
\code{\link{date_to_triangle_month}} works.
}
\details{
\code{plot_date_to_triangle_month} is typically used for
instruction or documentation, rather than for
actual data analysis.
}
\examples{
plot_date_to_triangle_month(date = c("2024-03-27",
                                     "2024-11-09"),
                            dob = c("2023-10-05",
                                    "2024-03-11"),
                            break_max = 12)
}
