% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_date_to_triangle.R
\name{plot_date_to_triangle_multi}
\alias{plot_date_to_triangle_multi}
\title{Depict the Lexis triangles created by
function 'date_to_triangle_multi'}
\usage{
plot_date_to_triangle_multi(
  date,
  dob,
  width = 5,
  break_max = 100,
  open_last = TRUE,
  origin = 2000,
  month_start = "Jan",
  show_months = FALSE
)
}
\arguments{
\item{date}{Dates of events.}

\item{dob}{Dates of birth.}

\item{width}{The width in years of the periods and
age groups. A positive integer defaulting to 5.}

\item{break_max}{An integer or \code{NULL}.
Defaults to 100.}

\item{open_last}{Whether the final age group
has no upper limit. Defaults to \code{TRUE}.}

\item{origin}{An integer. Defaults to 2000.}

\item{month_start}{An element of \code{\link[base]{month.name}},
or \code{\link[base]{month.abb}}. The period starts on
the first day of this month.}

\item{show_months}{Whether to include vertical
lines showing boundaries between months.
Defaults to \code{FALSE}.}
}
\description{
Create plot illustrating how function
\code{\link{date_to_triangle_multi}} works.
}
\details{
\code{plot_date_to_triangle_multi} is typically used for
instruction or documentation, rather than for
actual data analysis.
}
\examples{
plot_date_to_triangle_multi(date = c("2027-03-27",
                                     "2022-11-09"),
                            dob = c("2012-05-12",
                                    "2016-02-29"),
                            break_max = 20)

## width is 10
plot_date_to_triangle_multi(date = c("2027-03-27",
                                     "2022-11-09"),
                            dob = c("2012-05-12",
                                    "2016-02-29"),
                            width = 10,
                            break_max = 20)
}
\keyword{internal}
