% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_date_to_triangle.R
\name{plot_date_to_triangle_quarter}
\alias{plot_date_to_triangle_quarter}
\title{Depict the Lexis triangles created by
function 'date_to_triangle_quarter'}
\usage{
plot_date_to_triangle_quarter(
  date,
  dob,
  break_max = 400,
  open_last = TRUE,
  show_months = FALSE
)
}
\arguments{
\item{date}{Dates of events.}

\item{dob}{Dates of birth.}

\item{break_max}{An integer or \code{NULL}.
Defaults to 400.}

\item{open_last}{Whether the final age group
has no upper limit. Defaults to \code{TRUE}.}

\item{show_months}{Whether to include vertical
lines showing boundaries between months.
Defaults to \code{FALSE}.}
}
\description{
Create plot illustrating how function
\code{\link{date_to_triangle_quarter}} works.
}
\details{
\code{plot_date_to_triangle_quarter} is typically used for
instruction or documentation, rather than for
actual data analysis.
}
\examples{
plot_date_to_triangle_quarter(date = c("2024-03-27",
                                       "2022-11-09"),
                              dob = "2020-01-01")

## open age group starts at 40 quarters
date_to_triangle_quarter(date = c("2017-03-27",
                                  "2024-03-27"),
                         dob = "2010-01-01",
                         break_max = 40)
}
