% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_date_to_triangle.R
\name{plot_date_to_triangle_year}
\alias{plot_date_to_triangle_year}
\title{Depict the Lexis triangles created by
function 'date_to_triangle_year'}
\usage{
plot_date_to_triangle_year(
  date,
  dob,
  break_max = 100,
  open_last = TRUE,
  month_start = "Jan",
  show_months = FALSE
)
}
\arguments{
\item{date}{Dates of events.}

\item{dob}{Dates of birth.}

\item{break_max}{An integer or \code{NULL}.
Defaults to 100.}

\item{open_last}{Whether the final age group
has no upper limit. Defaults to \code{TRUE}.}

\item{month_start}{An element of \code{\link[base]{month.name}},
or \code{\link[base]{month.abb}}. The period starts on
the first day of this month.}

\item{show_months}{Whether to include vertical
lines showing boundaries between months.
Defaults to \code{FALSE}.}
}
\description{
Create plot illustrating how function
\code{\link{date_to_triangle_year}} works.
}
\details{
\code{plot_date_to_triangle_year} is typically used for
instruction or documentation, rather than for
actual data analysis.
}
\examples{
plot_date_to_triangle_year(date = c("2024-05-27",
                                    "2024-11-09"),
                           dob = c("2020-03-13",
                                   "2021-08-24"),
                           break_max = 5)

## final age group closed
plot_date_to_triangle_year(date = c("2024-05-27",
                                    "2024-11-09"),
                           dob = c("2020-03-13",
                                   "2021-08-24"),
                           break_max = 5,
                           open_last = FALSE)

## start on 1 July, rather than 1 January
plot_date_to_triangle_year(date = c("2024-05-27",
                                    "2024-11-09"),
                           dob = c("2020-03-13",
                                   "2021-08-24"),
                           break_max = 5,
                           month_start = "July")
}
